### 2023-08-02 #5730 @editorialbot
[PRE REVIEW]: konfound: An R Sensitivity Analysis Package to Quantify the Robustness of Causal
**Submitting author:** @jrosen48 (<a href="http://orcid.org/0000-0003-2170-0447">Joshua Rosenberg</a>)
**Repository:** https://github.com/konfound-project/konfound
**Branch with paper.md** (empty if default branch):
**Version:** 0.4.0
