+++ joss-paper/paper.bib
@@ -62,7 +62,8 @@ eprint = {https://doi.org/10.1111/j.1467-9531.2007.00186.x}
  publisher = {[American Educational Research Association, Sage Publications, Inc.]},
  title = {Does NBPTS Certification Affect the Number of Colleagues a Teacher Helps with Instructional Matters?},
  volume = {30},
- year = 2008
+ year = 2008,
+ doi = {10.3102/0162373707313781}
 }
 
 @article{frank2013,
@@ -110,7 +111,8 @@ url={https://papers.ssrn.com/sol3/papers.cfm?abstract_id=4305243},
 journal={SSRN}, 
 author={{Frank}, K.~A. and {Lin}, Q. and {Maroulis}, S. and {Dai}, S. and {Jess}, N. and {Lin}, H. and {Liu}, Y. and {Maestrales}, S. and {Searle}, E. and {Tait}, J.}, 
 year=2022, 
-month=dec
+month=dec,
+doi = {10.2139/ssrn.4305243}
 } 
 
 @article{xu2019,
