### 2023-08-14 #5730 @jrosen48
possible/suggested reviewers based on a review of the database of JOSS reviewers:

- @fartist (causal inference, R)
- @FATelarico (social science quantitative methods)
- @wjakethompson (quantitative methods in educational research contexts)
