+++ tests/testthat/test-konfound.R
@@ -18,7 +18,7 @@ m4 <- glm(outcome ~ condition, data = binary_dummy_data, family = binomial(link 
 output4 <- konfound(m4, condition, two_by_two = TRUE, n_treat = 55, to_return = "raw_output")
 
 test_that("konfound works for linear model", {
-    expect_equal(output1$percent_bias_to_change_inference, 66.629, tolerance = .001)
+    expect_equal(output1$perc_bias_to_change, 66.57696, tolerance = .001)
 })
 
 # test_that("konfound works for glm, 2x2 model", {
@@ -26,7 +26,7 @@ test_that("konfound works for linear model", {
 # })
 
 test_that("konfound works for lme4 model", {
-    expect_equal(output3$percent_bias_to_change_inference, 84.826, tolerance = .001)
+    expect_equal(output3$perc_bias_to_change, 84.826, tolerance = .001)
 })
 
 test_that("konfound works for glm, 2x2 model", {
