+++ R/helper_output_list.R
@@ -1,0 +1,35 @@
+output_list <- function(obs_r, critical_r, rxcv, rycv, rxcvGz, rycvGz, 
+                        itcvGz, itcv, rxz, ryz, 
+                        delta_star, delta_star_restricted, delta_exact, delta_pctbias, 
+                        cor_oster, cor_exact, 
+                        beta_threshold, perc_bias_to_change, 
+                        RIR, RIR_perc, fragility, starting_table, SE, 
+                        Fig_ITCV, Fig_RIR) {
+  result <- list(obs_r = obs_r,
+                 critical_r = critical_r, 
+                 rxcv = rxcv, 
+                 rycv = rycv, 
+                 rxcvGz = rxcvGz, 
+                 rycvGz = rycvGz, 
+                 itcvGz = itcvGz, 
+                 itcv = itcv, 
+                 rxz = rxz, 
+                 ryz = ryz, 
+                 delta_star = delta_star, 
+                 delta_star_restricted = delta_star_restricted, 
+                 delta_exact = delta_exact, 
+                 delta_pctbias = delta_pctbias, 
+                 cor_oster = cor_oster, 
+                 cor_exact = cor_exact, 
+                 beta_threshold = beta_threshold, 
+                 perc_bias_to_change = perc_bias_to_change, 
+                 RIR = RIR, 
+                 RIR_perc = RIR_perc, 
+                 fragility = fragility, 
+                 starting_table = starting_table, 
+                 SE = SE, 
+                 Fig_ITCV = Fig_ITCV,
+                 Fig_RIR = Fig_RIR)
+  result <- result[!is.na(result)]
+  return(result)
+}

[…]
