### 2023-08-20 c8a176a
Update test_sensitivity.R
take care of cases where eff_thr is specified but sdx and sdy are not given (assuming they are both 1)
