+++ R/test_sensitivity_ln.R
@@ -6,7 +6,7 @@ test_sensitivity_ln <- function(est_eff,
                                 n_covariates,
                                 n_treat,
                                 switch_trm = T,
-                                replace = "entire",
+                                replace = "control",
                                 alpha,
                                 tails,
                                 nu,
