### 2023-08-24 #5730 @fartist
Sure, I can take this one – but I would like to have more time for this one since I am in the process of relocation.

From: Fabian Scheipl ***@***.***>
Date: Thursday, August 24, 2023 at 10:20 AM
Cc: Feeeeeeeeeeng ***@***.***>, Mention ***@***.***>
Subject: Re: [openjournals/joss-reviews] [PRE REVIEW]: konfound: An R Sensitivity Analysis Package to Quantify the Robustness of Causal (Issue #5730)

👋 hi @fartist<https://github.com/fartist> @FATelarico<https://github.com/FATelarico> @wjakethompson<https://github.com/wjakethompson>
would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
