### 2023-08-24 konfound-project/konfound#50 @fabian-s
Code quality issues
1. goodpractice::gp  finds very low test coverage and some inconsistent / non-compliant coding style.
The former at least should be improved for your JOSS publication

2. cyclomatic complexity of many of your functions is extremely high, casting doubt on maintainability, legibility and correctness of the implementation.
maybe try refactoring/getting rid of some of your conditional logic and/or modularize your functions to get cyclomatic complexities closer to reasonable values of like 20-50?

```
goodpractice::gp("konfound")
Preparing: covr
Preparing: cyclocomp
Preparing: description
Preparing: lintr
Preparing: namespace
Preparing: rcmdcheck
── GP konfound ─────────────────────────────────────────────────────────────────

It is good practice to

  ✖ write unit tests for all functions, and all package code in
    general. 27% of code lines are covered by test cases.

    R/cop_pse_auxiliary.R:2:NA
    R/cop_pse_auxiliary.R:3:NA
    R/cop_pse_auxiliary.R:4:NA
    R/cop_pse_auxiliary.R:6:NA
    R/cop_pse_auxiliary.R:8:NA
    ... and 1745 more lines

  ✖ use '<-' for assignment instead of '='. '<-' is the
    standard, and R users and developers are used it and it is easier
    to read your code for them if you use '<-'.

    R/cop_pse_auxiliary.R:2:10
    R/cop_pse_auxiliary.R:4:13
    R/cop_pse_auxiliary.R:12:10
    R/cop_pse_auxiliary.R:15:9
    R/cop_pse_auxiliary.R:20:9
    ... and 174 more lines

    R/concord1.R:7:81
    R/cop_pse_auxiliary.R:24:81
    R/cop_pse_auxiliary.R:26:81
    R/cop_pse_auxiliary.R:38:81
    R/cop_pse_auxiliary.R:137:81
    ... and 447 more lines

  ✖ avoid 1:length(...), 1:nrow(...), 1:ncol(...), 1:NROW(...)
    and 1:NCOL(...) expressions. They are error prone and result 1:0 if
    the expression on the right hand side is zero. Use seq_len() or
    seq_along() instead.

    R/mkonfound.R:45:44

  ✖ not import packages as a whole, as this can cause name
    clashes between the imported packages. Instead, import only the
    specific functions you need.
  ✖ avoid 'T' and 'F', as they are just variables which are set
    to the logicals 'TRUE' and 'FALSE' by default, but are not reserved
    words and hence can be overwritten by the user.  Hence, one should
    always use 'TRUE' and 'FALSE' for the logicals.

    R/cop_pse_auxiliary.R:NA:NA
    R/cop_pse_auxiliary.R:NA:NA
    R/cop_pse_auxiliary.R:NA:NA
    R/cop_pse_auxiliary.R:NA:NA
    R/cop_pse_auxiliary.R:NA:NA
    ... and 21 more lines
```

```
> cyclocomp::cyclocomp_package("konfound")
                         name cyclocomp
43        test_sensitivity_ln       217
22           getswitch_fisher        79
45              tkonfound_fig        74
21            getswitch_chisq        70
44                  tkonfound        70
20                  getswitch        37
[....]
```

crossref: openjournals/joss-reviews#5779
