### 2023-09-11 2dc3fa1
use n-1 in calculating rxz and reporting r2xz and r2yz instead
