### 2023-09-11 b9defd2
update sign of rycv
rxcv should always be positive, rycv goes with the direction with ITCV
