### 2023-09-11 ac1fe19
update df extracted
the extracted df includes the predictor
