### 2023-09-11 d11c252
Update helper_output_list.R
