+++ R/helper_output_list.R
@@ -1,5 +1,5 @@
 output_list <- function(obs_r, critical_r, rxcv, rycv, rxcvGz, rycvGz, 
-                        itcvGz, itcv, rxz, ryz, 
+                        itcvGz, itcv, r2xz, r2yz, 
                         delta_star, delta_star_restricted, delta_exact, delta_pctbias, 
                         cor_oster, cor_exact, 
                         beta_threshold, perc_bias_to_change, 
@@ -13,8 +13,8 @@ output_list <- function(obs_r, critical_r, rxcv, rycv, rxcvGz, rycvGz,
                  rycvGz = rycvGz, 
                  itcvGz = itcvGz, 
                  itcv = itcv, 
-                 rxz = rxz, 
-                 ryz = ryz, 
+                 r2xz = r2xz, 
+                 r2yz = r2yz, 
                  delta_star = delta_star, 
                  delta_star_restricted = delta_star_restricted, 
                  delta_exact = delta_exact, 
