+++ R/pkonfound.R
@@ -207,7 +207,6 @@
     suppression = suppression,
     eff_thr = eff_thr,
     to_return = to_return
-    
   )
 } 
       
