+++ R/konfound-lm.R
@@ -33,7 +33,7 @@ konfound_lm <- function(model_object, tested_variable_string, test_all, alpha, t
       index = index,
       nu = 0,
       suppression = 0,
-      eff_thr = 0, 
+      eff_thr = NA, 
       to_return = to_return,
       model_object = model_object,
       tested_variable = tested_variable_string
