### 2023-09-18 konfound-project/konfound#50 @fabian-s
Any updates on this yet?
