### 2023-09-19 konfound-project/konfound#53 @wjakethompson
Test coverage
JOSS review criterion: Automated tests (https://github.com/openjournals/joss-reviews/issues/5779)

Using covr::package_coverage(), I see only 27.11% test coverage. I think this should be substantially higher in order to adequately verify that the package functions as expected and to ensure that code is immune to bugs that may be introduced by future changes.

```
covr::package_coverage()
konfound Coverage: 27.11%
R/cop_pse_auxiliary.R: 0.00%
R/helper_output_print.R: 0.00%
R/helper_output_table.R: 0.00%
R/helper_plot_correlation.R: 0.00%
R/helper_plot_threshold.R: 0.00%
R/konfound-glm.R: 0.00%
R/test_cop.R: 0.00%
R/test_pse.R: 0.00%
R/tkonfound_fig.R: 0.00%
R/test_sensitivity.R: 29.03%
R/pkonfound.R: 30.61%
R/nonlinear_auxiliary.R: 36.82%
R/tkonfound.R: 41.62%
R/test_sensitivity_ln.R: 43.81%
R/helper_output_dataframe.R: 44.44%
R/zzz.R: 50.00%
R/mkonfound.R: 60.78%
R/konfound.R: 66.13%
R/konfound-glm-dichotomous.R: 100.00%
R/konfound-lm.R: 100.00%
R/konfound-lmer.R: 100.00%
```
