### 2023-09-19 konfound-project/konfound#54 @wjakethompson
Add community guidelines
JOSS review criterion: Community guidelines (https://github.com/openjournals/joss-reviews/issues/5779)

No clear guidelines for external contributions are provided. `usethis::use_tidy_contributing()` provides a good starting point for creating this document (modifications will be needed, but most of the content is relevant).

I would also recommended adding a code of conduct, which can be added using `usethis::use_code_of_conduct()`, and then modified as needed.
