### 2023-09-19 konfound-project/konfound#53 @fabian-s
Seconded!
