### 2023-09-19 #5779 @wjakethompson
I have completed my review, with a few outstanding issues:

1. A few recommendations from `goodpractice::gp()` to ensure functionality of the package works as intended, as noted by @fabian-s. crossref: konfound-project/konfound#50
2. Test coverage is very low (~27%). I think this should be substantially higher to meet the documentation criterion. crossref: konfound-project/konfound#53
3. There are currently no contributor guidelines. Specific guidelines should be added, and I would also recommend adding a code of conduct. crossref: konfound-project/konfound#54
4. Finally, a two notes on the paper:
    * The statement of need does not address a specific target audience. To me the implied audience is applied researchers, but that's not explicitly stated anywhere.
    * The paper does not currently describe how this software compares to existing packages for sensitivity analyses, or mention what other packages are available.
    * Some of the output in the example uses runs off the edge of the page (this may be a typesetting issue?)
