### 2023-09-20 #5779 @fartist
I will get reviews in by this month – sorry for the delay.

From: Fabian Scheipl ***@***.***>
Date: Wednesday, September 20, 2023 at 3:52 AM
Cc: Feeeeeeeeeeng ***@***.***>, Mention ***@***.***>
Subject: Re: [openjournals/joss-reviews] [REVIEW]: konfound: An R Sensitivity Analysis Package to Quantify the Robustness of Causal (Issue #5779)

@jrosen48<https://github.com/jrosen48> : to speed up the review process it would be good to start working on these while we wait for the missing review by @fartist<https://github.com/fartist>. please let us know your timeframe for doing so.
