### 2023-09-27 konfound-project/konfound#50 @jrosen48
Hi, sorry for the delay - I'm on parental leave for another two weeks but making some progress with collaborators in some free moments. I will plan to update again in around three weeks or less, if that sounds okay.
