### 2023-10-10 d690abb
update signITCV to cover more scenarios
