### 2023-10-12 bd3f433
Update tkonfound.R
T to TRUE
