### 2023-10-13 bf509d6
Update tkonfound_fig.R
T to TRUE; F to FALSE
