### 2023-10-22 6ddf2cb
change all -3 to -2 for df calculation
