### 2023-10-23 #5779 @fartist
Hi I am in the middle of reviewing it – two more weeks to finish. Hope it is okay, sorry for the delay.

Best,

From: Fabian Scheipl ***@***.***>
Date: Monday, October 23, 2023 at 6:48 AM
Cc: Feeeeeeeeeeng ***@***.***>, Mention ***@***.***>
Subject: Re: [openjournals/joss-reviews] [REVIEW]: konfound: An R Sensitivity Analysis Package to Quantify the Robustness of Causal (Issue #5779)

hi @fartist<https://github.com/fartist>, we are still waiting for your review of this JOSS submission.
It's been a month since your last post to this thread, please let us know how much more time you will need.
