### 2023-10-23 #5779 @fabian-s
@fartist sure, it can't be helped if it takes this long. to speed things up at least a little bit:

- please do generate your checklist on github in this comment thread by typing
`@editorialbot generate my checklist`
- please check off the items there that you have already reviewed and found sufficient
- please open [issues in the project repository](https://github.com/konfound-project/konfound/issues) and link to them here by mentioning https://github.com/openjournals/joss-reviews/issues/5779,  or comment on the problems you've already identified in this thread

no need to  hold back your feedback until you are all done.
