### 2023-10-24 daf45c8
add r_final to check & adjust RIR (need to verify)
(1) add r_final based on conditional correlation calculation to verify itcv calculation for nonzero nu
(2) add beta_threshold_rir
All these changes could be temporary & need verification & finalization.
