### 2023-10-28 c40276e
testing the new process of determining mp and signITCV
in the process of testing...
