### 2023-10-28 1f7bcf0
remove beta_threshold_rir
