+++ R/test_sensitivity.R
@@ -203,7 +203,6 @@ test_sensitivity <- function(est_eff,
                        delta_exact = NA, delta_pctbias = NA, 
                        cor_oster = NA, cor_exact = NA, 
                        beta_threshold = beta_threshold,
-                       beta_threshold_rir = beta_threshold_rir,
                        perc_bias_to_change = perc_to_change, 
                        RIR = recase, RIR_perc = perc_to_change, 
                        fragility = NA, starting_table = NA, 
