+++ R/helper_output_list.R
@@ -2,7 +2,7 @@ output_list <- function(obs_r, critical_r, r_final, rxcv, rycv, rxcvGz, rycvGz,
                         itcvGz, itcv, r2xz, r2yz, 
                         delta_star, delta_star_restricted, delta_exact, delta_pctbias, 
                         cor_oster, cor_exact, 
-                        beta_threshold, beta_threshold_rir, perc_bias_to_change, 
+                        beta_threshold, perc_bias_to_change, 
                         RIR, RIR_perc, fragility, starting_table, SE, 
                         Fig_ITCV, Fig_RIR) {
   result <- list(obs_r = obs_r,
@@ -23,7 +23,6 @@ output_list <- function(obs_r, critical_r, r_final, rxcv, rycv, rxcvGz, rycvGz,
                  cor_oster = cor_oster, 
                  cor_exact = cor_exact, 
                  beta_threshold = beta_threshold, 
-                 beta_threshold_rir = beta_threshold_rir,
                  perc_bias_to_change = perc_bias_to_change, 
                  RIR = RIR, 
                  RIR_perc = RIR_perc, 
