### 2023-10-29 79584b2
Update helper_output_print.R
test code
