### 2023-11-07 913f771
Update test_sensitivity_ln.R
add fragility information in the output
