### 2023-11-07 c1d54d1
Update test_sensitivity_ln.R
fix the bug of no extra RIR when only one row needed
