### 2023-11-07 #5779 @fabian-s
copying the e-mail I received from our "reviewer 2" into this thread:

> Apologies for the delay. In brief, I lost access to 2FA and GitHub because my cellphone was stolen in an accident. I have been trying to recover my account but have not been able to fix it. Currently, I can only respond to your email and not access the thread. I have read the paper, examined, and tested the package, and everything seems good to me. I just have a few minor comments on the draft.

- Line 149, SE should have a hat and standard error should be standard error estimate.
- Number of digits should be consistent throughout the text: e.g., .568 in line 173 vs. .52 in line 109.
- APA style were not followed at multiple places e.g., line 108, line 134.

 @jrosen48 please address these in addition, to the issues pointed out by @wjakethompson and myself.
 could you give us a timeframe for your patches so we can move this along?
