### 2023-11-07 #5779 @fabian-s
thx

On Tue, Nov 7, 2023 at 5:26 AM Feeeeeeeeeeng ***@***.***>
wrote:

> Apologies for the delay. In brief, I lost access to 2FA and GitHub because
> my cellphone was stolen in an accident. I have been trying to recover my
> account but have not been able to fix it. Currently, I can only respond to
> your email and not access the thread. I have read the paper, examined, and
> tested the package, and everything seems good to me. I just have a few
> minor comments on the draft.
> Line 149, SE should have a hat and standard error should be standard error
> estimate.
> Number of digits should be consistent throughout the text: e.g., .568 in
> line 173 vs. .52 in line 109.
> APA style were not followed at multiple places e.g., line 108, line 134.
>
> Best.
>
> From: Fabian Scheipl ***@***.***>
> Date: Monday, November 6, 2023 at 5:42 AM
> To: openjournals/joss-reviews ***@***.***>
> Cc: Feeeeeeeeeeng ***@***.***>, Mention ***@***.***>
> Subject: Re: [openjournals/joss-reviews] [REVIEW]: konfound: An R
> Sensitivity Analysis Package to Quantify the Robustness of Causal (Issue
> #5779)
>
> @fartist<https://github.com/fartist> two more weeks have now passed, can
> we please have your review?
>
