### 2023-11-19 b16cd81
change -2 to -3 for critical_t and critical_r
