### 2023-11-19 57f0268
change -2 to -3 for critical_t and critical_r
