### 2023-11-20 d32fe06
Update helper_output_dataframe.R
Linefeed the parameter "output_df <- function" to ensure that the line of code does not exceed 80 characters.
