### 2023-11-20 3cbab8b
Update helper_output_print.R
Manual modification: no more than 80 characters per line of code
