+++ R/helper_output_print.R
@@ -1,24 +1,47 @@
 # Function to output printed text
 
-output_print <- function(eff_diff, beta_threshhold, bias = NULL, sustain = NULL, nu, recase, obs_r, critical_r, r_con, itcv, alpha, index) {
-  if (index == "RIR"){ 
-    cat(crayon::bold("Robustness of Inference to Replacement (RIR):\n"))
-    if (abs(eff_diff) > abs(beta_threshhold)) {
-      cat("To invalidate an inference, ", round(bias, 3), "% of the estimate would have to be due to bias. ")
-      cat("\n")
-      cat("This is based on a threshold of ", round(beta_threshhold, 3), " for statistical significance (alpha = ", alpha, ").\n", sep = "")
-      cat("\n")
-      cat("To invalidate an inference, ", round(recase, 3), " observations would have to be replaced with cases")
-      cat("\n")
-      cat("for which the effect is ", nu, " (RIR = ", round(recase, 3), ").\n", sep = "")
-      cat("\n")
+output_print <- function(eff_diff, 
+                         beta_threshhold, 
+                         bias = NULL, 
+                         sustain = NULL, 
+                         nu, 
+                         recase, 
+                         obs_r, 
+                         critical_r, 
+                         r_con, 
+                         itcv, 
+                         alpha, 
+                         index) {
+ if (index == "RIR") { 
+  cat(crayon::bold("Robustness of Inference to Replacement (RIR):\n"))
+  if (abs(eff_diff) > abs(beta_threshhold)) {
+    cat("To invalidate an inference, ", round(bias, 3), 
+        "% of the estimate would have to be due to bias. ")
+    cat("\n")
+    cat("This is based on a threshold of ", round(beta_threshhold, 3), 
+        " for statistical significance (alpha = ", alpha, ").\n", 
+        sep = "")
+    cat("\n")
+    cat("To invalidate an inference, ", round(recase, 3), 
+        " observations would have to be replaced with cases")
+    cat("\n")
+    cat("for which the effect is ", nu,
+        " (RIR = ", round(recase, 3), ").\n", sep = "")
+    cat("\n")
     }
     else if (abs(eff_diff) < abs(beta_threshhold)) {
-      cat("To sustain an inference, ", round(sustain, 3), "% of the estimate would have to be due to bias. ")
+      cat("To sustain an inference, ", round(sustain, 3), 
+          "% of the estimate would have to be due to bias. ")
       cat("\n")
-      cat("This is based on a threshold of ", round(beta_threshhold, 3), " for statistical significance (alpha = ", alpha, ").\n", sep = "")
+      cat("This is based on a threshold of ", round(beta_threshhold, 3), 
+          " for statistical significance (alpha = ", alpha, ").\n", 
+          sep = "")
       cat("\n")
-      cat("To sustain an inference, ", round(recase, 3), " of the cases with ", nu, " effect would have to be replaced with cases at the threshold of inference"," (RIR = ", round(recase, 3), ").\n", sep = "")
+      cat("To sustain an inference, ", round(recase, 3), 
+          " of the cases with ", nu, 
+          " effect would have to be replaced with cases at the 
+          threshold of inference",
+          " (RIR = ", round(recase, 3), ").\n", sep = "")
     }
     else if (eff_diff == beta_threshhold) {
       warning("The coefficient is exactly equal to the threshold.\n")
@@ -26,65 +49,94 @@ output_print <- function(eff_diff, beta_threshhold, bias = NULL, sustain = NULL,
     cat("See Frank et al. (2013) for a description of the method.")
     cat("\n")
     cat("\n")
-    cat(crayon::underline("Citation:"), "Frank, K.A., Maroulis, S., Duong, M., and Kelcey, B. (2013).")
[…]
