### 2023-11-20 4fc7208
Update helper_output_table.R
Manual modification: no more than 80 characters per line of code
