### 2023-11-20 eef4c49
Update helper_plot_correlation.R
Manual modification: no more than 80 characters per line of code
