+++ R/helper_plot_correlation.R
@@ -13,28 +13,111 @@ plot_correlation <- function(r_con, obs_r, critical_r) {
 
   p <- ggplot2::ggplot(d, ggplot2::aes_string(x = "x", y = "y")) +
 
-    # main arrows
-    ggplot2::geom_segment(ggplot2::aes(y = .1), xend = 0, yend = .9, arrow = ggplot2::arrow(), size = 2.5, color = "#A6CEE3") + # straight up
-    ggplot2::geom_segment(ggplot2::aes(x = .1), xend = 1, yend = .9, arrow = ggplot2::arrow(), size = 2.5, color = "#A6CEE3") + # hypotenuse
-    ggplot2::geom_segment(ggplot2::aes(x = .15, y = 1), xend = .9, yend = 1, arrow = ggplot2::arrow(), size = 2.5, color = "#A6CEE3") + # straight across
+# main arrows
+ggplot2::geom_segment(
+  ggplot2::aes(y = .1),
+  xend = 0,
+  yend = .9,
+  arrow = ggplot2::arrow(),
+  size = 2.5,
+  color = "#A6CEE3"
+) + # straight up
+ggplot2::geom_segment(
+  ggplot2::aes(x = .1),
+  xend = 1,
+  yend = .9,
+  arrow = ggplot2::arrow(),
+  size = 2.5,
+  color = "#A6CEE3"
+) + # hypotenuse
+ggplot2::geom_segment(
+  ggplot2::aes(x = .15, y = 1),
+  xend = .9,
+  yend = 1,
+  arrow = ggplot2::arrow(),
+  size = 2.5,
+  color = "#A6CEE3"
+) + # straight across
 
-    # lagel annotations
-    ggplot2::annotate("text", x = 0, y = 0, label = paste0("Confounding\nVariable"), fontface = 3) +
-    ggplot2::annotate("text", x = 0, y = 1, label = paste0("Predictor of Interest"), fontface = 3) +
-    ggplot2::annotate("text", x = 1, y = 1, label = paste0("Outcome"), fontface = 3) +
 
-    # CV arrows
-    # ggplot2::geom_segment(ggplot2::aes(x = .05, y = .25), xend = .275, yend = .65, arrow = ggplot2::arrow(), size = 2.5, color = "#1F78B4") + # straight across
-    # ggplot2::geom_segment(ggplot2::aes(x = .175, y = .15), xend = .3, yend = .625, arrow = ggplot2::arrow(), size = 2.5, color = "#1F78B4") + # straight across
-    ggplot2::geom_curve(ggplot2::aes(x = .04, y = .325, xend = .35, yend = .825), curvature = -.35, size = 2.5, color = "#1F78B4", arrow = ggplot2::arrow()) +
-    ggplot2::geom_curve(ggplot2::aes(x = .225, y = .23, xend = .4, yend = .8), curvature = .35, size = 2.5, color = "#1F78B4", arrow = ggplot2::arrow()) +
+# label annotations
+ggplot2::annotate(
+  "text",
+  x = 0,
+  y = 0,
+  label = paste0("Confounding\nVariable"),
+  fontface = 3
+) +
+ggplot2::annotate(
+  "text",
+  x = 0,
+  y = 1,
+  label = paste0("Predictor of Interest"),
+  fontface = 3
+) +
+ggplot2::annotate(
+  "text",
+  x = 1,
+  y = 1,
+  label = paste0("Outcome"),
+  fontface = 3
+) +
 
[…]
