### 2023-11-20 64e1584
Update helper_plot_threshold.R
Manual modification: no more than 80 characters per line of code
