### 2023-11-20 b0e5adc
Update konfound-glm-dichotomous.R
Manual modification: no more than 80 characters per line of code
