+++ R/konfound-glm-dichotomous.R
@@ -1,7 +1,13 @@
 # konfound-glm
 
-konfound_glm_dichotomous <- function(model_object, tested_variable_string, test_all, alpha, tails,
-                                     to_return, n_treat, switch_trm, replace) {
+konfound_glm_dichotomous <- function(model_object, 
+                                     tested_variable_string, 
+                                     test_all, 
+                                     alpha, tails,
+                                     to_return, 
+                                     n_treat, 
+                                     switch_trm, 
+                                     replace) {
   tidy_output <- broom::tidy(model_object) # tidying output
   glance_output <- broom::glance(model_object)
 
