### 2023-11-20 68d4824
Update konfound-glm.R
Manual modification: no more than 80 characters per line of code
