### 2023-11-20 5fd0bca
Update konfound-lm.R
Manual modification: no more than 80 characters per line of code
