### 2023-11-20 e2a806b
Update konfound-lmer.R
Manual modification: no more than 80 characters per line of code
