### 2023-11-20 1c0d2b4
Update konfound.R
some comment need to lines break
