### 2023-11-21 3926d82
Update helper_output_print.R
Changed the output language for output_print() following Dr. Frank's and Dr. Maroulis's recommendations. The major change includes a condition for the eff_thr status.
