### 2023-11-21 87b589a
Update test_sensitivity_ln.R
Changed the output language for test_sensitivity_ln() following Dr. Frank's and Dr. Maroulis' recommendations. The major change includes 1) changeSE status, 2) Fragility.
