### 2023-11-21 c070235
Update helper_output_print.R
