+++ R/pkonfound.R
@@ -214,6 +214,8 @@ if (!is.null(out)) { # dealing with a strange print issue
   return(out)
 }
 
+cat("\n")
+
 if (to_return == "print") {
   message("For other forms of output, run ?pkonfound and inspect the to_return argument")
 }
