### 2023-11-21 a4e97f4
Update pkonfound.R
