### 2023-11-21 a499746
Update helper_output_print.R
add "= NULL" for index due to error "Error in output_print(est_eff, beta_threshold, bias, sustain, nu, recase,  : 
  argument "index" is missing, with no default"
