### 2023-11-21 e3323dc
Update helper_output_print.R
