### 2023-11-21 e733766
Update test_sensitivity_ln.R
solve parentheses problem
