+++ R/test_sensitivity_ln.R
@@ -364,10 +364,10 @@ test_sensitivity_ln <- function(est_eff,
           #conclusion1b 
           cat(paste(
             sprintf("This is equivalent to \ntransferring %d", final_solution$final_switch), 
-            c("case from"), transferway, 
+            c("case from"), transferway), 
             sprintf("\nin the initial table (Fragility = %d).", result$fragility),
             sprintf("\nThis transfer of cases yields the following table:")
-          ))
+          )
           
         } else if (!final_solution$needtworows & final_solution$final_switch == 1) {
           #conclusion1 
@@ -385,10 +385,10 @@ test_sensitivity_ln <- function(est_eff,
           #conclusion1b 
           cat(paste(
             sprintf("This is equivalent to \ntransferring %d", final_solution$final_switch), 
-            c("case from"), transferway, 
+            c("case from"), transferway), 
             sprintf("\nin the initial table (Fragility = %d).", result$fragility),
             sprintf("\nThis transfer of cases yields the following table:")
-          ))
+          )
           
         } else {
           #conclusion1 
@@ -433,10 +433,11 @@ test_sensitivity_ln <- function(est_eff,
           #conclusion1b 
           cat(paste(
             sprintf("This is equivalent \nto transferring %d", final_solution$final_switch), 
-            c("case from"), transferway, 
+            c("case from"), transferway), 
             sprintf("\nin the initial table (Fragility = %d).", result$fragility),
-            sprintf("\nThis transfer of cases yields the following table:"))
+            sprintf("\nThis transfer of cases yields the following table:")
              )
+             
           
         } else if (!final_solution$needtworows & final_solution$final_switch == 1) {
           #conclusion1 
@@ -454,10 +455,11 @@ test_sensitivity_ln <- function(est_eff,
           #conclusion1b 
           cat(paste(
             sprintf("This is equivalent \nto transferring %d", final_solution$final_switch), 
-            c("case from"), transferway, 
+            c("case from"), transferway), 
             sprintf("\nin the initial table (Fragility = %d).", result$fragility),
-            sprintf("\nThis transfer of cases yields the following table:"))
+            sprintf("\nThis transfer of cases yields the following table:")
              )
+             
           
         } else {
           #conclusion1
@@ -520,10 +522,11 @@ test_sensitivity_ln <- function(est_eff,
           #conclusion1b 
           cat(paste(
             sprintf("This is equivalent to \ntransferring %d", final_solution$final_switch), 
-            c("case from"), transferway, 
+            c("case from"), transferway), 
             sprintf("\nin the initial table (Fragility = %d).\n", result$fragility),
-            sprintf("This transfer of cases yields the following table:"))
+            sprintf("This transfer of cases yields the following table:")
              )
+             
           
         } else if (!final_solution$needtworows & final_solution$final_switch == 1) {
           #conclusion1 
@@ -541,10 +544,10 @@ test_sensitivity_ln <- function(est_eff,
           #conclusion1b 
           cat(paste(
             sprintf("This is equivalent to \ntransferring %d", final_solution$final_switch), 
-            c("case from"), transferway, 
+            c("case from"), transferway), 
             sprintf("\nin the initial table (Fragility = %d).\n", result$fragility),
             sprintf("This transfer of cases yields the following table:")
-          ))
+          )
           
         } else {
           #conclusion1
@@ -588,9 +591,9 @@ test_sensitivity_ln <- function(est_eff,
           #conclusion1b
           cat(paste(
             sprintf("This is equivalent \nto transferring %d", final_solution$final_switch), 
-            c("case from"), transferway, 
+            c("case from"), transferway), 
             sprintf("\nin the initial table (Fragility = %d).\n", result$fragility),
[…]
