+++ R/test_sensitivity_ln.R
@@ -313,10 +313,11 @@ test_sensitivity_ln <- function(est_eff,
     
   } else  if (to_return == "print") {
     
+    
     result <- list(conclusion1,conclusion1b, conclusion1c,
                    Implied_Table = final_solution$table_start, notice, Transfer_Table = final_solution$table_final,
                    conclusion2, conclusion3,
-                   total_RIR, total_switch)
+                   total_RIR = total_RIR, total_switch = total_switch)
     
     fragility <- result$Implied_Table["Treatment", "Fail"] - result$Transfer_Table["Treatment", "Fail"]
     
@@ -333,40 +334,42 @@ test_sensitivity_ln <- function(est_eff,
     notice <- result$notice
     RIR_value <- result$RIR
     fragility <- result$fragility
-    total_switch <- result$total_switch
+    
+    
+    fragility <- result$Implied_Table["Treatment", "Fail"] - result$Transfer_Table["Treatment", "Fail"]
     
     if (changeSE) {
-     cat(sprintf("RIR = %d\n\n", RIR))
+    
+      cat(sprintf("RIR = %d\n\n", RIR))
       cat("The table you entered or is implied by your effect size:\n\n")
       print(Implied_Table)
       cat("\n")
-      cat(paste(sprintf("(Effect size = %.3f, SE = %.3f, t-ratio = %.3f. Values have\n",
-                        final_solution$est_eff_final, final_solution$std_err_final, final_solution$t_final),
-                sprintf("been rounded to the nearest integer. This may cause a little\n"), 
-                sprintf("change to the estimated effect for the table. In order to\n"),
-                sprintf("generate a usable implied contingency table, we increased the\n"), 
-                sprintf("standard error to %.3f (the original one is %.3f)).\n\n", std_err, user_std_err))
-      )
+      cat(sprintf("(Effect size = %.3f, SE = %.3f, t-ratio = %.3f. Values have
+      been rounded to the nearest integer. This may cause a little
+      change to the estimated effect for the table.)\n\n", 
+                  final_solution$est_eff_start, final_solution$std_err_start, final_solution$t_start))
       
-      ### start here (changeSE = T, invalidate_ob = T)
-      
+      ### start here
       if (invalidate_ob) {
+
+      ### start here (changeSE = T, invalidate_ob = T)
+        
         change <- sprintf("To invalidate the inference that the effect is greater than 0 \n(alpha =%.2f)", alpha)
         ### 
         if (!final_solution$needtworows & final_solution$final_switch > 1) {
-          #conclusion1 
+          #conclusion1 <- 
           cat(paste(
             change, sprintf("you would need to replace %d", RIR), RIRway, "\ncases "))
           
           if (replace == "control") {
-            #conclusion1a 
+            #conclusion1a <- 
             cat(sprintf("with cases for which the probability of failure in the \ncontrol group applies (RIR = %d).", RIR))
           } else {
-            #conclusion1a 
+            #conclusion1a <- 
             cat(sprintf("with cases for which the probability of failure in the \nentire sample applies (RIR = %d).", RIR))
           }
           
-          #conclusion1b 
+          #conclusion1b <- 
           cat(paste(
             sprintf("This is equivalent to \ntransferring %d", total_switch), 
             c("case from"), transferway, 
@@ -375,7 +378,7 @@ test_sensitivity_ln <- function(est_eff,
           ))
           
         } else if (!final_solution$needtworows & final_solution$final_switch == 1) {
-          #conclusion1 
+          #conclusion1 <- 
           cat(paste(
             change, sprintf("you would need to replace %d", RIR), RIRway, "\ncases"))
           
[…]
