### 2023-11-21 f91400e
Update helper_output_print.R
