### 2023-11-21 af9e55d
Update test_sensitivity.R
Add "eff_thr" due to added condition in output_print() function
