### 2023-11-21 4706575
Update mkonfound.R
Manual modification
