+++ R/mkonfound.R
@@ -1,13 +1,19 @@
 #' Perform meta-analyses including sensitivity analysis
-#' @description For fitted models, this command carries out sensitivity analysis for a number of models, when their parameters stored in a data.frame.
-#' @param d data.frame or tibble with the t-statistics and associated degrees of freedom
+#' @description For fitted models, this command carries out sensitivity analysis
+#' for a number of models, when their parameters stored in a data.frame.
+#' @param d data.frame or tibble with the t-statistics and 
+#' associated degrees of freedom
 #' @param t t-statistic or vector of t-statistics
-#' @param df degrees of freedom or vector of degrees of freedom associated with the t-statistics in the t argument
-#' @param return_plot whether to return a plot of the percent bias; defaults to FALSE
+#' @param df degrees of freedom or vector of degrees of 
+#' freedom associated with the t-statistics in the t argument
+#' @param return_plot whether to return a plot of the percent bias; 
+#' defaults to FALSE
 #' @inheritParams pkonfound
 #' @import rlang
 #' @import dplyr
-#' @return prints the bias and the number of cases that would have to be replaced with cases for which there is no effect to invalidate the inference for each of the cases in the data.frame
+#' @return prints the bias and the number of cases that would have to be 
+#' replaced with cases for which there is no effect to invalidate the inference 
+#' for each of the cases in the data.frame
 #' @examples
 #' \dontrun{
 #' mkonfound_ex
@@ -25,18 +31,23 @@ mkonfound <- function(d, t, df, alpha = .05, tails = 2, return_plot = FALSE) {
   df_vec <- pull(select(d, !!df_enquo))
 
   if (length(t_vec) <= 1) {
-    stop("To carry out sensitivity analysis for a single study, use pkonfound()")
+    stop("To carry out sensitivity analysis 
+         for a single study, use pkonfound()")
   }
 
-  results_df <- suppressWarnings(purrr::map2_dfr(.x = t_vec, .y = df_vec, .f = core_sensitivity_mkonfound))
-
+  results_df <- suppressWarnings(
+  purrr::map2_dfr(
+    .x = t_vec, .y = df_vec, .f = core_sensitivity_mkonfound)
+)
   if (return_plot == TRUE) {
     results_df$action <- dplyr::case_when(
       results_df$action == "to_invalidate" ~ "To Invalidate",
       results_df$action == "to_sustain" ~ "To Sustain"
     )
 
-    p <- ggplot2::ggplot(results_df, ggplot2::aes_string(x = "pct_bias_to_change_inference", fill = "action")) +
+    p <- ggplot2::ggplot(results_df, ggplot2::aes_string(
+  x = "pct_bias_to_change_inference", fill = "action"
+)) +
       ggplot2::geom_histogram() +
       ggplot2::scale_fill_manual("", values = c("#1F78B4", "#A6CEE3")) +
       ggplot2::theme_bw() +
@@ -86,8 +97,15 @@ core_sensitivity_mkonfound <- function(t, df, alpha = .05, tails = 2) {
   r_con <- round(sqrt(abs(itcv)), 3)
 
   out <- dplyr::data_frame(t, df, action, inference, pct_bias, itcv, r_con)
-  names(out) <- c("t", "df", "action", "inference", "pct_bias_to_change_inference", "itcv", "r_con")
-  out$pct_bias_to_change_inference <- round(out$pct_bias_to_change_inference, 3)
+  names(out) <- c("t", 
+                  "df", 
+                  "action", 
+                  "inference", 
+                  "pct_bias_to_change_inference", 
+                  "itcv", 
+                  "r_con")
+  out$pct_bias_to_change_inference <- round(
[…]
