### 2023-11-21 6489298
Update helper_output_print.R
change eff_diff into est_eff due to argument error
