+++ R/helper_output_print.R
@@ -1,62 +1,62 @@
 # Function to output printed text
 
-output_print <- function(eff_diff, beta_threshhold, bias = NULL, sustain = NULL, nu, eff_thr, recase, obs_r, critical_r, r_con, itcv, alpha, index) {
+output_print <- function(est_eff, beta_threshhold, bias = NULL, sustain = NULL, nu, eff_thr, recase, obs_r, critical_r, r_con, itcv, alpha, index) {
   if (index == "RIR"){ 
     cat(crayon::bold("Robustness of Inference to Replacement (RIR):\n"))
-    if ((abs(eff_diff) > abs(beta_threshhold)) & is.na(eff_thr) == TRUE) {
+    if ((abs(est_eff) > abs(beta_threshhold)) & is.na(eff_thr) == TRUE) {
       cat("TO INVALIDATE:\n")
       cat("\n")
       cat(paste0("RIR = ", round(recase, 3), "\n"))
       cat("\n")
-      cat(paste0("You entered an estimated effect of ", round(eff_diff, 3), ". To invalidate"))
+      cat(paste0("You entered an estimated effect of ", round(est_eff, 3), ". To invalidate"))
       cat("\n")
       cat(paste0("the inference of an effect using the threshold of ", round(beta_threshhold, 3), " for"))
       cat("\n")
-      cat(paste0("statistical significance with alpha= ", alpha, " ,", round(bias, 3), " % of the (", round(eff_diff, 3), ")"))
+      cat(paste0("statistical significance with alpha= ", alpha, " ,", round(bias, 3), " % of the (", round(est_eff, 3), ")"))
       cat("\n")
       cat(paste0("estimate would have to be due to bias. This implies ", round(recase, 3)))
       cat("\n")
       cat("observations would have to be replaced with cases for which the\n")
       cat(paste0("effect is 0 (RIR = ", round(recase, 3), ").\n"))
       cat("\n")
-    } else if ((abs(eff_diff) > abs(beta_threshhold)) & is.na(eff_thr) == FALSE) {
+    } else if ((abs(est_eff) > abs(beta_threshhold)) & is.na(eff_thr) == FALSE) {
       cat("TO INVALIDATE:\n")
       cat("\n")
       cat(paste0("RIR = ", round(recase, 3), "\n"))
       cat("\n")
-      cat(paste0("You entered an effect of ", round(eff_diff, 3), ", and specified a threshold"))
+      cat(paste0("You entered an effect of ", round(est_eff, 3), ", and specified a threshold"))
       cat("\n")
       cat(paste0("value of ", round(eff_thr, 3), ". To invalidate the inference based on your estimate"))
       cat("\n")
-      cat(paste0(round(bias, 3), " % of the (", round(eff_diff, 3), ") estimate would have to be due to bias."))
+      cat(paste0(round(bias, 3), " % of the (", round(est_eff, 3), ") estimate would have to be due to bias."))
       cat("\n")
       cat(paste0("This implies ", round(recase, 3), " observations would have to be replaced with cases"))
       cat("\n")
       cat("for which the effect is", nu, "(RIR = ", round(recase, 3),").\n")
       cat("\n")
-    } else if ((abs(eff_diff) < abs(beta_threshhold)) & is.na(eff_thr) == TRUE) {
+    } else if ((abs(est_eff) < abs(beta_threshhold)) & is.na(eff_thr) == TRUE) {
       cat("TO SUSTAIN:\n", sep = "")
       cat("\n")
       cat(paste0("RIR = ", round(recase, 3), "\n"))
       cat("\n")
-      cat(paste0("You entered an estimated effect of ", round(eff_diff, 3), ". The threshold value for"))
[…]
