+++ R/test_sensitivity_ln.R
@@ -370,7 +370,7 @@ test_sensitivity_ln <- function(est_eff,
           cat(paste(
             sprintf("This is equivalent to \ntransferring %d", final_solution$final_switch), 
             c("case from"), transferway, 
-            sprintf("\nin the initial table (Fragility = %d).", fragility),
+            sprintf("\nin the initial table (Fragility = %d).", total_switch),
             sprintf("This transfer of cases \nyields the following table:")
           ))
           
@@ -391,7 +391,7 @@ test_sensitivity_ln <- function(est_eff,
           cat(paste(
             sprintf("This is equivalent to \ntransferring %d", final_solution$final_switch), 
             c("case from"), transferway, 
-            sprintf("\nin the initial table (Fragility = %d).", fragility),
+            sprintf("\nin the initial table (Fragility = %d).", total_switch),
             sprintf("This transfer of cases \nyields the following table:")
           ))
           
@@ -436,7 +436,7 @@ test_sensitivity_ln <- function(est_eff,
           cat(paste(
             sprintf("This is equivalent \nto transferring %d", final_solution$final_switch), 
             c("case from"), transferway, 
-            sprintf("\nin the initial table (Fragility = %d).", fragility),
+            sprintf("\nin the initial table (Fragility = %d).", total_switch),
             sprintf("This transfer of cases \nyields the following table:")
           ))
           
@@ -457,7 +457,7 @@ test_sensitivity_ln <- function(est_eff,
           cat(paste(
             sprintf("This is equivalent \nto transferring %d", final_solution$final_switch), 
             c("case from"), transferway, 
-            sprintf("\nin the initial table (Fragility = %d).", fragility),
+            sprintf("\nin the initial table (Fragility = %d).", total_switch),
             sprintf("This transfer of cases \nyields the following table:")
           ))
           
@@ -523,7 +523,7 @@ test_sensitivity_ln <- function(est_eff,
           cat(paste(
             sprintf("This is equivalent to \ntransferring %d", final_solution$final_switch), 
             c("case from"), transferway, 
-            sprintf("\nin the initial table (Fragility = %d).", fragility),
+            sprintf("\nin the initial table (Fragility = %d).", total_switch),
             sprintf("This transfer of cases \nyields the following table:")
           ))
           
@@ -544,7 +544,7 @@ test_sensitivity_ln <- function(est_eff,
           cat(paste(
             sprintf("This is equivalent to \ntransferring %d", final_solution$final_switch), 
             c("case from"), transferway, 
-            sprintf("\nin the initial table (Fragility = %d).", fragility),
+            sprintf("\nin the initial table (Fragility = %d).", total_switch),
             sprintf("This transfer of cases \nyields the following table:")
           ))
           
@@ -588,7 +588,7 @@ test_sensitivity_ln <- function(est_eff,
           cat(paste(
             sprintf("This is equivalent \nto transferring %d", final_solution$final_switch), 
             c("case from"), transferway, 
-            sprintf("\nin the initial table (Fragility = %d).", fragility),
+            sprintf("\nin the initial table (Fragility = %d).", total_switch),
             sprintf("This transfer of cases \nyields the following table:")
           ))
           
@@ -609,7 +609,7 @@ test_sensitivity_ln <- function(est_eff,
           cat(paste(
             sprintf("This is equivalent \nto transferring %d", final_solution$final_switch), 
             c("case from"), transferway, 
-            sprintf("\nin the initial table (Fragility = %d).", fragility),
+            sprintf("\nin the initial table (Fragility = %d).", total_switch),
             sprintf("This transfer of cases \nyields the following table:")
           ))
           
