### 2023-11-21 c4cc779
Update nonlinear_auxiliary.R
All manual modifications, except for the calculation formula
