+++ R/nonlinear_auxiliary.R
@@ -8,7 +8,7 @@ isinvalidate <- function(thr_t, ob_t) {
   return(x)
 }
 
-# to evaluate what kind of switches we need - increase or decrease the odds ratio
+#to evaluate what kind of switches we need-increase or decrease the odds ratio
 isdcroddsratio <- function(thr_t, ob_t) {
   if (thr_t < ob_t) {
     x <- T
@@ -18,7 +18,8 @@ isdcroddsratio <- function(thr_t, ob_t) {
   return(x)
 }
 
-# get the a cell (control failure) for given odds_ratio, se and treatment cases - first solution
+# get the a cell (control failure) for given odds_ratio, 
+# se and treatment cases - first solution
 get_a1_kfnl <- function(odds_ratio, std_err, n_obs, n_trm) {
   a1 <- -(1 / (2 * (1 + odds_ratio^2 + odds_ratio * (-2 + n_trm * std_err^2)))) *
     (2 * n_trm * (-1 + odds_ratio) * odds_ratio + n_trm^2 * odds_ratio * std_err^2 -
@@ -28,7 +29,8 @@ get_a1_kfnl <- function(odds_ratio, std_err, n_obs, n_trm) {
   return(a1)
 }
 
-# get the c cell (treatment failure) for given odds_ratio, se and treatment cases - first solution
+# get the c cell (treatment failure) for given odds_ratio, 
+# se and treatment cases - first solution
 get_c1_kfnl <- function(odds_ratio, std_err, n_obs, n_trm) {
   c1 <- -((-2 * n_trm + 2 * n_trm * odds_ratio - n_obs * n_trm * odds_ratio * std_err^2 +
     n_trm^2 * odds_ratio * std_err^2 +
@@ -38,7 +40,8 @@ get_c1_kfnl <- function(odds_ratio, std_err, n_obs, n_trm) {
   return(c1)
 }
 
-# get the a cell (control failure) for given odds_ratio, se and treatment cases - second solution
+# get the a cell (control failure) for given odds_ratio, 
+# se and treatment cases - second solution
 get_a2_kfnl <- function(odds_ratio, std_err, n_obs, n_trm) {
   a2 <- (1 / (2 * (1 + odds_ratio^2 + odds_ratio * (-2 + n_trm * std_err^2)))) *
     (-2 * n_trm * (-1 + odds_ratio) * odds_ratio - n_trm^2 * odds_ratio * std_err^2 +
@@ -48,7 +51,8 @@ get_a2_kfnl <- function(odds_ratio, std_err, n_obs, n_trm) {
   return(a2)
 }
 
-# get the c cell (treatment failure) for given odds_ratio, se and treatment cases - second solution
+# get the c cell (treatment failure) for given odds_ratio, 
+# se and treatment cases - second solution
 get_c2_kfnl <- function(odds_ratio, std_err, n_obs, n_trm) {
   c2 <- (2 * n_trm - 2 * n_trm * odds_ratio + n_obs * n_trm * odds_ratio * std_err^2 -
     n_trm^2 * odds_ratio * std_err^2 +
@@ -75,8 +79,10 @@ taylorexp <- function(a, b, c, d, q, thr) {
   d1square <- Num1 / Den1 - Num2 / Den2
   # d1unsquare is the first derivative of the unsquared term
   d1unsquare <- d1square / (2 * log(a * (d - q) / (b * (c + q))) / sqrt(1 / a + 1 / b + 1 / (c + q) + 1 / (d - q)))
-  # x is the number of cases need to be replaced solved based on the taylor expansion
-  # this is the (linear approximation) of the original/unsquared term around the value of q
+  # x is the number of cases need to be replaced 
[…]
