### 2023-11-21 50da8c9
Update pkonfound.R
Manual modification
