+++ R/pkonfound.R
@@ -1,34 +1,59 @@
 #' Perform sensitivity analysis for published studies
-#' @description For published studies, this command calculates (1) how much bias there must be in an estimate to invalidate/sustain an inference; (2) the impact of an omitted variable necessary to invalidate/sustain an inference for a regression coefficient.
-#' @param est_eff the estimated effect (such as an unstandardized beta coefficient or a group mean difference)
-#' @param std_err the standard error of the estimate of the unstandardized regression coefficient
+#' @description For published studies, this command calculates 
+#' (1) how much bias there must be in an estimate to invalidate/sustain 
+#' an inference; (2) the impact of an omitted variable necessary to 
+#' invalidate/sustain an inference for a regression coefficient.
+#' @param est_eff the estimated effect (such as an unstandardized beta 
+#' coefficient or a group mean difference)
+#' @param std_err the standard error of the estimate of the unstandardized 
+#' regression coefficient
 #' @param n_obs the number of observations in the sample
 #' @param n_covariates the number of covariates in the regression model
 #' @param alpha probability of rejecting the null hypothesis (defaults to 0.05)
-#' @param tails integer whether hypothesis testing is one-tailed (1) or two-tailed (2; defaults to 2)
-#' @param index whether output is RIR or IT (impact threshold); defaults to "RIR"
-#' @param nu what hypothesis to be tested; defaults to testing whether est_eff is significantly different from 0
-#' @param model_type the type of model being estimated; defaults to "ols" for a linear regression model; the other option is "logistic"
-#' @param n_treat the number of cases associated with the treatment condition; applicable only when model_type = "logistic"
-#' @param switch_trm whether to switch the treatment and control cases; defaults to FALSE; applicable only when model_type = "logistic"
-#' @param a cell is the number of cases in the control group showing unsuccessful results
-#' @param b cell is the number of cases in the control group showing successful results
-#' @param c cell is the number of cases in the treatment group showing unsuccessful results
-#' @param d cell is the number of cases in the treatment group showing successful results
-#' @param two_by_two_table table that is a matrix or can be coerced to one (data.frame, tibble, tribble) from which the a, b, c, and d arguments can be extracted
-#' @param test whether using Fisher's Exact Test or A chi-square test; defaults to Fisher's Exact Test
-#' @param replace whether using entire sample or the control group to calculate the base rate; default is the control group
+#' @param tails integer whether hypothesis testing is one-tailed (1) or 
+#' two-tailed (2; defaults to 2)
+#' @param index whether output is RIR or IT (impact threshold); 
+#' defaults to "RIR"
+#' @param nu what hypothesis to be tested; defaults to testing whether 
+#' est_eff is significantly different from 0
+#' @param model_type the type of model being estimated; defaults to "ols" for 
+#' a linear regression model; the other option is "logistic"
+#' @param n_treat the number of cases associated with the treatment condition; 
+#' applicable only when model_type = "logistic"
+#' @param switch_trm whether to switch the treatment and control cases; defaults
+#'  to FALSE; applicable only when model_type = "logistic"
+#' @param a cell is the number of cases in the control group showing 
+#' unsuccessful results
+#' @param b cell is the number of cases in the control group showing successful 
+#' results
+#' @param c cell is the number of cases in the treatment group showing 
+#' unsuccessful results
+#' @param d cell is the number of cases in the treatment group showing 
+#' successful results
[…]
