### 2023-11-21 76eb2a6
Update test_cop.R
Manual modification
