### 2023-11-21 b5e431d
Update test_pse.R
Manual modification
