+++ R/test_pse.R
@@ -17,13 +17,21 @@ test_pse <- function(est_eff,
     df = n_obs - n_covariates - 3
     
     ## error message if input is inappropriate
-    if (!(std_err > 0)) {stop("Did not run! Standard error needs to be greater than zero.")}
-    if (!(sdx > 0)) {stop("Did not run! Standard deviation of x needs to be greater than zero.")}
-    if (!(sdy > 0)) {stop("Did not run! Standard deviation of y needs to be greater than zero.")}
-    if (!(n_obs > n_covariates + 3)) {stop("Did not run! There are too few observations relative to the number of observations and covariates. Please specify a less complex model to use KonFound-It.")}
+    if (!(std_err > 0)) {stop("Did not run! Standard error needs 
+                              to be greater than zero.")}
+    if (!(sdx > 0)) {stop("Did not run! Standard deviation of x 
+                          needs to be greater than zero.")}
+    if (!(sdy > 0)) {stop("Did not run! Standard deviation of y 
+                          needs to be greater than zero.")}
+    if (!(n_obs > n_covariates + 3)) 
+       {stop("Did not run! There are too few observations relative to the 
+              number of observations and covariates. Please specify a less 
+          complex model to use KonFound-It.")}
     if (!(0 < R2)) {stop("Did not run! R2 needs to be greater than zero.")}
     if (!(R2 < 1)) {stop("Did not run! R2 needs to be less than one.")}
-    if (!(1-((sdy^2/sdx^2)*(1-R2)/((df+1) * std_err^2))>0)) {stop("Did not run! Entered values produced Rxz^2 <=0, consider adding more significant digits to your entered values.")}
+    if (!(1-((sdy^2/sdx^2)*(1-R2)/((df+1) * std_err^2))>0)) 
+       {stop("Did not run! Entered values produced Rxz^2 <=0, 
+              consider adding more significant digits to your entered values.")}
     
     
     ## now standardize 
@@ -51,19 +59,24 @@ test_pse <- function(est_eff,
     rxcvGz = as.numeric(Gz_pse[1])
     rycvGz = as.numeric(Gz_pse[2])
     
-    # convert conditional correlations to unconditional correlations to be used in new regression
+    # convert conditional correlations to unconditional 
+    # correlations to be used in new regression
     rxcv = rxcvGz * sqrt((1 - rcvz^2) * (1 - rxz^2)) + rxz * rcvz
     rycv = rycvGz * sqrt((1 - rcvz^2) * (1 - rzy^2)) + rzy * rcvz
     
-    verify_pse_reg_M3 = verify_reg_Gzcv(n_obs, sdx, sdy, sdz, sdcv, rxy, rxz, rzy, rycv, rxcv, rcvz)
-    verfiy_pse_manual_thr = verify_manual(rxy, rxz, rxcv, ryz, rycv, rzcv, sdy, sdx)
+    verify_pse_reg_M3 = verify_reg_Gzcv(n_obs, sdx, sdy, sdz, sdcv, rxy, rxz, 
+                                        rzy, rycv, rxcv, rcvz)
+    verfiy_pse_manual_thr = verify_manual(rxy, rxz, rxcv, ryz, rycv, 
+                                          rzcv, sdy, sdx)
     cov_pse = verify_pse_reg_M3[[11]]
     
     # prepare some other values in the final Table (long output)
     R2_M3 = as.numeric(verify_pse_reg_M3[1])
-    eff_x_M3 = as.numeric(verify_pse_reg_M3[2]) # should be equivalent or very close to eff_thr
+    eff_x_M3 = as.numeric(verify_pse_reg_M3[2]) 
+    # should be equivalent or very close to eff_thr
     se_x_M3 = as.numeric(verify_pse_reg_M3[3])
[…]
