### 2023-11-21 c049eef
Update test_sensitivity_ln.R
Manual modification
