### 2023-11-21 6991bf8
Update tkonfound.R
Manual modification
