+++ R/tkonfound.R
@@ -4,254 +4,276 @@ tkonfound <- function(a, b, c, d,
                       test = "fisher", 
                       replace = "control", 
                       to_return = to_return){
-  # a <- 35
-  # b <- 17
-  # c <- 17
-  # d <- 38
-  # alpha <- 0.05
-  # switch_trm <- T
-  # test <- "fisher"
-  
-  # stop message
-  if (a < 0 || b < 0 || c < 0 || d < 0) {
-    stop("Please enter non-negative integers for each cell.")
-  }
-  
-  if (a != as.integer(a) || b != as.integer(b) || c != as.integer(c) || d != as.integer(d)) {
-    stop("Please enter non-negative integers for each cell.")
-  }
-  
-  # use fisher if any of the cell is smaller than 5
-  if (a < 5 || b < 5 || c < 5 || d < 5){
-    test <- "fisher"
-  }
-  
-  # odds_ratio <- a*d/(b*c)
-  n_cnt <- a+b
-  n_trm <- c+d
-  n_obs <- n_cnt + n_trm
-  # est <- log(odds_ratio)
-  
-  # this is the 2 by 2 table we start with
-  table_ob <- matrix(c(a, b, c, d), byrow = TRUE, 2, 2)
-  
-  if (test == "fisher") {
-    p_ob <- fisher_p(a, b, c, d)
-    fisher_ob <- fisher_oddsratio(a, b, c, d)
-  }
-  if (test == "chisq") {
-    p_ob <- chisq_p(a, b, c, d)
-    chisq_ob <- chisq_value(a, b, c, d)
-  }
-  
-  # get solution
-  if (test == "chisq"){
-    solution <- getswitch_chisq(a, b, c, d, alpha, switch_trm)
-    chisq_final <- solution$chisq_final
-  }
-  
-  if (test == "fisher"){
-    solution <- getswitch_fisher(a, b, c, d, alpha, switch_trm)
-    fisher_final <- solution$fisher_final
-  }
-  
-  table_final <- solution$Transfer_Table
-  table_start <- table_ob
-  dcroddsratio_ob <- solution$dcroddsratio_ob
-  allnotenough <- solution$needtworows
-  final <- solution$final_switch
-  p_final <- solution$p_final
-  taylor_pred <- solution$taylor_pred
-  perc_bias_pred <- solution$perc_bias_pred
-  final_extra <- solution$final_extra
-  total_switch <- solution$total_switch
-  
-  ### add column and row names to contingency tables
-  rownames(table_start) <- rownames(table_final) <- c("Control", "Treatment")
-  colnames(table_start) <- colnames(table_final) <- c("Fail", "Success")
-  
-  if (switch_trm && dcroddsratio_ob) {
-    transferway <- "treatment success to treatment failure"
-    RIR <- ceiling(final/((a+c)/n_obs))*(replace=="entire") + ceiling(final/(a/(a+b)))*(1-(replace=="entire"))
-    RIRway <- "treatment success"
-  }
-  if (switch_trm && !dcroddsratio_ob) {
-    transferway <- "treatment failure to treatment success"
-    RIR <- ceiling(final/((b+d)/n_obs))*(replace=="entire") + ceiling(final/(b/(a+b)))*(1-(replace=="entire"))
-    RIRway <- "treatment failure"
-  }
-  if (!switch_trm && dcroddsratio_ob) {
-    transferway <- "control failure to control success"
-    RIR <- ceiling(final/((b+d)/n_obs))*(replace=="entire") + ceiling(final/(b/(a+b)))*(1-(replace=="entire"))
-    RIRway <- "control failure"
-  }
[…]
