### 2023-11-21 1b9157e
Update tkonfound_fig.R
Manual modification
