+++ R/zzz.R
@@ -1,8 +1,14 @@
-## quiets concerns (notes) of R CMD check re: the vars that are evaluated using non-standard evaluation
-if (getRversion() >= "2.15.1") utils::globalVariables(c("inference", "key", "replace_null_cases", "percent_bias", "val"))
+## quiets concerns (notes) of R CMD check re: the vars that are evaluated 
+## using non-standard evaluation
+if (getRversion() >= "2.15.1") 
+    utils::globalVariables(c("inference", "key", "replace_null_cases", 
+                             "percent_bias", "val"))
 
 .onAttach <- function(libname, pkgname) {
-  packageStartupMessage("Sensitivity analysis as described in Frank, Maroulis, Duong, and Kelcey (2013) and in Frank (2000).\nFor more information visit http://konfound-it.com.")
+    packageStartupMessage("Sensitivity analysis as described in Frank, 
+                          Maroulis, Duong, and Kelcey (2013) and 
+                          in Frank (2000).\nFor more information visit 
+                          http://konfound-it.com.")
 }
 
 #' Open interactive web application for konfound
@@ -11,8 +17,10 @@ if (getRversion() >= "2.15.1") utils::globalVariables(c("inference", "key", "rep
 #' @export
 
 launch_shiny <- function() {
-  utils::browseURL("http://konfound-it.com")
+    utils::browseURL("http://konfound-it.com")
 }
 
-# addresses concerns (notes) of R CMD check re: the vars that are evaluated using non-standard evaluation
-# if (getRversion() >= "2.15.1") utils::globalVariables(c("itcv", "term", "unstd_beta1", "var_name", "x", "y"))
+# addresses concerns (notes) of R CMD check re: the vars that are 
+# evaluated using non-standard evaluation
+# if (getRversion() >= "2.15.1") 
+# utils::globalVariables(c("itcv", "term", "unstd_beta1", "var_name", "x", "y"))
