+++ R/konfound.R
@@ -1,12 +1,22 @@
 #' Perform sensitivity analysis on fitted models
-#' @description For fitted models, this command calculates (1) how much bias there must be in an estimate to invalidate/sustain an inference; (2) the impact of an omitted variable necessary to invalidate/sustain an inference for a regression coefficient. Currently works for: models created with lm() (linear models).
+#' @description For fitted models, this command calculates 
+#' (1) how much bias there must be in an estimate to invalidate/sustain 
+#' an inference; (2) the impact of an omitted variable necessary to 
+#' invalidate/sustain an inference for a regression coefficient. 
+#' Currently works for: models created with lm() (linear models).
 #' @param model_object output from a model (currently works for: lm)
-#' @param tested_variable Variable associated with the unstandardized beta coefficient to be tested
+#' @param tested_variable Variable associated with the unstandardized beta 
+#' coefficient to be tested
 #' @inheritParams pkonfound
-#' @param index whether output is RIR or IT (impact threshold); defaults to "RIR"
-#' @param two_by_two whether or not the tested variable is a dichotomous variable in a GLM; if so, the 2X2 table approach is used; only works for single variables at present (so test_all = TRUE will return an error)
-#' @param test_all whether to carry out the sensitivity test for all of the coefficients (defaults to FALSE)
-#' @return prints the bias and the number of cases that would have to be replaced with cases for which there is no effect to invalidate the inference
+#' @param index whether output is RIR or IT (impact threshold); 
+#' defaults to "RIR"
+#' @param two_by_two whether or not the tested variable is a dichotomous 
+#' variable in a GLM; if so, the 2X2 table approach is used; only works for 
+#' single variables at present (so test_all = TRUE will return an error)
+#' @param test_all whether to carry out the sensitivity test for all of the 
+#' coefficients (defaults to FALSE)
+#' @return prints the bias and the number of cases that would have to be 
+#' replaced with cases for which there is no effect to invalidate the inference
 #' @importFrom rlang .data
 #' @examples
 #' # using lm() for linear models
@@ -32,7 +42,8 @@
 #'   konfound(m3, Days)
 #' }
 #' 
-#' m4 <- glm(outcome ~ condition, data = binary_dummy_data, family = binomial(link = "logit"))
+#' m4 <- glm(outcome ~ condition, data = binary_dummy_data, 
+#' family = binomial(link = "logit"))
 #' konfound(m4, condition, two_by_two = TRUE, n_treat = 55)
 #' 
 
@@ -49,120 +60,125 @@ konfound <- function(model_object,
                      n_treat = NULL,
                      switch_trm = TRUE,
                      replace = "control") {
-  
-  # Stop messages
-  if (!(class(model_object)[1] %in% c("lm", "glm", "lmerMod"))) {
-    stop("konfound() is currently implemented for models estimated with
-         lm(), glm(), and lme4::lmer(); consider using pkonfound() instead")
-  }
-  
-  if ("table" %in% to_return & test_all == TRUE){
-    stop("cannot return a table when test_all is set to TRUE")
-  }
-  # Dealing with non-standard evaluation
-
-  # dealing with non-standard evaluation
-  #(so unquoted names for tested_variable can be used)
-  tested_variable_enquo <- rlang::enquo(tested_variable) 
-  tested_variable_string <- rlang::quo_name(tested_variable_enquo)
-  
-  # Dispatching based on class
-  if (class(model_object)[1] == "lm") {
-    output <- konfound_lm(
-      model_object = model_object,
-      tested_variable_string = tested_variable_string,
-      test_all = test_all,
-      alpha = alpha,
-      tails = tails,
-      index = index,
[…]
