### 2023-11-21 4b6cafc
Update cop_pse_auxiliary.R
"R:222:235
R:248
over 80 characters cause of calculation formula"
