+++ R/cop_pse_auxiliary.R
@@ -21,9 +21,17 @@ cal_rxy <- function(ryxGz, rxz, ryz){
     return(rxy)
 }
 
-cal_delta_star <- function(FR2max, R2, R2_uncond, est_eff, eff_thr, var_x, var_y, est_uncond, rxz, n_obs){
+cal_delta_star <- function(FR2max, 
+                           R2, R2_uncond, 
+                           est_eff, 
+                           eff_thr, 
+                           var_x, 
+                           var_y, 
+                           est_uncond, 
+                           rxz, 
+                           n_obs){
     if (FR2max > .99) {FR2max = .99}
-    # if (FR2max < R2 + inci) {FR2max = R2 + inci} check with Ken what this means
+    # if (FR2max < R2 + inci) {FR2max = R2 + inci}check with Ken what this means
     if (FR2max > R2) {D = sqrt(FR2max - R2)}
     
     #elements for computing Oster's delta_star
@@ -35,7 +43,8 @@ cal_delta_star <- function(FR2max, R2, R2_uncond, est_eff, eff_thr, var_x, var_y
     t_x = var_x * (n_obs / (n_obs - 1)) * (1 - rxz^2)
     ## adjust df for var_x 
     ## var_x is population variance, need sample variance from x
-    ## this adjustment is to get closer to what robomit generates as they run regression using the sample data 
+    ## this adjustment is to get closer to what robomit generates as they 
+    ## run regression using the sample data 
     num1 = bt_m_b * rt_m_ro_t_syy * t_x
     num2 = bt_m_b * var_x * t_x * b0_m_b1^2
     num3 = 2 * bt_m_b^2 * (t_x * b0_m_b1 * var_x)
@@ -115,8 +124,8 @@ verify_reg_Gzcv = function(n_obs, sdx, sdy, sdz, sdcv,
     flag_cov <- tryCatch(
         expr = {
             lavaan::sem(model, 
-                sample.cov = cov.matrix, 
-                sample.nobs = n_obs)
+                        sample.cov = cov.matrix, 
+                        sample.nobs = n_obs)
         },
         error = function(e){
             flag_cov = F
@@ -130,18 +139,24 @@ verify_reg_Gzcv = function(n_obs, sdx, sdy, sdz, sdcv,
     #if model can be run to verify true delta, then run it can save results
     if (class(flag_cov) == "lavaan") {
         fit <- lavaan::sem(model,
-                      sample.cov = cov.matrix,
-                      sample.nobs = n_obs)
+                           sample.cov = cov.matrix,
+                           sample.nobs = n_obs)
         ## the R2 extracted from summary is NOT right, do the calculation below
         R2 <- (sdy^2 - lavaan::parameterEstimates(fit)[4,]$est) / sdy^2
-        betaX <- lavaan::parameterEstimates(fit)[lavaan::parameterEstimates(fit)$label == 'beta1',]$est
-        seX <- lavaan::parameterEstimates(fit)[lavaan::parameterEstimates(fit)$label == 'beta1',]$se
-        betaZ <- lavaan::parameterEstimates(fit)[lavaan::parameterEstimates(fit)$label == 'beta2',]$est
-        seZ <- lavaan::parameterEstimates(fit)[lavaan::parameterEstimates(fit)$label == 'beta2',]$se
-        betaCV <- lavaan::parameterEstimates(fit)[lavaan::parameterEstimates(fit)$label == 'beta3',]$est
-        seCV <- lavaan::parameterEstimates(fit)[lavaan::parameterEstimates(fit)$label == 'beta3',]$se
+        betaX <- lavaan::parameterEstimates(fit)[
[…]
