### 2023-11-21 39f3418
Update mkonfound-data.R
Manual modification: no more than 80 characters per line of code
