### 2023-11-28 #5779 @jrosen48
We are actively working on both the patches / additions and the edits to the paper. Aim to be completed/resubmit by EOD on 12/12. Is that okay?
