### 2023-11-28 2fbcab4
add CONTRIBUTING and update README; also change email for JR
