+++ README.Rmd
@@ -114,4 +114,4 @@ We prefer for issues to be filed via GitHub (link to the issues page for `konfou
 
 ### Code of Conduct
 
-Please note that this project is released with a Contributor Code of Conduct available at [https://www.contributor-covenant.org/version/1/0/0/](https://www.contributor-covenant.org/version/1/0/0/)
+Please note that the konfound project is released with a [Contributor Code of Conduct](https://contributor-covenant.org/version/2/1/CODE_OF_CONDUCT.html). By contributing to this project, you agree to abide by its terms.
