### 2023-11-28 2ab5d4a
add CONTRIBUTING and update README; also change email for S narvaiz
