+++ joss-paper/paper.bib
@@ -12,6 +12,16 @@
 	keywords = {Decision making, Machine learning, Mathematical modeling, Model robustness, Model validation and verification, Policy support, Sensitivity analysis, Uncertainty quantification}
 }
 
+@article{frank2023,
+  title={Quantifying the robustness of causal inferences: Sensitivity analysis for pragmatic social science},
+  author={Frank, Kenneth A and Lin, Qinyun and Xu, Ran and Maroulis, Spiro and Mueller, Anna},
+  journal={Social Science Research},
+  volume={110},
+  pages={102815},
+  year={2023},
+  publisher={Elsevier}
+}
+
 @article{cinelli2019,
     author = {{Cinelli}, C. and {Hazlett}, C.},
     title = "{Making Sense of Sensitivity: Extending Omitted Variable Bias}",
@@ -105,15 +115,12 @@ journal = {Social Science Research},
 publisher = {Academic Press Inc.}
 }
 
-@article{frank2022, 
-title={Improving Oster’s δ*: Exact calculation for the coefficient of proportionality without subjective specification of a baseline model},
-url={https://papers.ssrn.com/sol3/papers.cfm?abstract_id=4305243}, 
-journal={SSRN}, 
-author={{Frank}, K.~A. and {Lin}, Q. and {Maroulis}, S. and {Dai}, S. and {Jess}, N. and {Lin}, H. and {Liu}, Y. and {Maestrales}, S. and {Searle}, E. and {Tait}, J.}, 
-year=2022, 
-month=dec,
-doi = {10.2139/ssrn.4305243}
-} 
+@article{frank2022,
+  title={Improving Oster’s $\delta$*: Exact Calculation for the Coefficient of Proportionality Without Subjective Specification of a Baseline Model},
+  author={Frank, Ken and Lin, Qinyun and Maroulis, Spiro and Dai, Shimeng and Jess, Nicole and Lin, Hung-chang and Maestrales, Sarah and Searle, Ellen and Tait, Jordan and others},
+  journal={Ellen and Tait, Jordan, Improving Oster’s $\delta$*: Exact Calculation for the Coefficient of Proportionality Without Subjective Specification of a Baseline Model (December 16, 2022)},
+  year={2022}
+}
 
 @article{xu2019,
 author = {{Xu}, R. and {Frank}, K.~A. and {Maroulis}, S.~J. and {Rosenberg}, J.~M.},
