+++ joss-paper/paper.md
@@ -7,6 +7,7 @@ tags:
 - R
 date: "31 July 2023"
 output:
+  pdf_document: default
   html_document:
     df_print: paged
 authors:
@@ -45,7 +46,9 @@ affiliations:
 
 # Quantifying the Robustness of Inferences
 
-Statistical methods which quantify the conditions necessary to alter inferences are important to a variety of disciplines [@razavi2021]. One line of work is rooted in linear models and foregrounds the sensitivity of inferences to the strength of omitted variables [@frank2000; @cinelli2019]. A more recent approach is rooted in the potential outcomes framework for causal inference and foregrounds how hypothetical changes in a sample would alter an inference if such variables or cases were otherwise observed [@frank2007; @frank2008; @frank2013; @xu2019). We have implemented two measures from these lines of work within R via the `konfound` R package. One measure is the Impact Threshold of a Confounding Variable (ITCV), which generates statements such as "to invalidate an inference of an effect, an omitted variable would have to be correlated at \_\_ with the predictor of interest and with the outcome" [@frank2000]. This sensitivity analysis can be calculated for any linear model. The second measure is the Robustness of an Inference to Replacement (RIR) which generates statements such as "to invalidate the inference, \_\_ % of the cases would have to be replaced with counterfactual cases with zero effect of the treatment" [@frank2013]. The RIR represents a more generally applicable approach not limited to linear models, and is recommended for all cases that use binary outcomes [@frank2021].
+Statistical methods which quantify the conditions necessary to alter inferences are important to a variety of disciplines [@razavi2021]. One line of work is rooted in linear models and foregrounds the sensitivity of inferences to the strength of omitted variables [@frank2000; @cinelli2019]. A more recent approach is rooted in the potential outcomes framework for causal inference and foregrounds how hypothetical changes in a sample would alter an inference if such variables or cases were otherwise observed [@frank2007; @frank2008; @frank2013; @xu2019). **It differs from other sensitivity analysis approaches in the variety of dependent variable types it accomodates, its internal logic, and its analytic (rather than simulation-based) approach. [@frank2023].**
+
+We have implemented two measures from these lines of work within R via the `konfound` R package. **The audience for this R package is primarily social scientists, as well as interested individuals in other fields**. One measure is the Impact Threshold of a Confounding Variable (ITCV), which generates statements such as "to invalidate an inference of an effect, an omitted variable would have to be correlated at \_\_ with the predictor of interest and with the outcome" [@frank2000]. This sensitivity analysis can be calculated for any linear model. The second measure is the Robustness of an Inference to Replacement (RIR) which generates statements such as "to invalidate the inference, \_\_ % of the cases would have to be replaced with counterfactual cases with zero effect of the treatment" [@frank2013]. The RIR represents a more generally applicable approach not limited to linear models, and is recommended for all cases that use binary outcomes [@frank2021].
 
 # Statement of Need: The Need for an R Package
 
@@ -65,7 +68,8 @@ We will model the impact the following variables have on household water consump
 
     library(konfound)
 
-    ## Sensitivity analysis as described in Frank, Maroulis, Duong, and Kelcey (2013) and in Frank (2000).
+    ## Sensitivity analysis as described in Frank, Maroulis, Duong, and Kelcey 
+    ## (2013) and in Frank (2000).
     ## For more information visit http://konfound-it.com.
 
     m <- lm(water81 ~ water80 + income + educat + retire + peop80, data = concord1)
@@ -104,23 +108,27 @@ Now let's examine the robustness of the `peop80` effect by calculating the ITCV.
[…]
