### 2023-11-28 konfound-project/konfound#54 @jrosen48
thank you, addressed in https://github.com/konfound-project/konfound/commit/2fbcab46c1efedae07c59acd53a8c6fb027d9d45
