### 2023-11-28 konfound-project/konfound#50 @jrosen48
@fabian-s do you know of any clever way to identify the functions imported from other packages? I've tried pkgnet and lintr to no avail. Of course, a manual approach may be best - planning to do that.

I think we've addressed the other issues; will detail how and in which commits shortly.
