### 2023-11-29 konfound-project/konfound#50 @fabian-s
sounds dumb but should work:

- remove all Imports/Depends/Suggests from the DESCRIPTION
- run R CMD CHECK
- note all the "no visible global function definition for" errors that pop up and add import statements for those

alternatively, run `codetools::findGlobals` on each of your package's functions.
