### 2023-11-29 konfound-project/konfound#50 @jrosen48
For some reason, after removing imports and suggests, rebuilding, and checking, I am only seeing:

```
Namespace dependencies missing from DESCRIPTION Imports/Depends entries:
  'dplyr', 'rlang'
```

When there are far more.
