### 2023-11-30 f0d5cc5
Update helper_output_print.R
