+++ R/helper_output_print.R
@@ -14,10 +14,10 @@ output_print <- function(est_eff, beta_threshhold, bias = NULL, sustain = NULL, 
       cat("\n")
       cat(paste0("statistical significance with alpha= ", alpha, " ,", round(bias, 3), " % of the (", round(est_eff, 3), ")"))
       cat("\n")
-      cat(paste0("estimate would have to be due to bias. This implies ", round(recase, 3)))
+      cat(paste0("estimate would have to be due to bias. This implies that", round(recase, 3)))
       cat("\n")
-      cat("observations would have to be replaced with cases for which the\n")
-      cat(paste0("effect is 0 (RIR = ", round(recase, 3), ").\n"))
+      cat("to change invalidate the inference one would expect to have to replace\n")
+      cat(paste0(round(recase, 3), "observations with cases for which the treatment effect is 0 (RIR = ", round(recase, 3), ").\n"))
       cat("\n")
     } else if ((abs(est_eff) > abs(beta_threshhold)) & is.na(eff_thr) == FALSE) {
       cat("TO INVALIDATE:\n")
@@ -30,9 +30,9 @@ output_print <- function(est_eff, beta_threshhold, bias = NULL, sustain = NULL, 
       cat("\n")
       cat(paste0(round(bias, 3), " % of the (", round(est_eff, 3), ") estimate would have to be due to bias."))
       cat("\n")
-      cat(paste0("This implies ", round(recase, 3), " observations would have to be replaced with cases"))
+      cat(paste0("This implies that to invalidate the inference one would expect to have to replace "))
       cat("\n")
-      cat("for which the effect is", nu, "(RIR = ", round(recase, 3),").\n")
+      cat(round(recase, 3), "observations with cases for which the treatment effect is ", nu, "(RIR = ", round(recase, 3),").\n")
       cat("\n")
     } else if ((abs(est_eff) < abs(beta_threshhold)) & is.na(eff_thr) == TRUE) {
       cat("TO SUSTAIN:\n", sep = "")
@@ -45,9 +45,9 @@ output_print <- function(est_eff, beta_threshhold, bias = NULL, sustain = NULL, 
       cat("\n")
       cat(paste0("threshold, ", round(bias, 3), "% of the (", round(est_eff, 3), ") estimate would have to be due to"))
       cat("\n")
-      cat("bias. This implies", round(recase, 3), "observations with", nu, "effect would have to")
+      cat("bias. This implies that to sustain the inference on would expect to have")
       cat("\n")
-      cat(paste0("be replaced with cases for which the effect is ", round(beta_threshhold, 3), " (RIR = ", round(recase, 3), ")."))
+      cat(paste0("to replace ", round(recase, 3), "observation with effect of", nu, "with cases with effect of ", round(beta_threshhold, 3), "(RIR = ", round(recase, 3), ")."))
       cat("\n")
     } else if ((abs(est_eff) < abs(beta_threshhold)) & is.na(eff_thr) == FALSE) {
       cat("TO SUSTAIN:\n", sep = "")
@@ -58,11 +58,11 @@ output_print <- function(est_eff, beta_threshhold, bias = NULL, sustain = NULL, 
       cat("\n")
       cat(paste0("value of ", round(eff_thr, 3), ". To reach that threshold, ", round(bias, 3), "% of the (", round(est_eff, 3), ")"))
       cat("\n")
-      cat(paste0("estimate would have to be due to bias. This implies ", round(recase, 3)))
+      cat(paste0("estimate would have to be due to bias. This implies that"))
       cat("\n")
-      cat("observations with", nu, "effect would have to be replaced with cases")
+      cat("to sustain the inference one would expect to have to replace ")
       cat("\n")
[…]
