### 2023-11-30 6cf8cb6
Update helper_output_print.R
