+++ R/helper_output_print.R
@@ -12,13 +12,15 @@ output_print <- function(est_eff, beta_threshhold, bias = NULL, sustain = NULL, 
       cat("\n")
       cat(paste0("the inference of an effect using the threshold of ", round(beta_threshhold, 3), " for"))
       cat("\n")
-      cat(paste0("statistical significance with alpha= ", alpha, " ,", round(bias, 3), " % of the (", round(est_eff, 3), ")"))
+      cat(paste0("statistical significance with alpha = ", alpha, " ,", round(bias, 3), "% of the"))
       cat("\n")
-      cat(paste0("estimate would have to be due to bias. This implies that", round(recase, 3)))
+      cat(paste0("(", round(est_eff, 3), ") estimate would have to be due to bias. This implies that"))
       cat("\n")
-      cat("to change invalidate the inference one would expect to have to replace\n")
-      cat(paste0(round(recase, 3), "observations with cases for which the treatment effect is 0 (RIR = ", round(recase, 3), ").\n"))
-      cat("\n")
+      cat("to invalidate the inference one would expect to have to\n")
+      cat(paste0("replace ", round(recase, 3), " observations with cases for which the treatment effect is ", nu))
+      cat("\n")    
+      cat(paste0("(RIR = ", round(recase, 3), ").\n"))
+      cat("\n")    
     } else if ((abs(est_eff) > abs(beta_threshhold)) & is.na(eff_thr) == FALSE) {
       cat("TO INVALIDATE:\n")
       cat("\n")
