+++ R/helper_output_print.R
@@ -14,12 +14,12 @@ output_print <- function(est_eff, beta_threshhold, bias = NULL, sustain = NULL, 
       cat("\n")
       cat(paste0("statistical significance with alpha = ", alpha, " ,", round(bias, 3), "% of the"))
       cat("\n")
-      cat(paste0("(", round(est_eff, 3), ") estimate would have to be due to bias. This implies that"))
+      cat(paste0("(", round(est_eff, 3), ") estimate would have to be due to bias. This implies"))
       cat("\n")
-      cat("to invalidate the inference one would expect to have to\n")
-      cat(paste0("replace ", round(recase, 3), " observations with cases for which the treatment effect is ", nu))
+      cat("that to invalidate the inference one would expect to have to\n")
+      cat(paste0("replace ", round(recase, 3), " observations with cases for which the treatment"))
       cat("\n")    
-      cat(paste0("(RIR = ", round(recase, 3), ").\n"))
+      cat(paste0("effect is ", nu, "(RIR = ", round(recase, 3), ").\n"))
       cat("\n")    
     } else if ((abs(est_eff) > abs(beta_threshhold)) & is.na(eff_thr) == FALSE) {
       cat("TO INVALIDATE:\n")
@@ -32,10 +32,12 @@ output_print <- function(est_eff, beta_threshhold, bias = NULL, sustain = NULL, 
       cat("\n")
       cat(paste0(round(bias, 3), " % of the (", round(est_eff, 3), ") estimate would have to be due to bias."))
       cat("\n")
-      cat(paste0("This implies that to invalidate the inference one would expect to have to replace "))
+      cat(paste0("This implies that to invalidate the inference one would expect"))
       cat("\n")
-      cat(round(recase, 3), "observations with cases for which the treatment effect is ", nu, "(RIR = ", round(recase, 3),").\n")
+      cat("to have to replace ", round(recase, 3), "observations with cases for which")
       cat("\n")
+      cat(paste0("the treatment effect is ", nu, "(RIR = ", round(recase, 3), ").\n"))
+      cat("\n") 
     } else if ((abs(est_eff) < abs(beta_threshhold)) & is.na(eff_thr) == TRUE) {
       cat("TO SUSTAIN:\n", sep = "")
       cat("\n")
