### 2023-11-30 98350d3
Update helper_output_print.R
