+++ R/test_sensitivity_ln.R
@@ -147,22 +147,22 @@ test_sensitivity_ln <- function(est_eff,
   d <- final_solution$table_start[2,2]
   
   if (switch_trm && dcroddsratio_ob) {
-    transferway <- "treatment success to treatment failure,"
+    transferway <- "treatment success to treatment failure"
     RIR <- ceiling(final/((a+c)/n_obs))*(replace=="entire") + ceiling(final/(a/(a+b)))*(1-(replace=="entire"))
     RIRway <- "treatment success"
   }
   if (switch_trm && !dcroddsratio_ob) {
-    transferway <- "treatment failure to treatment success,"
+    transferway <- "treatment failure to treatment success"
     RIR <- ceiling(final/((b+d)/n_obs))*(replace=="entire") + ceiling(final/(b/(a+b)))*(1-(replace=="entire"))
     RIRway <- "treatment failure"
   }
   if (!switch_trm && dcroddsratio_ob) {
-    transferway <- "control failure to control success,"
+    transferway <- "control failure to control success"
     RIR <- ceiling(final/((b+d)/n_obs))*(replace=="entire") + ceiling(final/(b/(a+b)))*(1-(replace=="entire"))
     RIRway <- "control failure"
   }
   if (!switch_trm && !dcroddsratio_ob) {
-    transferway <- "control success to control failure,"
+    transferway <- "control success to control failure"
     RIR <- ceiling(final/((a+c)/n_obs))*(replace=="entire") + ceiling(final/(a/(a+b)))*(1-(replace=="entire"))
     RIRway <- "control success"
   }
@@ -170,25 +170,25 @@ test_sensitivity_ln <- function(est_eff,
   if (final_solution$needtworows) {
     final_extra <- final_solution$final_extra
     if (switch_trm && dcroddsratio_ob) {
-      transferway_extra <- "control failure to control success,"
+      transferway_extra <- "control failure to control success"
       RIR_extra <- ceiling(final_extra/((b+d)/n_obs))*(replace=="entire") + 
         ceiling(final_extra/(b/(b+d)))*(1-(replace=="entire"))
       RIRway_extra <- "control failure"
     }
     if (switch_trm && !dcroddsratio_ob) {
-      transferway_extra <- "control success to control failure,"
+      transferway_extra <- "control success to control failure"
       RIR_extra <- ceiling(final_extra/((a+c)/n_obs))*(replace=="entire") +
         ceiling(final_extra/(a/(a+b)))*(1-(replace=="entire"))
       RIRway_extra <- "control success"
     }
     if (!switch_trm && dcroddsratio_ob) {
-      transferway_extra <- "treatment success to treatment failure,"
+      transferway_extra <- "treatment success to treatment failure"
       RIR_extra <- ceiling(final_extra/((a+c)/n_obs))*(replace=="entire") +
         ceiling(final_extra/(a/(a+b)))*(1-(replace=="entire"))
       RIRway_extra <- "treatment success"
     }
     if (!switch_trm && !dcroddsratio_ob) {
-      transferway_extra <- "treatment failure to treatment success,"
+      transferway_extra <- "treatment failure to treatment success"
       RIR_extra <- ceiling(final_extra/((b+d)/n_obs))*(replace=="entire") +
         ceiling(final_extra/(b/(b+d)))*(1-(replace=="entire"))
       RIRway_extra <- "treatment failure"
@@ -265,7 +265,7 @@ test_sensitivity_ln <- function(est_eff,
   
   if (changeSE) {
     notice_SE <- sprintf(
-      "In order to generate a usable implied contingency table, we increased the standard error to %.3f (the original one is %.3f).",
+      "In order to generate a usable implied contingency table, we increased the standard error to %.3f (the reported standard error is %.3f).",
       std_err, user_std_err)
   }
   
[…]
