### 2023-11-30 konfound-project/konfound#50 @fabian-s
yeah, sorry about that, -- need to delete the NAMESPACE directives as well. if i do so and run

`rcmdcheck::rcmdcheck(".", build_args = "--no-build-vignettes")`

I see:

```
✔  checking startup messages can be suppressed (341ms)
W  checking dependencies in R code ...
   '::' or ':::' imports not declared from:
     ‘broom’ ‘broom.mixed’ ‘crayon’ ‘dplyr’ ‘ggplot2’ ‘ggrepel’ ‘lavaan’
     ‘lme4’ ‘margins’ ‘pbkrtest’ ‘purrr’ ‘rlang’ ‘tidyr’
✔  checking S3 generic/method consistency ...
✔  checking replacement functions ...
✔  checking foreign function calls ...
N  checking R code for possible problems (4.8s)
   chisq_p: no visible global function definition for ‘chisq.test’
   chisq_value: no visible global function definition for ‘chisq.test’
   fisher_oddsratio: no visible global function definition for
     ‘fisher.test’
   fisher_p: no visible global function definition for ‘fisher.test’
   konfound_glm: no visible binding for global variable ‘.data’
   konfound_lm: no visible binding for global variable ‘.data’
   konfound_lmer: no visible global function definition for ‘filter’
   konfound_lmer: no visible binding for global variable ‘.data’
   konfound_lmer: no visible global function definition for ‘bind_cols’
   mkonfound: no visible global function definition for ‘enquo’
   mkonfound: no visible global function definition for ‘pull’
   mkonfound: no visible global function definition for ‘select’
   test_cop: no visible binding for global variable ‘ModelLabel’
   test_cop: no visible binding for global variable ‘coef_X’
     .data ModelLabel bind_cols chisq.test coef_X enquo filter fisher.test
     pull select
   Consider adding
     importFrom("stats", "chisq.test", "filter", "fisher.test")
   to your NAMESPACE file.
✔  checking Rd files ...
✔  checking Rd metadata ...
✔  checking Rd cross-references ...
✔  checking for missing documentation entries ...
W  checking for code/documentation mismatches ...
   Codoc mismatches from documentation object 'tkonfound_fig':
   tkonfound_fig
     Code: function(a, b, c, d, thr_p = 0.05, switch_trm = TRUE, test =
                    "fisher", replace = "control")
     Docs: function(a, b, c, d, thr_p = 0.05, switch_trm = T, test =
                    "fisher", replace = "control")
     Mismatches in argument default values:
       Name: 'switch_trm' Code: TRUE Docs: T

✔  checking Rd \usage sections (662ms)
✔  checking Rd contents ...
W  checking for unstated dependencies in examples ...
   '::' or ':::' imports not declared from:
     ‘forcats’ ‘lme4’
   'library' or 'require' calls not declared from:
     ‘forcats’ ‘lme4’
```

proceed from there, I guess, but it seems to me most packages you import now are not actually needed (maybe for the vignettes, but then they should go into "Suggests:", I think)
