### 2023-12-01 bd9d7c4
Update helper_output_print.R
