### 2023-12-01 c590701
Update helper_output_print.R
