+++ R/helper_output_print.R
@@ -47,7 +47,7 @@ output_print <- function(est_eff, beta_threshhold, bias = NULL, sustain = NULL, 
       cat("\n")
       cat(paste0("statistical significance is ", round(beta_threshhold, 3), " (alpha = ", alpha, "). To reach that"))
       cat("\n")
-      cat(paste0("threshold, ", round(bias, 3), "% of the (", round(est_eff, 3), ") estimate would have to be due to"))
+      cat(paste0("threshold, ", round(sustain, 3), "% of the (", round(est_eff, 3), ") estimate would have to be due to"))
       cat("\n")
       cat("bias. This implies that to sustain the inference on would expect to have")
       cat("\n")
@@ -60,7 +60,7 @@ output_print <- function(est_eff, beta_threshhold, bias = NULL, sustain = NULL, 
       cat("\n")
       cat(paste0("You entered an effect size of ", round(est_eff, 3), ", and specified a threshold"))
       cat("\n")
-      cat(paste0("value of ", round(eff_thr, 3), ". To reach that threshold, ", round(bias, 3), "% of the (", round(est_eff, 3), ")"))
+      cat(paste0("value of ", round(eff_thr, 3), ". To reach that threshold, ", round(sustain, 3), "% of the (", round(est_eff, 3), ")"))
       cat("\n")
       cat(paste0("estimate would have to be due to bias. This implies that"))
       cat("\n")
