### 2023-12-01 a79c816
Update helper_output_print.R
