### 2023-12-01 4ee215e
Update helper_output_print.R
