### 2023-12-01 502fff6
Update helper_output_print.R
