+++ R/test_sensitivity_ln.R
@@ -372,8 +372,8 @@ test_sensitivity_ln <- function(est_eff,
           cat(paste(
             sprintf("This is equivalent to \ntransferring %d", final_solution$final_switch), 
             c("case from"), transferway, 
-            sprintf("\nin the initial table (Fragility = %d).", total_switch),
-            sprintf("This transfer of cases \nyields the following table:")
+            sprintf("\n(Fragility = %d).", total_switch),
+            sprintf("This transfer of cases yields the following table:")
           ))
           
         } else if (!final_solution$needtworows & final_solution$final_switch == 1) {
@@ -393,8 +393,8 @@ test_sensitivity_ln <- function(est_eff,
           cat(paste(
             sprintf("This is equivalent to \ntransferring %d", final_solution$final_switch), 
             c("case from"), transferway, 
-            sprintf("\nin the initial table (Fragility = %d).", total_switch),
-            sprintf("This transfer of cases \nyields the following table:")
+            sprintf("\ntable (Fragility = %d).", total_switch),
+            sprintf("This transfer of cases yields the following table:")
           ))
           
         } else {
@@ -438,8 +438,8 @@ test_sensitivity_ln <- function(est_eff,
           cat(paste(
             sprintf("This is equivalent \nto transferring %d", final_solution$final_switch), 
             c("case from"), transferway, 
-            sprintf("\nin the initial table (Fragility = %d).", total_switch),
-            sprintf("This transfer of cases \nyields the following table:")
+            sprintf("\ntable (Fragility = %d).", total_switch),
+            sprintf("This transfer of cases yields the following table:")
           ))
           
         } else if (!final_solution$needtworows & final_solution$final_switch == 1) {
@@ -459,8 +459,8 @@ test_sensitivity_ln <- function(est_eff,
           cat(paste(
             sprintf("This is equivalent \nto transferring %d", final_solution$final_switch), 
             c("case from"), transferway, 
-            sprintf("\nin the initial table (Fragility = %d).", total_switch),
-            sprintf("This transfer of cases \nyields the following table:")
+            sprintf("\ntable (Fragility = %d).", total_switch),
+            sprintf("This transfer of cases yields the following table:")
           ))
           
         } else {
@@ -525,8 +525,8 @@ test_sensitivity_ln <- function(est_eff,
           cat(paste(
             sprintf("This is equivalent to \ntransferring %d", final_solution$final_switch), 
             c("case from"), transferway, 
-            sprintf("\nin the initial table (Fragility = %d).", total_switch),
-            sprintf("This transfer of cases \nyields the following table:")
+            sprintf("\ntable (Fragility = %d).", total_switch),
+            sprintf("This transfer of cases yields the following table:")
           ))
           
         } else if (!final_solution$needtworows & final_solution$final_switch == 1) {
@@ -546,8 +546,8 @@ test_sensitivity_ln <- function(est_eff,
           cat(paste(
             sprintf("This is equivalent to \ntransferring %d", final_solution$final_switch), 
             c("case from"), transferway, 
-            sprintf("\nin the initial table (Fragility = %d).", total_switch),
-            sprintf("This transfer of cases \nyields the following table:")
+            sprintf("\ntable (Fragility = %d).", total_switch),
+            sprintf("This transfer of cases yields the following table:")
           ))
           
         } else {
@@ -590,8 +590,8 @@ test_sensitivity_ln <- function(est_eff,
           cat(paste(
             sprintf("This is equivalent \nto transferring %d", final_solution$final_switch), 
             c("case from"), transferway, 
-            sprintf("\nin the initial table (Fragility = %d).", total_switch),
-            sprintf("This transfer of cases \nyields the following table:")
[…]
