+++ R/test_sensitivity_ln.R
@@ -372,8 +372,8 @@ test_sensitivity_ln <- function(est_eff,
           cat(paste(
             sprintf("This is equivalent to transferring \n%d", final_solution$final_switch), 
             c("case from"), transferway, 
-            sprintf("(Fragility = %d).\n", total_switch),
-            sprintf("This transfer of cases yields the following table:")
+            sprintf("(Fragility = %d).", total_switch),
+            sprintf("\nThis transfer of cases yields the following table:")
           ))
           
         } else if (!final_solution$needtworows & final_solution$final_switch == 1) {
@@ -393,8 +393,8 @@ test_sensitivity_ln <- function(est_eff,
           cat(paste(
             sprintf("This is equivalent to transferring \n%d", final_solution$final_switch), 
             c("case from"), transferway, 
-            sprintf("(Fragility = %d).\n", total_switch),
-            sprintf("This transfer of cases yields the following table:")
+            sprintf("(Fragility = %d).", total_switch),
+            sprintf("\nThis transfer of cases yields the following table:")
           ))
           
         } else {
