+++ R/test_sensitivity_ln.R
@@ -401,11 +401,11 @@ test_sensitivity_ln <- function(est_eff,
           #conclusion1 <- 
         cat(paste0(
             sprintf("The inference cannot be invalidated merely by switching cases in"),
-            sprintf("\nonly one treatment condition. Therefore,cases have been"),
-            c("\nswitched from"), transferway, c("and from"),
-            c("\n"), transferway_extra, c("."), 
-            sprintf("The final Fragility(= %d) and RIR\n", total_switch),
-            c("reflects both sets of changes. Please compare the after transfer table with the implied table")
+            sprintf("\nonly one treatment condition. Therefore,cases have been switched from"),
+            c("\n"), transferway, c(" and from"),
+            transferway_extra, c("."), c("\n")
+            sprintf("The final Fragility(= %d) and RIR", total_switch),
+            c("reflects both sets of changes. \nPlease compare the after transfer table with the implied table")
             )
               )
         }
@@ -462,13 +462,13 @@ test_sensitivity_ln <- function(est_eff,
           
         } else {
           #conclusion1 <- 
-          cat(paste0(
+        cat(paste0(
             sprintf("The inference cannot be invalidated merely by switching cases in"),
-            sprintf("\nonly one treatment condition. Therefore,cases have been"),
-            c("\nswitched from"), transferway, c("and from"),
-            c("\n"), transferway_extra, c("."), 
-            sprintf("The final Fragility(= %d) and RIR\n", total_switch),
-            c("reflects both sets of changes. Please compare the after transfer table with the implied table")
+            sprintf("\nonly one treatment condition. Therefore,cases have been switched from"),
+            c("\n"), transferway, c(" and from"),
+            transferway_extra, c("."), c("\n")
+            sprintf("The final Fragility(= %d) and RIR", total_switch),
+            c("reflects both sets of changes. \nPlease compare the after transfer table with the implied table")
             )
               )
         }
@@ -548,11 +548,11 @@ test_sensitivity_ln <- function(est_eff,
           #conclusion1 <- 
             cat(paste0(
             sprintf("The inference cannot be invalidated merely by switching cases in"),
-            sprintf("\nonly one treatment condition. Therefore,cases have been"),
-            c("\nswitched from"), transferway, c("and from"),
-            c("\n"), transferway_extra, c("."), 
-            sprintf("The final Fragility(= %d) and RIR\n", total_switch),
-            c("reflects both sets of changes. Please compare the after transfer table with the implied table")
+            sprintf("\nonly one treatment condition. Therefore,cases have been switched from"),
+            c("\n"), transferway, c(" and from"),
+            transferway_extra, c("."), c("\n")
+            sprintf("The final Fragility(= %d) and RIR", total_switch),
+            c("reflects both sets of changes. \nPlease compare the after transfer table with the implied table")
             )
               )
                    }
@@ -608,13 +608,13 @@ test_sensitivity_ln <- function(est_eff,
           
         } else {
           #conclusion1 <- 
-           cat(paste0(
+            cat(paste0(
             sprintf("The inference cannot be invalidated merely by switching cases in"),
-            sprintf("\nonly one treatment condition. Therefore,cases have been"),
-            c("\nswitched from"), transferway, c("and from"),
-            c("\n"), transferway_extra, c("."), 
-            sprintf("The final Fragility(= %d) and RIR\n", total_switch),
-            c("reflects both sets of changes. Please compare the after transfer table with the implied table")
+            sprintf("\nonly one treatment condition. Therefore,cases have been switched from"),
+            c("\n"), transferway, c(" and from"),
+            transferway_extra, c("."), c("\n")
+            sprintf("The final Fragility(= %d) and RIR", total_switch),
+            c("reflects both sets of changes. \nPlease compare the after transfer table with the implied table")
             )
               )
         }
