+++ R/test_sensitivity_ln.R
@@ -403,7 +403,7 @@ test_sensitivity_ln <- function(est_eff,
             sprintf("The inference cannot be invalidated merely by switching cases in"),
             sprintf("\nonly one treatment condition. Therefore,cases have been switched from"),
             c("\n"), transferway, c(" and from"),
-            transferway_extra, c("."), c("\n")
+            transferway_extra, c("."), c("\n"),
             sprintf("The final Fragility(= %d) and RIR", total_switch),
             c("reflects both sets of changes. \nPlease compare the after transfer table with the implied table")
             )
@@ -466,7 +466,7 @@ test_sensitivity_ln <- function(est_eff,
             sprintf("The inference cannot be invalidated merely by switching cases in"),
             sprintf("\nonly one treatment condition. Therefore,cases have been switched from"),
             c("\n"), transferway, c(" and from"),
-            transferway_extra, c("."), c("\n")
+            transferway_extra, c("."), c("\n"),
             sprintf("The final Fragility(= %d) and RIR", total_switch),
             c("reflects both sets of changes. \nPlease compare the after transfer table with the implied table")
             )
@@ -550,7 +550,7 @@ test_sensitivity_ln <- function(est_eff,
             sprintf("The inference cannot be invalidated merely by switching cases in"),
             sprintf("\nonly one treatment condition. Therefore,cases have been switched from"),
             c("\n"), transferway, c(" and from"),
-            transferway_extra, c("."), c("\n")
+            transferway_extra, c("."), c("\n"),
             sprintf("The final Fragility(= %d) and RIR", total_switch),
             c("reflects both sets of changes. \nPlease compare the after transfer table with the implied table")
             )
@@ -612,7 +612,7 @@ test_sensitivity_ln <- function(est_eff,
             sprintf("The inference cannot be invalidated merely by switching cases in"),
             sprintf("\nonly one treatment condition. Therefore,cases have been switched from"),
             c("\n"), transferway, c(" and from"),
-            transferway_extra, c("."), c("\n")
+            transferway_extra, c("."), c("\n"),
             sprintf("The final Fragility(= %d) and RIR", total_switch),
             c("reflects both sets of changes. \nPlease compare the after transfer table with the implied table")
             )
