### 2023-12-04 a014089
deal with signsuppression == 1
