+++ R/test_sensitivity.R
@@ -32,9 +32,11 @@ test_sensitivity <- function(est_eff,
                              tails = 2,
                              index,
                              nu = 0, # null hypothesis 
-                             suppression = 0,
+                             signsuppression = 0,
+                             ## signsuprresion  means towards the other side of nu
                              ## by default is zero
                              ## alternative is one  
+                             
                              eff_thr = NA, # another non-zero and arbitrary threshold in terms of beta
                              to_return,
                              model_object,
@@ -42,7 +44,7 @@ test_sensitivity <- function(est_eff,
   
   ## warning messages for potential confusion 
   
-  if (suppression == 1) warning("suppression is defined by a threshold of opposite sign of the estimated effect.")
+  if (signsuppression == 1) warning("signsuppression is defined by a threshold of opposite sign of the estimated effect.")
 
   if (nu != 0) warning("You entered a non-zero null hypothesis about an effect. 
                        ITCV is calculated assuming omitted variable is equally 
@@ -79,12 +81,21 @@ test_sensitivity <- function(est_eff,
   LWbound <- nu - abs(critical_t * std_err)
 
   # determine mp 
+  # qqq user specified suppression argument should NOT impact this 
+  ##### qqqqq when mp == 1, actually mp may be affected need to try both ways 
   if ((est_eff > LWbound) & (est_eff < UPbound)) {mp <- 1}
   if (est_eff < LWbound | est_eff > UPbound) {mp <- -1}
   
   # determine beta_threshold 
-  if (est_eff < nu) {beta_threshold <- LWbound}
-  if (est_eff > nu) {beta_threshold <- UPbound}
+  # if signsuppression == 1, then the opposite 
+  if (est_eff < nu) {
+      beta_threshold <- (signsuppression == 0) * LWbound +
+          (signsuppression == 1) * UPbound
+      }
+  if (est_eff >= nu) {
+      beta_threshold <- (signsuppression == 1) * LWbound +
+          (signsuppression == 0) * UPbound
+      }
   
   # determine signITCV
   if (est_eff < beta_threshold) {signITCV <- -1}
@@ -94,14 +105,34 @@ test_sensitivity <- function(est_eff,
   # I. for RIR
 
   # calculating percentage of effect and number of observations to sustain or invalidate inference
-  if (abs(est_eff) > abs(beta_threshold)) {
-    perc_to_change <- bias <- 100 * (1 - (beta_threshold / est_eff))
-    recase <- round(n_obs * (bias / 100))
-  } else if (abs(est_eff) < abs(beta_threshold)) {
-    perc_to_change <- sustain <- 100 * (1 - (est_eff / beta_threshold))
-    recase <- round(n_obs * (sustain / 100))
-  } else if (est_eff == beta_threshold) {
-    stop("The coefficient is exactly equal to the threshold.")
+  if (signsuppression == 0) {
+      if (abs(est_eff) > abs(beta_threshold)) {
+        perc_to_change <- bias <- 100 * (1 - (beta_threshold / est_eff))
+        recase <- round(n_obs * (bias / 100))
+      } else if (abs(est_eff) < abs(beta_threshold)) {
+        perc_to_change <- sustain <- 100 * (1 - (est_eff / beta_threshold))
+        recase <- round(n_obs * (sustain / 100))
+      } else if (est_eff == beta_threshold) {
+        stop("The coefficient is exactly equal to the threshold.")
+      }
+  }
+  
[…]
