### 2023-12-04 783a500
add act_r into raw_output
