+++ R/helper_output_list.R
@@ -1,4 +1,5 @@
-output_list <- function(obs_r, critical_r, r_final, rxcv, rycv, rxcvGz, rycvGz, 
+output_list <- function(obs_r, act_r,
+                        critical_r, r_final, rxcv, rycv, rxcvGz, rycvGz, 
                         itcvGz, itcv, r2xz, r2yz, 
                         delta_star, delta_star_restricted, delta_exact, delta_pctbias, 
                         cor_oster, cor_exact, 
@@ -6,6 +7,7 @@ output_list <- function(obs_r, critical_r, r_final, rxcv, rycv, rxcvGz, rycvGz,
                         RIR, RIR_perc, fragility, starting_table, SE, 
                         Fig_ITCV, Fig_RIR) {
   result <- list(obs_r = obs_r,
+                 act_r = act_r, 
                  critical_r = critical_r, 
                  r_final = r_final,
                  rxcv = rxcv, 
