### 2023-12-04 503f3fd
align raw_output through output_list
