+++ R/helper_output_list.R
@@ -4,7 +4,8 @@ output_list <- function(obs_r, act_r,
                         delta_star, delta_star_restricted, delta_exact, delta_pctbias, 
                         cor_oster, cor_exact, 
                         beta_threshold, perc_bias_to_change, 
-                        RIR, RIR_perc, fragility, starting_table, SE, 
+                        RIR, RIR_perc, fragility, starting_table, final_table, 
+                        user_SE, analysis_SE, 
                         Fig_ITCV, Fig_RIR) {
   result <- list(obs_r = obs_r,
                  act_r = act_r, 
@@ -30,7 +31,9 @@ output_list <- function(obs_r, act_r,
                  RIR_perc = RIR_perc, 
                  fragility = fragility, 
                  starting_table = starting_table, 
-                 SE = SE, 
+                 final_table = final_table, 
+                 user_SE = user_SE,
+                 analysis_SE = analysis_SE, 
                  Fig_ITCV = Fig_ITCV,
                  Fig_RIR = Fig_RIR)
   result <- result[!is.na(result)]
