+++ R/test_sensitivity.R
@@ -242,8 +242,9 @@ test_sensitivity <- function(est_eff,
                        beta_threshold = beta_threshold,
                        perc_bias_to_change = perc_to_change, 
                        RIR = recase, RIR_perc = perc_to_change, 
-                       fragility = NA, starting_table = NA, 
-                       SE = NA, 
+                       fragility = NA, 
+                       starting_table = NA, final_table = NA,
+                       user_SE = NA, analysis_SE = NA,  
                        Fig_ITCV = 
                          plot_correlation(r_con = r_con, obs_r = obs_r, critical_r = critical_r),
                        Fig_RIR = plot_threshold(beta_threshold = beta_threshold, est_eff = est_eff)))
