+++ R/tkonfound.R
@@ -91,6 +91,8 @@ tkonfound <- function(a, b, c, d,
     RIRway <- "control success"
   }
   
+  RIR_extra <- 0
+  
   if (allnotenough) {
     if (switch_trm && dcroddsratio_ob) {
       transferway_extra <- "control failure to control success"
@@ -193,7 +195,30 @@ tkonfound <- function(a, b, c, d,
   info5 <- "the treatment success cell to the treatment failure cell."
   
   if (to_return == "raw_output") {
-    
+
+  return(output_list(obs_r = NA, act_r = NA, 
+                     critical_r = NA, r_final = NA,
+                     rxcv = NA, rycv = NA, 
+                     rxcvGz = NA, rycvGz = NA, 
+                     itcvGz = NA, itcv = NA, 
+                     r2xz = NA, r2yz = NA, 
+                     delta_star = NA, delta_star_restricted = NA, 
+                     delta_exact = NA, delta_pctbias = NA, 
+                     cor_oster = NA, cor_exact = NA, 
+                     beta_threshold = NA,
+                     perc_bias_to_change = NA, 
+                     RIR = RIR + RIR_extra, 
+                     RIR_perc = NA,  # need to discuss the denominator
+                     fragility = total_switch, 
+                     starting_table = table_start,
+                     final_table = table_final,
+                     user_SE = NA,
+                     analysis_SE = NA, 
+                     Fig_ITCV = NA,
+                     Fig_RIR = NA))
+      
+      
+          
     result <- list(info1,
                    info2,
                    conclusion1,
