### 2023-12-04 c310aca
add note for list output
