+++ R/test_sensitivity_ln.R
@@ -423,7 +423,7 @@ test_sensitivity_ln <- function(est_eff,
             sprintf("\nonly one treatment condition. Therefore,cases have been switched from"),
             c("\n"), transferway, c(" and from "),
             transferway_extra, c("."), c("\n"),
-            sprintf("The final Fragility(= %d) and RIR(= %d)", final_switch, RIR),
+            sprintf("The final Fragility(= %d) and RIR(= %d)", final_solution$final_switch, RIR),
             c(" reflects both sets of changes. \nPlease compare the after transfer table with the implied table.")
             )
               )
@@ -491,7 +491,7 @@ test_sensitivity_ln <- function(est_eff,
             sprintf("\nonly one treatment condition. Therefore,cases have been switched from"),
             c("\n"), transferway, c(" and from "),
             transferway_extra, c("."), c("\n"),
-            sprintf("The final Fragility(= %d) and RIR(= %d)", final_switch, RIR),
+            sprintf("The final Fragility(= %d) and RIR(= %d)", final_solution$final_switch, RIR),
             c(" reflects both sets of changes. \nPlease compare the after transfer table with the implied table.")
             )
               )
@@ -582,7 +582,7 @@ test_sensitivity_ln <- function(est_eff,
             sprintf("\nonly one treatment condition. Therefore,cases have been switched from"),
             c("\n"), transferway, c(" and from "),
             transferway_extra, c("."), c("\n"),
-            sprintf("The final Fragility(= %d) and RIR(= %d)", final_switch, RIR),
+            sprintf("The final Fragility(= %d) and RIR(= %d)", final_solution$final_switch, RIR),
             c(" reflects both sets of changes. \nPlease compare the after transfer table with the implied table.")
             )
               )
@@ -650,7 +650,7 @@ test_sensitivity_ln <- function(est_eff,
             sprintf("\nonly one treatment condition. Therefore,cases have been switched from"),
             c("\n"), transferway, c(" and from "),
             transferway_extra, c("."), c("\n"),
-            sprintf("The final Fragility(= %d) and RIR(= %d)", final_switch, RIR),
+            sprintf("The final Fragility(= %d) and RIR(= %d)", final_solution$final_switch, RIR),
             c(" reflects both sets of changes. \nPlease compare the after transfer table with the implied table.")
             )
               )
