### 2023-12-05 78ec6f9
fix bug error with thresh_plot output #63
