### 2023-12-07 d778b11
Update helper_output_print.R
