+++ R/helper_output_print.R
@@ -16,10 +16,10 @@ output_print <- function(est_eff, beta_threshhold, bias = NULL, sustain = NULL, 
       cat("\n")
       cat(paste0("(", round(est_eff, 3), ") estimate would have to be due to bias. This implies"))
       cat("\n")
-      cat("that to invalidate the inference one would expect to have to\n")
-      cat(paste0("replace ", round(recase, 3), " observations with cases for which the treatment"))
+      cat("that to invalidate the inference one would expect to have to replace\n")
+      cat(paste0("replace ", round(recase, 3), "(", round(bias, 3), "%) observations with cases for which the"))
       cat("\n")    
-      cat(paste0("effect is ", nu, " (RIR = ", round(recase, 3), ").\n"))
+      cat(paste0("treatment effect is ", nu, " (RIR = ", round(recase, 3), ").\n"))
       cat("\n")    
     } else if ((abs(est_eff) > abs(beta_threshhold)) & is.na(eff_thr) == FALSE) {
       cat("TO INVALIDATE:\n")
@@ -28,15 +28,15 @@ output_print <- function(est_eff, beta_threshhold, bias = NULL, sustain = NULL, 
       cat("\n")
       cat(paste0("You entered an effect of ", round(est_eff, 3), ", and specified a threshold"))
       cat("\n")
-      cat(paste0("value of ", round(eff_thr, 3), ". To invalidate the inference based on your estimate"))
+      cat(paste0("for inference of ", round(eff_thr, 3), ". To invalidate the inference based on your estimate"))
       cat("\n")
       cat(paste0(round(bias, 3), " % of the (", round(est_eff, 3), ") estimate would have to be due to bias."))
       cat("\n")
       cat(paste0("This implies that to invalidate the inference one would expect"))
       cat("\n")
-      cat("to have to replace", round(recase, 3), "observations with cases for which")
+      cat("to have to replace", round(recase, 3), "(", round(bias, 3), "%) observations with cases")
       cat("\n")
-      cat(paste0("the treatment effect is ", nu, " (RIR = ", round(recase, 3), ").\n"))
+      cat(paste0("for which the treatment effect is ", nu, " (RIR = ", round(recase, 3), ").\n"))
       cat("\n") 
     } else if ((abs(est_eff) < abs(beta_threshhold)) & is.na(eff_thr) == TRUE) {
       cat("TO SUSTAIN:\n", sep = "")
@@ -49,10 +49,10 @@ output_print <- function(est_eff, beta_threshhold, bias = NULL, sustain = NULL, 
       cat("\n")
       cat(paste0("threshold, ", round(sustain, 3), "% of the (", round(est_eff, 3), ") estimate would have to be due to"))
       cat("\n")
-      cat("bias. This implies that to sustain the inference on would expect to")
+      cat("bias. This implies that to sustain an inference one would expect to")
       cat("\n")
-      cat(paste0("have to replace ", round(recase, 3), " observation with effect of ", nu, " with cases with\n"))
-      cat(paste0("effect of ", round(beta_threshhold, 3), "(RIR = ", round(recase, 3), ").\n"))
+      cat(paste0("have to replace ", round(recase, 3), "(", round(bias, 3), "%) observations with effect of ", nu, " with\n"))
+      cat(paste0("cases with effect of ", round(beta_threshhold, 3), " (RIR = ", round(recase, 3), ").\n"))
       cat("\n")
     } else if ((abs(est_eff) < abs(beta_threshhold)) & is.na(eff_thr) == FALSE) {
       cat("TO SUSTAIN:\n", sep = "")
[…]
