### 2023-12-07 90f2f3b
Update helper_output_print.R
