+++ R/helper_output_print.R
@@ -76,6 +76,7 @@ output_print <- function(est_eff, beta_threshhold, bias = NULL, sustain = NULL, 
     }
     cat("See Frank et al. (2013) for a description of the method.")
     cat("\n")
+    "<a href='https://journals.sagepub.com/doi/10.3102/0162373713493129' target='_blank'>",
     cat("\n")
     cat(crayon::underline("Citation:"), "Frank, K.A., Maroulis, S., Duong, M., and Kelcey, B. (2013).")
     cat("\n")
@@ -85,11 +86,7 @@ output_print <- function(est_eff, beta_threshhold, bias = NULL, sustain = NULL, 
     cat("\n")
     cat(crayon::italic("Education, Evaluation and Policy Analysis, 35"), "437-460.")
     cat("\n")
-    if(tolower(readline()) == 'y') {
-        utils::browseURL("https://journals.sagepub.com/doi/10.3102/0162373713493129")  # Replace with the actual URL
-    }
-        cat("\n")
-
+   
   } 
   if (index == "IT") { 
     cat(crayon::bold("Impact Threshold for a Confounding Variable:\n"))
