### 2023-12-07 f844120
Update helper_output_print.R
