### 2023-12-07 63cdb7f
Update tkonfound.R
changed following Dr.Lin's modification
