### 2023-12-07 3d6a5e6
Update helper_output_print.R
