### 2023-12-07 1f863d1
Update helper_output_print.R
