### 2023-12-07 728da60
Update helper_output_print.R
