+++ R/test_sensitivity_ln.R
@@ -363,7 +363,7 @@ test_sensitivity_ln <- function(est_eff,
       print(Implied_Table)
       cat("\n")
       cat(paste(sprintf("The reported effect size = %.3f, SE = %.3f.",
-                        est_eff, std_err),
+                        est_eff, user_std_err),
                 sprintf("\nThe SE has been adjusted to %.3f to generate real number in the", final_solution$std_err_start), 
                 sprintf("\nimplied table. Numbers in the table cells have been rounded"),
                 sprintf("\nto integers, which may slightly alter the estimated effect from"), 
@@ -523,7 +523,7 @@ test_sensitivity_ln <- function(est_eff,
       print(Implied_Table)
       cat("\n")
       cat(paste(sprintf("(The reported effect size = %.3f, and SE = %.3f.",
-                        est_eff, std_err),
+                        est_eff, user_std_err),
                 sprintf("\nValues have been rounded to the nearest integer. This may cause"), 
                 sprintf("\na little change to the estimated effect for the table.)\n\n"))
       )
