### 2023-12-08 c70d30a
align raw output
now we RIR_primary, RIR_supplemental, RIR_perc, fragility_primary, fragility_supplemental
