+++ R/test_sensitivity.R
@@ -243,8 +243,8 @@ test_sensitivity <- function(est_eff,
                        cor_oster = NA, cor_exact = NA, 
                        beta_threshold = beta_threshold,
                        perc_bias_to_change = perc_to_change, 
-                       RIR = recase, RIR_perc = perc_to_change, 
-                       fragility = NA, 
+                       RIR_primary = recase, RIR_supplemental = NA, RIR_perc = perc_to_change, 
+                       fragility_primary = NA, fragility_supplemental = NA, 
                        starting_table = NA, final_table = NA,
                        user_SE = NA, analysis_SE = NA,  
                        Fig_ITCV = 
