+++ R/helper_output_list.R
@@ -4,7 +4,9 @@ output_list <- function(obs_r, act_r,
                         delta_star, delta_star_restricted, delta_exact, delta_pctbias, 
                         cor_oster, cor_exact, 
                         beta_threshold, perc_bias_to_change, 
-                        RIR, RIR_perc, fragility, starting_table, final_table, 
+                        RIR_primary, RIR_supplemental, RIR_perc, 
+                        fragility_primary, fragility_supplemental,
+                        starting_table, final_table, 
                         user_SE, analysis_SE, 
                         Fig_ITCV, Fig_RIR) {
   result <- list(obs_r = obs_r,
@@ -27,9 +29,11 @@ output_list <- function(obs_r, act_r,
                  cor_exact = cor_exact, 
                  beta_threshold = beta_threshold, 
                  perc_bias_to_change = perc_bias_to_change, 
-                 RIR = RIR, 
+                 RIR_primary = RIR_primary, 
+                 RIR_supplemental = RIR_supplemental, 
                  RIR_perc = RIR_perc, 
-                 fragility = fragility, 
+                 fragility_primary = fragility_primary, 
+                 fragility_supplemental = fragility_supplemental, 
                  starting_table = starting_table, 
                  final_table = final_table, 
                  user_SE = user_SE,
