+++ R/test_sensitivity_ln.R
@@ -150,24 +150,33 @@ test_sensitivity_ln <- function(est_eff,
     transferway <- "treatment success to treatment failure,"
     RIR <- ceiling(final/((a+c)/n_obs))*(replace=="entire") + ceiling(final/(a/(a+b)))*(1-(replace=="entire"))
     RIRway <- "treatment success"
+    RIR_pi <- RIR / d * 100
   }
   if (switch_trm && !dcroddsratio_ob) {
     transferway <- "treatment failure to treatment success,"
     RIR <- ceiling(final/((b+d)/n_obs))*(replace=="entire") + ceiling(final/(b/(a+b)))*(1-(replace=="entire"))
     RIRway <- "treatment failure"
+    RIR_pi <- RIR / c * 100
   }
   if (!switch_trm && dcroddsratio_ob) {
     transferway <- "control failure to control success,"
     RIR <- ceiling(final/((b+d)/n_obs))*(replace=="entire") + ceiling(final/(b/(a+b)))*(1-(replace=="entire"))
     RIRway <- "control failure"
+    RIR_pi <- RIR / a * 100
   }
   if (!switch_trm && !dcroddsratio_ob) {
     transferway <- "control success to control failure,"
     RIR <- ceiling(final/((a+c)/n_obs))*(replace=="entire") + ceiling(final/(a/(a+b)))*(1-(replace=="entire"))
     RIRway <- "control success"
+    RIR_pi <- RIR / b * 100
   }
   
+  RIR_extra <- final_extra <- NA
+  
   if (final_solution$needtworows) {
+    # if need two rows, then do not report RIR_pi 
+    ## because denominator is tricky 
+    RIR_pi <- NA
     final_extra <- final_solution$final_extra
     if (switch_trm && dcroddsratio_ob) {
       transferway_extra <- "control failure to control success,"
@@ -296,9 +305,11 @@ test_sensitivity_ln <- function(est_eff,
                   cor_oster = NA, cor_exact = NA, 
                   beta_threshold = NA,
                   perc_bias_to_change = NA, 
-                  RIR = total_RIR, 
-                  RIR_perc = NA,  # need to discuss the denominator
-                  fragility = total_switch, 
+                  RIR_primary = RIR,
+                  RIR_supplemental = RIR_extra, 
+                  RIR_perc = RIR_pi,  # need to discuss the denominator
+                  fragility_primary = final,
+                  fragility_supplemental = final_extra,
                   starting_table = final_solution$table_start,
                   final_table = final_solution$table_final,
                   user_SE = user_std_err,
