### 2023-12-12 12ea71b
Update test_sensitivity_ln.R
align with Dr.Lin's change
