### 2023-12-12 161878b
Update tkonfound.R
