### 2023-12-12 1dbaf8b
Update helper_output_list.R
