+++ R/helper_output_list.R
@@ -1,12 +1,18 @@
-output_list <- function(obs_r, critical_r, rxcv, rycv, rxcvGz, rycvGz, 
+output_list <- function(obs_r, act_r,
+                        critical_r, r_final, rxcv, rycv, rxcvGz, rycvGz,  
                         itcvGz, itcv, r2xz, r2yz, 
                         delta_star, delta_star_restricted, delta_exact, delta_pctbias, 
                         cor_oster, cor_exact, 
                         beta_threshold, perc_bias_to_change, 
-                        RIR, RIR_perc, fragility, starting_table, SE, 
+                        RIR_primary, RIR_supplemental, RIR_perc, 
+                        fragility_primary, fragility_supplemental,
+                        starting_table, final_table,
+                        user_SE, analysis_SE, 
                         Fig_ITCV, Fig_RIR) {
   result <- list(obs_r = obs_r,
+                 act_r = act_r, 
                  critical_r = critical_r, 
+                 r_final = r_final,
                  rxcv = rxcv, 
                  rycv = rycv, 
                  rxcvGz = rxcvGz, 
@@ -22,14 +28,20 @@ output_list <- function(obs_r, critical_r, rxcv, rycv, rxcvGz, rycvGz,
                  cor_oster = cor_oster, 
                  cor_exact = cor_exact, 
                  beta_threshold = beta_threshold, 
-                 perc_bias_to_change = perc_bias_to_change, 
-                 RIR = RIR, 
+                 perc_bias_to_change = perc_bias_to_change,
+                 RIR_primary = RIR_primary, 
+                 RIR_supplemental = RIR_supplemental, 
                  RIR_perc = RIR_perc, 
-                 fragility = fragility, 
+                 fragility_primary = fragility_primary, 
+                 fragility_supplemental = fragility_supplemental, 
                  starting_table = starting_table, 
+                 final_table = final_table, 
+                 user_SE = user_SE,
+                 analysis_SE = analysis_SE
                  SE = SE, 
                  Fig_ITCV = Fig_ITCV,
                  Fig_RIR = Fig_RIR)
   result <- result[!is.na(result)]
+  message("For interpretation, check out to_return = 'print'.")
   return(result)
 }
