+++ R/test_sensitivity.R
@@ -32,7 +32,8 @@ test_sensitivity <- function(est_eff,
                              tails = 2,
                              index,
                              nu = 0, # null hypothesis 
-                             suppression = 0,
+                             signsuppression = 0,
+                             ## signsuprresion  means towards the other side of nu
                              ## by default is zero
                              ## alternative is one  
                              eff_thr = NA, # another non-zero and arbitrary threshold in terms of beta
@@ -42,8 +43,8 @@ test_sensitivity <- function(est_eff,
   
   ## warning messages for potential confusion 
   
-  if (suppression == 1) warning("suppression is defined by a threshold of opposite sign of the estimated effect.")
-
+if (signsuppression == 1) warning("signsuppression is defined by a threshold of opposite sign of the estimated effect.")
+  
   if (nu != 0) warning("You entered a non-zero null hypothesis about an effect. 
                        ITCV is calculated assuming omitted variable is equally 
                        correlated with predictor of interest and outcome. 
