+++ R/test_sensitivity.R
@@ -76,23 +76,69 @@ if (signsuppression == 1) warning("signsuppression is defined by a threshold of 
     critical_t <- stats::qt(1 - (alpha / tails), n_obs - n_covariates - 2)
   }
 
-  beta_threshold <- critical_t * std_err + nu * std_err 
-  # to account for non-zero nu, assuming stat sig is determined by non-zero nu
+  # create CI centered nu    
+  UPbound <- nu + abs(critical_t * std_err)
+  LWbound <- nu - abs(critical_t * std_err)
+
+  # determine mp 
+  # qqq user specified suppression argument should NOT impact this 
+  ##### qqqqq when mp == 1, actually mp may be affected need to try both ways 
+  if ((est_eff > LWbound) & (est_eff < UPbound)) {mp <- 1}
+  if (est_eff < LWbound | est_eff > UPbound) {mp <- -1}
   
-  # for replacement of cases approach
+  # determine beta_threshold 
+  # if signsuppression == 1, then the opposite 
+  if (est_eff < nu) {
+      beta_threshold <- (signsuppression == 0) * LWbound +
+          (signsuppression == 1) * UPbound
+      }
+  if (est_eff >= nu) {
+      beta_threshold <- (signsuppression == 1) * LWbound +
+          (signsuppression == 0) * UPbound
+      }
+  
+  # determine signITCV
+  if (est_eff < beta_threshold) {signITCV <- -1}
+  if (est_eff > beta_threshold) {signITCV <- 1}
+  if (est_eff == beta_threshold) {signITCV <- 0}
+  
+  # I. for RIR
+  # right now calculation in terms of effect size (not correlation)
+  # if swtich to correlation could do A D for signsuppression as well 
+  ## using -1 and +1 in the replacement 
 
   # calculating percentage of effect and number of observations to sustain or invalidate inference
-  if (abs(est_eff) > abs(beta_threshold)) {
-    perc_to_change <- bias <- 100 * (1 - (beta_threshold / est_eff))
-    recase <- round(n_obs * (bias / 100))
-  } else if (abs(est_eff) < abs(beta_threshold)) {
-      perc_to_change <- sustain <- 100 * (1 - (est_eff / beta_threshold))
-    recase <- round(n_obs * (sustain / 100))
-  } else if (est_eff == beta_threshold) {
-    stop("The coefficient is exactly equal to the threshold.")
+  if (signsuppression == 0) {
+      if (abs(est_eff) > abs(beta_threshold)) {
+        perc_to_change <- bias <- 100 * (1 - (beta_threshold / est_eff))
+        recase <- round(n_obs * (bias / 100))
+      } else if (abs(est_eff) < abs(beta_threshold)) {
+        perc_to_change <- sustain <- 100 * (1 - (est_eff / beta_threshold))
+        recase <- round(n_obs * (sustain / 100))
+      } else if (est_eff == beta_threshold) {
+        stop("The coefficient is exactly equal to the threshold.")
+      }
+  }
+  
+  if (signsuppression == 1) {
+      if (est_eff < LWbound | est_eff > UPbound) {
+          stop("Such scenarios do not make sense to consider RIR.")
+      } 
+      if (est_eff >= LWbound & est_eff < nu) {
+          ## B case
+          ## consider est_eff as a combination of pi*LB+(1-pi)*UB
+          perc_to_change <- bias <- 100 * (UPbound - est_eff)/(UPbound - LWbound)
+          recase <- round(n_obs * (bias / 100))
+      }
[…]
