+++ R/test_sensitivity_ln.R
@@ -431,7 +431,7 @@ test_sensitivity_ln <- function(est_eff,
         } else if (!final_solution$needtworows & final_solution$final_switch == 1) {
           #conclusion1 <- 
           cat(paste(
-            change, sprintf("one would need to replace %d", RIR), RIRway, "\ncases"))
+            change, sprintf("one would need to replace %d (%d%)", RIR, RIR_pi), RIRway, "\ncases"))
           
           if (replace == "control") {
             #conclusion1a <- 
