+++ R/test_sensitivity_ln.R
@@ -431,7 +431,7 @@ test_sensitivity_ln <- function(est_eff,
         } else if (!final_solution$needtworows & final_solution$final_switch == 1) {
           #conclusion1 <- 
           cat(paste(
-            change, sprintf("one would need to replace %d (%d%)", RIR, RIR_pi), RIRway, "\ncases"))
+            change, sprintf("one would need to replace %d (%.2f%)", RIR, RIR_pi), RIRway, "\ncases"))
           
           if (replace == "control") {
             #conclusion1a <- 
@@ -478,7 +478,7 @@ test_sensitivity_ln <- function(est_eff,
         if (!final_solution$needtworows & final_solution$final_switch > 1) {
           #conclusion1 <- 
           cat(paste(
-            change, sprintf("one would need to replace \n%d", RIR), RIRway, "cases "))
+            change, sprintf("one would need to replace %d (%.2f%)", RIR, RIR_pi), RIRway, "cases "))
           
           if (replace == "control") {
             #conclusion1a <- 
@@ -499,7 +499,7 @@ test_sensitivity_ln <- function(est_eff,
         } else if (!final_solution$needtworows & final_solution$final_switch == 1) {
           #conclusion1 <- 
           cat(paste(
-            change, sprintf("one would need to replace \n%d", RIR), RIRway, "\ncases"))
+            change, sprintf("one would need to replace %d (%.2f%)", RIR, RIR_pi), RIRway, "\ncases"))
           
           if (replace == "control") {
             #conclusion1a <- 
@@ -569,7 +569,7 @@ test_sensitivity_ln <- function(est_eff,
         if (!final_solution$needtworows & final_solution$final_switch > 1) {
           #conclusion1 <- 
           cat(paste(
-            change, sprintf("one would need to replace %d", RIR), RIRway, "\ncases "))
+            change, sprintf("one would need to replace %d (%.2f%)", RIR, RIR_pi), RIRway, "\ncases "))
           
           if (replace == "control") {
             #conclusion1a <- 
@@ -590,7 +590,7 @@ test_sensitivity_ln <- function(est_eff,
         } else if (!final_solution$needtworows & final_solution$final_switch == 1) {
           #conclusion1 <- 
           cat(paste(
-            change, sprintf("one would need to replace %d", RIR), RIRway, "\ncases"))
+            change, sprintf("one would need to replace %d (%.2f%)", RIR, RIR_pi), RIRway, "\ncases"))
           
           if (replace == "control") {
             #conclusion1a <- 
@@ -637,7 +637,7 @@ test_sensitivity_ln <- function(est_eff,
         if (!final_solution$needtworows & final_solution$final_switch > 1) {
           #conclusion1 <- 
           cat(paste(
-            change, sprintf("one would need to replace \n%d", RIR), RIRway, "cases "))
+            change, sprintf("one would need to replace %d (%.2f%)", RIR, RIR_pi), RIRway, "cases "))
           
           if (replace == "control") {
             #conclusion1a <- 
@@ -658,7 +658,7 @@ test_sensitivity_ln <- function(est_eff,
         } else if (!final_solution$needtworows & final_solution$final_switch == 1) {
           #conclusion1 <- 
           cat(paste(
-            change, sprintf("one would need to replace \n%d", RIR), RIRway, "\ncases"))
+            change, sprintf("one would need to replace %d (%.2f%)", RIR, RIR_pi), RIRway, "\ncases"))
           
           if (replace == "control") {
             #conclusion1a <- 
