### 2023-12-12 f87d6c5
Update helper_output_list.R
