+++ R/test_sensitivity_ln.R
@@ -405,7 +405,7 @@ test_sensitivity_ln <- function(est_eff,
         if (!final_solution$needtworows & final_solution$final_switch > 1) {
           #conclusion1 <- 
           cat(paste(
-            change, sprintf("one would need to replace %d (%.2f%)", RIR, RIR_pi), RIRway, "\ncases "))
+            change, sprintf("one would need to replace %d (%d%)", RIR, RIR_pi), RIRway, "\ncases "))
           
           if (replace == "control") {
             #conclusion1a <- 
@@ -426,7 +426,7 @@ test_sensitivity_ln <- function(est_eff,
         } else if (!final_solution$needtworows & final_solution$final_switch == 1) {
           #conclusion1 <- 
           cat(paste(
-            change, sprintf("one would need to replace %d (%.2f%)", RIR, RIR_pi), RIRway, "\ncases"))
+            change, sprintf("one would need to replace %d (%d%)", RIR, RIR_pi), RIRway, "\ncases"))
           
           if (replace == "control") {
             #conclusion1a <- 
@@ -473,7 +473,7 @@ test_sensitivity_ln <- function(est_eff,
         if (!final_solution$needtworows & final_solution$final_switch > 1) {
           #conclusion1 <- 
           cat(paste(
-            change, sprintf("one would need to replace %d (%.2f%)", RIR, RIR_pi), RIRway, "cases "))
+            change, sprintf("one would need to replace %d (%d%)", RIR, RIR_pi), RIRway, "cases "))
           
           if (replace == "control") {
             #conclusion1a <- 
@@ -494,7 +494,7 @@ test_sensitivity_ln <- function(est_eff,
         } else if (!final_solution$needtworows & final_solution$final_switch == 1) {
           #conclusion1 <- 
           cat(paste(
-            change, sprintf("one would need to replace %d (%.2f%)", RIR, RIR_pi), RIRway, "\ncases"))
+            change, sprintf("one would need to replace %d (%d%)", RIR, RIR_pi), RIRway, "\ncases"))
           
           if (replace == "control") {
             #conclusion1a <- 
@@ -564,7 +564,7 @@ test_sensitivity_ln <- function(est_eff,
         if (!final_solution$needtworows & final_solution$final_switch > 1) {
           #conclusion1 <- 
           cat(paste(
-            change, sprintf("one would need to replace %d (%.2f%)", RIR, RIR_pi), RIRway, "\ncases "))
+            change, sprintf("one would need to replace %d (%d%)", RIR, RIR_pi), RIRway, "\ncases "))
           
           if (replace == "control") {
             #conclusion1a <- 
@@ -585,7 +585,7 @@ test_sensitivity_ln <- function(est_eff,
         } else if (!final_solution$needtworows & final_solution$final_switch == 1) {
           #conclusion1 <- 
           cat(paste(
-            change, sprintf("one would need to replace %d (%.2f%)", RIR, RIR_pi), RIRway, "\ncases"))
+            change, sprintf("one would need to replace %d (%d%)", RIR, RIR_pi), RIRway, "\ncases"))
           
           if (replace == "control") {
             #conclusion1a <- 
@@ -632,7 +632,7 @@ test_sensitivity_ln <- function(est_eff,
         if (!final_solution$needtworows & final_solution$final_switch > 1) {
           #conclusion1 <- 
           cat(paste(
-            change, sprintf("one would need to replace %d (%.2f%)", RIR, RIR_pi), RIRway, "cases "))
+            change, sprintf("one would need to replace %d (%d%)", RIR, RIR_pi), RIRway, "cases "))
           
           if (replace == "control") {
             #conclusion1a <- 
[…]
