+++ R/test_sensitivity_ln.R
@@ -506,7 +506,7 @@ test_sensitivity_ln <- function(est_eff,
         } else if (!final_solution$needtworows & final_solution$final_switch == 1) {
           #conclusion1 <- 
           cat(paste(
-            change, sprintf("one would need to replace %d (%.2f)", RIR, RIR_pi), RIRway, "\ncases"))
+            change, sprintf("one would need to replace %d \n(%.2f)", RIR, RIR_pi), RIRway, "\ncases"))
           
           if (replace == "control") {
             #conclusion1a <- 
@@ -644,7 +644,7 @@ test_sensitivity_ln <- function(est_eff,
         if (!final_solution$needtworows & final_solution$final_switch > 1) {
           #conclusion1 <- 
           cat(paste(
-            change, sprintf("one would need to replace %d (%.2f)", RIR, RIR_pi), RIRway, "cases "))
+            change, sprintf("one would need to replace %d \n(%.2f)", RIR, RIR_pi), RIRway, "cases "))
           
           if (replace == "control") {
             #conclusion1a <- 
