### 2023-12-15 2d21536
Update helper_output_print.R
