### 2023-12-15 e5e8619
Update helper_output_print.R
