### 2023-12-15 9043cd8
Update helper_output_print.R
