### 2023-12-15 6bb8043
Update helper_output_print.R
