### 2023-12-15 1c3df38
Update helper_output_print.R
