+++ R/test_sensitivity_ln.R
@@ -431,9 +431,25 @@ test_sensitivity_ln <- function(est_eff,
           cat(paste(
             sprintf(" This is equivalent to transferring \n%d", final_solution$final_switch), 
             c("case from"), transferway, 
-            sprintf("(Fragility = %d).", total_switch),
-            sprintf("\nThis transfer of cases yields the following table:")
-          ))
+            sprintf("(Fragility = %d).", total_switch))
+              )
+            
+          if (RIR_pi > 100) {
+            a <- switch(RIRway,
+                        "treatment success" = d,
+                        "treatment failure" = c,
+                        "control failure" = a,
+                        "control success" = b,
+                        NA)
+            if (!is.na(a)) {
+              # Calculate the value for the %d placeholder
+              successRate <- 1 - Fragility / a
+              # Display the message
+              cat(sprintf("Note the RIR exceeds 100%. Generating the transfer of %d cases would require replacing\n cases with a probability of success of %d of %s which is smaller than the \nprobability of success for the control group used to calculate the RIR.", a, successRate, RIRway))
+            }
+          }
+          
+          cat(sprintf("\nThis transfer of cases yields the following table:"))
           
         } else if (!final_solution$needtworows & final_solution$final_switch == 1) {
           #conclusion1 <- 
@@ -452,9 +468,24 @@ test_sensitivity_ln <- function(est_eff,
           cat(paste(
             sprintf(" This is equivalent to transferring \n%d", final_solution$final_switch), 
             c("case from"), transferway, 
-            sprintf("(Fragility = %d).", total_switch),
-            sprintf("\nThis transfer of cases yields the following table:")
-          ))
+            sprintf("(Fragility = %d).", total_switch)))
+       
+          if (RIR_pi > 100) {
+            a <- switch(RIRway,
+                        "treatment success" = d,
+                        "treatment failure" = c,
+                        "control failure" = a,
+                        "control success" = b,
+                        NA)
+            if (!is.na(a)) {
+              # Calculate the value for the %d placeholder
+              successRate <- 1 - Fragility / a
+              # Display the message
+              cat(sprintf("Note the RIR exceeds 100%. Generating the transfer of %d cases would require replacing\n cases with a probability of success of %d of %s which is smaller than the \nprobability of success for the control group used to calculate the RIR.", a, successRate, RIRway))
+            }
+          }
+            
+          cat(sprintf("\nThis transfer of cases yields the following table:"))
           
         } else {
 
