+++ R/test_sensitivity_ln.R
@@ -443,7 +443,7 @@ test_sensitivity_ln <- function(est_eff,
                         NA)
             if (!is.na(a)) {
               # Calculate the value for the %d placeholder
-              successRate <- 1 - Fragility / a
+              successRate <- 1 - total_switch / a
               # Display the message
               cat(sprintf("Note the RIR exceeds 100%. Generating the transfer of %d cases would require replacing\n cases with a probability of success of %d of %s which is smaller than the \nprobability of success for the control group used to calculate the RIR.", a, successRate, RIRway))
             }
@@ -479,7 +479,7 @@ test_sensitivity_ln <- function(est_eff,
                         NA)
             if (!is.na(a)) {
               # Calculate the value for the %d placeholder
-              successRate <- 1 - Fragility / a
+              successRate <- 1 - total_switch / a
               # Display the message
               cat(sprintf("Note the RIR exceeds 100%. Generating the transfer of %d cases would require replacing\n cases with a probability of success of %d of %s which is smaller than the \nprobability of success for the control group used to calculate the RIR.", a, successRate, RIRway))
             }
@@ -541,7 +541,7 @@ test_sensitivity_ln <- function(est_eff,
                         NA)
             if (!is.na(a)) {
               # Calculate the value for the %d placeholder
-              successRate <- 1 - Fragility / a
+              successRate <- 1 - total_switch / a
               # Display the message
               cat(sprintf("Note the RIR exceeds 100%. Generating the transfer of %d cases would require replacing\n cases with a probability of success of %d of %s which is smaller than the \nprobability of success for the control group used to calculate the RIR.", a, successRate, RIRway))
             }
@@ -577,7 +577,7 @@ test_sensitivity_ln <- function(est_eff,
                         NA)
             if (!is.na(a)) {
               # Calculate the value for the %d placeholder
-              successRate <- 1 - Fragility / a
+              successRate <- 1 - total_switch / a
               # Display the message
               cat(sprintf("Note the RIR exceeds 100%. Generating the transfer of %d cases would require replacing\n cases with a probability of success of %d of %s which is smaller than the \nprobability of success for the control group used to calculate the RIR.", a, successRate, RIRway))
             }
@@ -754,7 +754,7 @@ test_sensitivity_ln <- function(est_eff,
                         NA)
             if (!is.na(a)) {
               # Calculate the value for the %d placeholder
-              successRate <- 1 - Fragility / a
+              successRate <- 1 - total_switch / a
               # Display the message
               cat(sprintf("Note the RIR exceeds 100%. Generating the transfer of %d cases would require replacing\n cases with a probability of success of %d of %s which is smaller than the \nprobability of success for the control group used to calculate the RIR.", a, successRate, RIRway))
             }
