+++ R/test_sensitivity_ln.R
@@ -543,7 +543,7 @@ test_sensitivity_ln <- function(est_eff,
               # Calculate the value for the %d placeholder
               successRate <- 1 - total_switch / a
               # Display the message
-              cat(sprintf("Note the RIR exceeds 100%. Generating the transfer of %d cases would require replacing\n cases with a probability of success of %d of %s which is smaller than the \nprobability of success for the control group used to calculate the RIR.", a, successRate, RIRway))
+              cat(sprintf("Note the RIR exceeds 100%. Generating the transfer of %d cases would require replacing\n cases with a probability of success of %.3f of %s which is smaller than the \nprobability of success for the control group used to calculate the RIR.", a, successRate, RIRway))
             }
           }
           
