+++ R/test_sensitivity_ln.R
@@ -544,7 +544,7 @@ test_sensitivity_ln <- function(est_eff,
               # Calculate the value for the %d placeholder
               successRate <- 1 - total_switch / a
               # Display the message
-              cat(sprintf("Note the RIR exceeds 100%. Generating the transfer of %s cases would require replacing\n cases with a probability of success of %s of %s which is smaller than the \nprobability of success for the control group used to calculate the RIR.", a, successRate, RIRway))
+              cat(paste(sprintf("Note the RIR exceeds 100%. Generating the transfer of %d cases would require replacing\n cases with a probability of success of %d of", a, successRate), RIRway, "which is smaller than the \nprobability of success for the control group used to calculate the RIR."))
             }
           }
           
@@ -580,7 +580,7 @@ test_sensitivity_ln <- function(est_eff,
               # Calculate the value for the %d placeholder
               successRate <- 1 - total_switch / a
               # Display the message
-              cat(sprintf("Note the RIR exceeds 100%. Generating the transfer of %s cases would require replacing\n cases with a probability of success of %s of %s which is smaller than the \nprobability of success for the control group used to calculate the RIR.", a, successRate, RIRway))
+              cat(paste(sprintf("Note the RIR exceeds 100%. Generating the transfer of %d cases would require replacing\n cases with a probability of success of %d of", a, successRate), RIRway, "which is smaller than the \nprobability of success for the control group used to calculate the RIR."))
             }
           }
           
