+++ R/test_sensitivity_ln.R
@@ -445,7 +445,7 @@ test_sensitivity_ln <- function(est_eff,
               # Calculate the value for the %d placeholder
               successRate <- 1 - total_switch / a
               # Display the message
-              cat(paste(sprintf("Note the RIR exceeds 100%. Generating the transfer of %d cases would require replacing\n cases with a probability of success of %d of", a, successRate), RIRway, "which is smaller than the \nprobability of success for the control group used to calculate the RIR."))
+              cat(paste(sprintf("Note the RIR exceeds 100%. Generating the transfer of %f cases would require replacing\n cases with a probability of success of %f of", a, successRate), RIRway, "which is smaller than the \nprobability of success for the control group used to calculate the RIR."))
               
             }
           }
@@ -482,7 +482,7 @@ test_sensitivity_ln <- function(est_eff,
               # Calculate the value for the %d placeholder
               successRate <- 1 - total_switch / a
               # Display the message
-              cat(paste(sprintf("Note the RIR exceeds 100%. Generating the transfer of %d cases would require replacing\n cases with a probability of success of %d of", a, successRate), RIRway, "which is smaller than the \nprobability of success for the control group used to calculate the RIR."))
+              cat(paste(sprintf("Note the RIR exceeds 100%. Generating the transfer of %f cases would require replacing\n cases with a probability of success of %f of", a, successRate), RIRway, "which is smaller than the \nprobability of success for the control group used to calculate the RIR."))
             }
           }
             
@@ -544,7 +544,7 @@ test_sensitivity_ln <- function(est_eff,
               # Calculate the value for the %d placeholder
               successRate <- 1 - total_switch / a
               # Display the message
-              cat(paste(sprintf("Note the RIR exceeds 100%. Generating the transfer of %d cases would require replacing\n cases with a probability of success of %d of", a, successRate), RIRway, "which is smaller than the \nprobability of success for the control group used to calculate the RIR."))
+              cat(paste(sprintf("Note the RIR exceeds 100%. Generating the transfer of %f cases would require replacing\n cases with a probability of success of %f of", a, successRate), RIRway, "which is smaller than the \nprobability of success for the control group used to calculate the RIR."))
             }
           }
           
@@ -580,7 +580,7 @@ test_sensitivity_ln <- function(est_eff,
               # Calculate the value for the %d placeholder
               successRate <- 1 - total_switch / a
               # Display the message
-              cat(paste(sprintf("Note the RIR exceeds 100%. Generating the transfer of %d cases would require replacing\n cases with a probability of success of %d of", a, successRate), RIRway, "which is smaller than the \nprobability of success for the control group used to calculate the RIR."))
+              cat(paste(sprintf("Note the RIR exceeds 100%. Generating the transfer of %f cases would require replacing\n cases with a probability of success of %f of", a, successRate), RIRway, "which is smaller than the \nprobability of success for the control group used to calculate the RIR."))
             }
           }
           
