+++ R/test_sensitivity_ln.R
@@ -543,18 +543,8 @@ test_sensitivity_ln <- function(est_eff,
             
     if (!is.na(transfer)) {
         # Calculate the value for the successRate
-        successRate <- 1 - total_switch / transfer
-        cat("Transfer:", transfer, "Class:", class(transfer), "\n")
-        cat("Success Rate:", successRate, "Class:", class(successRate), "\n")
-        cat("total_switch:", total_switch, "Class:", class(total_switch), "\n")
-        cat("RIRway:", RIRway, "Class:", class(RIRway), "\n")
-
-       # Separate sprintf statements
-        transferMsg <- sprintf("1Note the RIR exceeds 100%%. Generating the transfer of %d cases", total_switch)
-        successRateMsg <- sprintf("would require replacing cases with a probability of success of %.2f of", successRate)
-
-        # Combine with cat
-        cat(transferMsg, successRateMsg, RIRway, "which is smaller than the probability of success for the control group used to calculate the RIR.")
+      successRate <- 1 - total_switch / transfer
+      cat(paste(sprintf("Note the RIR exceeds 100%%. Generating\n the transfer of %d cases would require replacing cases with a probability\n of success of %.2f of", total_switch, successRate), RIRway, "which is smaller\n than the probability of success for the control group used to calculate the RIR."))
     }
 }
           
