+++ R/test_sensitivity_ln.R
@@ -545,11 +545,13 @@ test_sensitivity_ln <- function(est_eff,
     if (!is.na(transfer)) {
         # Calculate the value for the successRate
       successRate <- 1 - total_switch / transfer
-      cat("\n")
-      cat(paste(sprintf("Note the RIR exceeds 100%%. Generating the transfer of %d cases would\n", total_switch),
-                sprintf("require replacing cases with a probability of success of %.2f of\n", successRate), RIRway, 
-                sprintf("which is smaller than the probability of success for the control group \nused to calculate the RIR.")))
+
     }
+
+     cat("\n")
+     cat(paste(sprintf("Note the RIR exceeds 100%%. Generating the transfer of %d cases would\n", total_switch),
+                sprintf("require replacing cases with a probability of success of %.2f of\n", successRate), RIRway, 
+                sprintf("which is smaller than the probability of success \nfor the control group used to calculate the RIR.")))
 }
           
           
