+++ R/test_sensitivity_ln.R
@@ -555,6 +555,7 @@ test_sensitivity_ln <- function(est_eff,
                 sprintf("\nrequire replacing cases with a probability of success of %.2f of\n", successRate), RIRway, 
                 sprintf("which is smaller than the probability of success \nfor the control group used to calculate the RIR.")))
         } 
+        }
           
           
         } else if (!final_solution$needtworows & final_solution$final_switch == 1) {
