+++ R/test_sensitivity_ln.R
@@ -547,15 +547,14 @@ test_sensitivity_ln <- function(est_eff,
             c("case from"), transferway, 
             sprintf("\ntable (Fragility = %d).", total_switch)))
           cat(sprintf("This transfer of cases yields the following \ntable:\n"))
-
-
-
+          
      cat("\n")
+          
      cat(paste0(sprintf("Note the RIR exceeds 100%%. Generating the transfer of %d cases would", total_switch),
                 sprintf("\nrequire replacing cases with a probability of success of %.2f of\n", successRate), RIRway, 
                 sprintf("which is smaller than the probability of success \nfor the control group used to calculate the RIR.")))
-        } 
-        }
+        
+        
           
           
         } else if (!final_solution$needtworows & final_solution$final_switch == 1) {
