### 2023-12-16 010d575
Update helper_output_print.R
