### 2023-12-16 5f9371e
Update helper_output_print.R
