+++ R/test_sensitivity_ln.R
@@ -495,7 +495,7 @@ test_sensitivity_ln <- function(est_eff,
           #conclusion1 <- 
         cat(paste0(
             sprintf("The inference cannot be invalidated merely by switching cases in"),
-            sprintf("\nonly one treatment condition. Therefore,cases have been switched from"),
+            sprintf("\nonly one treatment condition. Therefore, cases have been switched from"),
             c("\n"), transferway, c(" and from "),
             transferway_extra, c("."), c("\n"),
             sprintf("The final Fragility(= %d) and RIR(= %d)", final_solution$final_switch, RIR),
@@ -578,7 +578,7 @@ test_sensitivity_ln <- function(est_eff,
           #conclusion1 <- 
         cat(paste0(
             sprintf("The inference cannot be sustained merely by switching cases in"),
-            sprintf("\nonly one treatment condition. Therefore,cases have been switched from"),
+            sprintf("\nonly one treatment condition. Therefore, cases have been switched from"),
             c("\n"), transferway, c(" and from "),
             transferway_extra, c("."), c("\n"),
             sprintf("The final Fragility(= %d) and RIR(= %d)", final_solution$final_switch, RIR),
@@ -679,7 +679,7 @@ test_sensitivity_ln <- function(est_eff,
           #conclusion1 <- 
             cat(paste0(
             sprintf("The inference cannot be invalidated merely by switching cases in"),
-            sprintf("\nonly one treatment condition. Therefore,cases have been switched from"),
+            sprintf("\nonly one treatment condition. Therefore, cases have been switched from"),
             c("\n"), transferway, c(" and from "),
             transferway_extra, c("."), c("\n"),
             sprintf("The final Fragility(= %d) and RIR(= %d)", final_solution$final_switch, RIR),
@@ -754,7 +754,7 @@ test_sensitivity_ln <- function(est_eff,
           #conclusion1 <- 
             cat(paste0(
             sprintf("The inference cannot be sustained merely by switching cases in"),
-            sprintf("\nonly one treatment condition. Therefore,cases have been switched from"),
+            sprintf("\nonly one treatment condition. Therefore, cases have been switched from"),
             c("\n"), transferway, c(" and from "),
             transferway_extra, c("."), c("\n"),
             sprintf("The final Fragility(= %d) and RIR(= %d)", final_solution$final_switch, RIR),
