+++ R/test_sensitivity_ln.R
@@ -499,7 +499,7 @@ test_sensitivity_ln <- function(est_eff,
             c("\n"), transferway, c(" and from "),
             transferway_extra, c("."), c("\n"),
             sprintf("The final Fragility(= %d) and RIR(= %d)", final_solution$final_switch, RIR),
-            c(" reflects both sets of changes. \nPlease compare the after transfer table with the implied table.")
+            c(" reflect both sets of changes. \nPlease compare the after transfer table with the implied table.")
             )
               )
         }
@@ -582,7 +582,7 @@ test_sensitivity_ln <- function(est_eff,
             c("\n"), transferway, c(" and from "),
             transferway_extra, c("."), c("\n"),
             sprintf("The final Fragility(= %d) and RIR(= %d)", final_solution$final_switch, RIR),
-            c(" reflects both sets of changes. \nPlease compare the after transfer table with the implied table.")
+            c(" reflect both sets of changes. \nPlease compare the after transfer table with the implied table.")
             )
               )
         }
@@ -683,7 +683,7 @@ test_sensitivity_ln <- function(est_eff,
             c("\n"), transferway, c(" and from "),
             transferway_extra, c("."), c("\n"),
             sprintf("The final Fragility(= %d) and RIR(= %d)", final_solution$final_switch, RIR),
-            c(" reflects both sets of changes. \nPlease compare the after transfer table with the implied table.")
+            c(" reflect both sets of changes. \nPlease compare the after transfer table with the implied table.")
             )
               )
                    }
@@ -758,7 +758,7 @@ test_sensitivity_ln <- function(est_eff,
             c("\n"), transferway, c(" and from "),
             transferway_extra, c("."), c("\n"),
             sprintf("The final Fragility(= %d) and RIR(= %d)", final_solution$final_switch, RIR),
-            c(" reflects both sets of changes. \nPlease compare the after transfer table with the implied table.")
+            c(" reflect both sets of changes. \nPlease compare the after transfer table with the implied table.")
             )
               )
         }
