### 2023-12-21 1375226
Update tkonfound.R
