+++ R/test_sensitivity_ln.R
@@ -461,8 +461,9 @@ p_start_chi_final <- chisq.test(final_solution$table_final,correct = FALSE)$p.va
             sprintf(" This is equivalent to transferring \n%d", final_solution$final_switch), 
             c("cases from"), transferway, 
             sprintf("(Fragility = %d).", total_switch),
-            c("Note that RIR = Fragility/[1-P(", RIRway_phrase, c(")]")))
-             )       
+            c("Note that RIR = Fragility/[1-P("), RIRway_phrase, c(")]"))
+             )
+                    
 
           cat("\n")
 
