+++ R/test_sensitivity_ln.R
@@ -459,10 +459,10 @@ p_start_chi_final <- chisq.test(final_solution$table_final,correct = FALSE)$p.va
           #conclusion1b <- 
           cat(paste0(
             sprintf(" This is equivalent to transferring \n%d", final_solution$final_switch), 
-            c("cases from"), transferway, 
-            sprintf("(Fragility = %d).", total_switch),
+            c(" cases from "), transferway, 
+            sprintf(" (Fragility = %d).", total_switch),
             c("\nNote that RIR = Fragility/[1-P("), RIRway_phrase, c(")]"))
-             )
+             ) 
                     
 
           cat("\n")
@@ -491,19 +491,19 @@ p_start_chi_final <- chisq.test(final_solution$table_final,correct = FALSE)$p.va
           }
           
           #conclusion1b <- 
-          cat(paste(
+          cat(paste0(
             sprintf(" This is equivalent to transferring \n%d", final_solution$final_switch), 
-            c("cases from"), transferway, 
-            sprintf("(Fragility = %d).", total_switch),
-            sprintf("Note that RIR = Fragility/[1-P(", RIRway_phrase, ")]"))
-             )   
+            c(" cases from "), transferway, 
+            sprintf(" (Fragility = %d).", total_switch),
+            c("\nNote that RIR = Fragility/[1-P("), RIRway_phrase, c(")]"))
+             )  
           
           cat("\n")
 
           ### for RIR_perc larger than 100%
           if (RIR_pi > 100){
-            cat(paste0(sprintf("Note the RIR exceeds 100%%. Generating the transfer of %d cases would", total_switch),
-                sprintf("\nrequire replacing more cases that are in the", RIRway, " condition\n"))
+            cat(paste0(sprintf("\nNote the RIR exceeds 100%%. Generating the transfer of %d cases would", total_switch),
+                c("\nrequire replacing more cases that are in the"), RIRway, c(" condition\n"))
                ) 
             }
 
@@ -549,19 +549,19 @@ p_start_chi_final <- chisq.test(final_solution$table_final,correct = FALSE)$p.va
           }
           
           #conclusion1b <- 
-          cat(paste(
+          cat(paste0(
             sprintf(" This is equivalent to transferring \n%d", final_solution$final_switch), 
-            c("cases from"), transferway, 
-            sprintf("(Fragility = %d).", total_switch),
-            sprintf("Note that RIR = Fragility/[1-P(", RIRway_phrase, ")]"))
+            c(" cases from "), transferway, 
+            sprintf(" (Fragility = %d).", total_switch),
+            c("\nNote that RIR = Fragility/[1-P("), RIRway_phrase, c(")]"))
              )  
           
      cat("\n")
 
           ### for RIR_perc larger than 100%
           if (RIR_pi > 100){
-            cat(paste0(sprintf("Note the RIR exceeds 100%%. Generating the transfer of %d cases would", total_switch),
-                sprintf("\nrequire replacing more cases that are in the", RIRway, " condition\n"))
+            cat(paste0(sprintf("\nNote the RIR exceeds 100%%. Generating the transfer of %d cases would", total_switch),
+                c("\nrequire replacing more cases that are in the"), RIRway, c(" condition\n"))
                ) 
             }
 
@@ -582,19 +582,19 @@ p_start_chi_final <- chisq.test(final_solution$table_final,correct = FALSE)$p.va
           }
           
           #conclusion1b <- 
-          cat(paste(
+          cat(paste0(
             sprintf(" This is equivalent to transferring \n%d", final_solution$final_switch), 
-            c("cases from"), transferway, 
-            sprintf("(Fragility = %d).", total_switch),
[…]
