+++ R/test_sensitivity_ln.R
@@ -458,18 +458,18 @@ p_start_chi_final <- chisq.test(final_solution$table_final,correct = FALSE)$p.va
           
           if (replace == "control") {
             #conclusion1a <- 
-            cat(sprintf("with cases for which the probability of failure in the \ncontrol group (%.0f%%) applies (RIR = %d).", prob_replace, RIR))
+            cat(sprintf("with cases for which the probability of failure in the control \ngroup (%.2f%%) applies (RIR = %d).", prob_replace, RIR))
           } else {
             #conclusion1a <- 
-            cat(sprintf("with cases for which the probability of failure in the \nentire sample (%.0f%%) applies (RIR = %d).", prob_replace, RIR))
+            cat(sprintf("with cases for which the probability of failure in the entire \nsample (%.2f%%) applies (RIR = %d).", prob_replace, RIR))
           }
           
           #conclusion1b <- 
           cat(paste0(
-            sprintf(" This is equivalent to transferring \n%d", final_solution$final_switch), 
+            sprintf(" This is equivalent to \ntransferring %d", final_solution$final_switch), 
             c(" cases from "), transferway, 
-            sprintf(" (Fragility = %d).", total_switch),
-            c("\nNote that RIR = Fragility/[1-P("), RIRway_phrase, c(")]"))
+            sprintf("\n(Fragility = %d).", total_switch),
+            c("\n\nNote that RIR = Fragility/[1-P("), RIRway_phrase, c(")]"))
              ) 
                     
 
@@ -492,18 +492,18 @@ p_start_chi_final <- chisq.test(final_solution$table_final,correct = FALSE)$p.va
           
           if (replace == "control") {
             #conclusion1a <- 
-            cat(sprintf("with cases for which the probability of failure in the \ncontrol group (%.0f%%) applies (RIR = %d).", prob_replace, RIR))
+            cat(sprintf("with cases for which the probability of failure in the control \ngroup (%.2f%%) applies (RIR = %d).", prob_replace, RIR))
           } else {
             #conclusion1a <- 
-            cat(sprintf("with cases for which the probability of failure in the \nentire sample (%.0f%%) applies (RIR = %d).", prob_replace, RIR))
+            cat(sprintf("with cases for which the probability of failure in the entire \nsample (%.2f%%) applies (RIR = %d).", prob_replace, RIR))
           }
           
           #conclusion1b <- 
           cat(paste0(
-            sprintf(" This is equivalent to transferring \n%d", final_solution$final_switch), 
+            sprintf(" This is equivalent to \ntransferring %d", final_solution$final_switch), 
             c(" cases from "), transferway, 
-            sprintf(" (Fragility = %d).", total_switch),
-            c("\nNote that RIR = Fragility/[1-P("), RIRway_phrase, c(")]"))
+            sprintf("\n(Fragility = %d).", total_switch),
+            c("\n\nNote that RIR = Fragility/[1-P("), RIRway_phrase, c(")]"))
              )  
           
           cat("\n")
@@ -546,14 +546,14 @@ p_start_chi_final <- chisq.test(final_solution$table_final,correct = FALSE)$p.va
         if (!final_solution$needtworows & final_solution$final_switch > 1) {
           #conclusion1 <- 
           cat(paste(
-            change, sprintf("one would need to replace %d \n(%.0f%%)", RIR, RIR_pi), RIRway, "cases "))
+            change, sprintf("one would need to replace %d \n(%.2f%%)", RIR, RIR_pi), RIRway, "cases "))
           
           if (replace == "control") {
             #conclusion1a <- 
[…]
