+++ R/test_sensitivity_ln.R
@@ -652,7 +652,7 @@ test_sensitivity_ln <- function(est_eff,
       print(Implied_Table)
       cat("\n")
       cat(paste(sprintf("The reported effect size = %.3f, and SE = %.3f, p-value = %.3f.",
-                        est_eff, user_std_err, p_final),
+                        est_eff, user_std_err, p_start),
                 sprintf("\nValues have been rounded to the nearest integer. This may cause"), 
                 sprintf("\na little change to the estimated effect for the table.\n\n"))
       )
