+++ R/test_sensitivity_ln.R
@@ -636,9 +636,9 @@ test_sensitivity_ln <- function(est_eff,
 
       cat("\n\n")
       print(Transfer_Table)
-      cat(sprintf("(Effect size = %.3f, SE = %.3f, p-value = %.3f. \nIndicate that",
+      cat(sprintf("Effect size = %.3f, SE = %.3f, p-value = %.3f. \nIndicate that",
                   final_solution$est_eff_final, final_solution$std_err_final, p_final),
-                 c(" this is based on t = estimated effect/standard error)")
+                 c("this is based on t = estimated effect/standard error")
          )
       
       
@@ -840,9 +840,9 @@ test_sensitivity_ln <- function(est_eff,
 
       cat("\n\n")
       print(Transfer_Table)
-      cat(sprintf("(Effect size = %.3f, SE = %.3f, p-value = %.3f. \nIndicate that",
+      cat(sprintf("Effect size = %.3f, SE = %.3f, p-value = %.3f. \nIndicate that",
                   final_solution$est_eff_final, final_solution$std_err_final, p_final),
-                 c(" this is based on t = estimated effect/standard error)")
+                 c("this is based on t = estimated effect/standard error")
          )
       
       
