+++ R/test_sensitivity_ln.R
@@ -663,9 +663,9 @@ test_sensitivity_ln <- function(est_eff,
 
       cat("\n\n")
       print(Transfer_Table)
-      cat(sprintf("Effect size = %.3f, SE = %.3f, p-value = %.3f. \n",
+      cat(sprintf("Effect size = %.3f, SE = %.3f, p-value = %.3f.",
                   final_solution$est_eff_final, final_solution$std_err_final, p_final),
-                 c("this is based on t = estimated effect/standard error")
+                 c("\nThis is based on t = estimated effect/standard error")
          )
       
       
@@ -867,9 +867,9 @@ test_sensitivity_ln <- function(est_eff,
 
       cat("\n\n")
       print(Transfer_Table)
-      cat(sprintf("Effect size = %.3f, SE = %.3f, p-value = %.3f. \n",
+      cat(sprintf("Effect size = %.3f, SE = %.3f, p-value = %.3f.",
                   final_solution$est_eff_final, final_solution$std_err_final, p_final),
-                 c("this is based on t = estimated effect/standard error")
+                 c("\nThis is based on t = estimated effect/standard error")
          )
       
       
