### 2024-01-08 konfound-project/konfound#50 @wwang93
@fabian-s Thanks to your suggestion, we roxygen2 commented the dependencies of each function, regenerated the NAMESPACE file, and updated the Description file, and in the end it basically passed the R CMD check!

> rcmdcheck::rcmdcheck(".")
── R CMD build ─────────────────────────────────────────────────────────────────────────────────────────────────
✔  checking for file ‘.../DESCRIPTION’ ...
─  preparing ‘konfound’: (393ms)
✔  creating vignettes (7.7s)
─  building ‘konfound_0.4.0.tar.gz’

── R CMD check ─────────────────────────────────────────────────────────────────────────────────────────────────
─  using log directory ‘/private/var/folders/hd/3_rrkpw50r10pbd7p9pxm_s40000gn/T/RtmpHIclFs/file119002eec5a2f/konfound.Rcheck’
─  using R version 4.3.0 (2023-04-21)
─  using platform: aarch64-apple-darwin20 (64-bit)
─  R was compiled by
       Apple clang version 14.0.0 (clang-1400.0.29.202)
       GNU Fortran (GCC) 12.2.0
─  running under: macOS Big Sur 11.5.1
─  using session charset: UTF-8
✔  checking for file ‘konfound/DESCRIPTION’ ...
─  checking extension type ... Package
─  this is package ‘konfound’ version ‘0.4.0’
─  package encoding: UTF-8
✔  checking package namespace information
✔  checking package dependencies (1.1s)
✔  checking if this is a source package ...
✔  checking if there is a namespace
✔  checking for executable files ...
N  checking for hidden files and directories ...
   Found the following hidden files and directories:
     joss-paper/.github
   These were most likely included in error. See section ‘Package
   structure’ in the ‘Writing R Extensions’ manual.
✔  checking for portable file names
✔  checking for sufficient/correct file permissions
✔  checking whether package ‘konfound’ can be installed (4.7s)
✔  checking installed package size
✔  checking package directory
✔  checking ‘build’ directory
✔  checking top-level files
✔  checking for left-over files
✔  checking index information ...
✔  checking package subdirectories ...
✔  checking R files for non-ASCII characters ...
✔  checking R files for syntax errors ...
✔  checking whether the package can be loaded (1.1s)
✔  checking whether the package can be loaded with stated dependencies (947ms)
✔  checking whether the package can be unloaded cleanly (915ms)
✔  checking whether the namespace can be loaded with stated dependencies (906ms)
✔  checking whether the namespace can be unloaded cleanly (1.1s)
✔  checking startup messages can be suppressed (2.1s)
✔  checking dependencies in R code (1s)
✔  checking S3 generic/method consistency (1.1s)
✔  checking replacement functions (910ms)
✔  checking foreign function calls (1s)
✔  checking R code for possible problems (4.1s)
✔  checking Rd files ...
✔  checking Rd metadata ...
✔  checking Rd cross-references ...
✔  checking for missing documentation entries (933ms)
✔  checking for code/documentation mismatches (2.8s)
✔  checking Rd \usage sections (1.2s)
✔  checking Rd contents ...
✔  checking for unstated dependencies in examples ...
✔  checking contents of ‘data’ directory
✔  checking data for non-ASCII characters ...
✔  checking LazyData
✔  checking data for ASCII and uncompressed saves ...
✔  checking installed files from ‘inst/doc’
✔  checking files in ‘vignettes’ ...
✔  checking examples (3.5s)
✔  checking for unstated dependencies in ‘tests’ ...
─  checking tests ...
✔  Running ‘testthat.R’ (1.7s)
✔  checking for unstated dependencies in vignettes ...
✔  checking package vignettes in ‘inst/doc’ ...
─  checking running R code from vignettes
     ‘introduction-to-konfound.Rmd’ using ‘UTF-8’... OK
    NONE
✔  checking re-building of vignette outputs (3.5s)
✔  checking PDF version of manual (2.7s)

   See
     ‘/private/var/folders/hd/3_rrkpw50r10pbd7p9pxm_s40000gn/T/RtmpHIclFs/file119002eec5a2f/konfound.Rcheck/00check.log’
   for details.

── R CMD check results ───────────────────────────────────────────────────────────────────── konfound 0.4.0 ────
Duration: 39.6s

❯ checking for hidden files and directories ... NOTE
  Found the following hidden files and directories:
    joss-paper/.github
  These were most likely included in error. See section ‘Package
  structure’ in the ‘Writing R Extensions’ manual.

0 errors ✔ | 0 warnings ✔ | 1 note ✖
