### 2024-01-09 konfound-project/konfound#50 @fabian-s
not the case for me on commit https://github.com/konfound-project/konfound/commit/78ec6f95156855bf38d10ebf82379cc6c6b841df, I see

```
==> devtools::check(document = FALSE)

══ Building ═══════════════════════════════════════════════════════════
Setting env vars:
• CFLAGS    : -Wall -pedantic -fdiagnostics-color=always
• CXXFLAGS  : -Wall -pedantic -fdiagnostics-color=always
• CXX11FLAGS: -Wall -pedantic -fdiagnostics-color=always
• CXX14FLAGS: -Wall -pedantic -fdiagnostics-color=always
• CXX17FLAGS: -Wall -pedantic -fdiagnostics-color=always
• CXX20FLAGS: -Wall -pedantic -fdiagnostics-color=always
── R CMD build ────────────────────────────────────────────────────────
✔  checking for file ‘/home/fabians/reviews/JOSS/konfound/DESCRIPTION’ ...
─  preparing ‘konfound’:
✔  creating vignettes (8.4s)
─  building ‘konfound_0.4.0.tar.gz’

══ Checking ═══════════════════════════════════════════════════════════
Setting env vars:
• _R_CHECK_CRAN_INCOMING_USE_ASPELL_           : TRUE
• _R_CHECK_CRAN_INCOMING_REMOTE_               : FALSE
• _R_CHECK_CRAN_INCOMING_                      : FALSE
• _R_CHECK_FORCE_SUGGESTS_                     : FALSE
• _R_CHECK_PACKAGES_USED_IGNORE_UNUSED_IMPORTS_: FALSE
• NOT_CRAN                                     : true
── R CMD check ────────────────────────────────────────────────────────
─  using log directory ‘/home/fabians/reviews/JOSS/konfound.Rcheck’
─  using R version 4.3.2 (2023-10-31)
─  using platform: x86_64-pc-linux-gnu (64-bit)
─  R was compiled by
       gcc (Ubuntu 9.4.0-1ubuntu1~20.04.2) 9.4.0
       GNU Fortran (Ubuntu 9.4.0-1ubuntu1~20.04.2) 9.4.0
─  running under: Linux Mint 20
─  using session charset: UTF-8
─  using options ‘--no-manual --as-cran’
✔  checking for file ‘konfound/DESCRIPTION’
─  checking extension type ... Package
─  this is package ‘konfound’ version ‘0.4.0’
─  package encoding: UTF-8
✔  checking package namespace information
✔  checking package dependencies (1.3s)
✔  checking if this is a source package
✔  checking if there is a namespace
✔  checking for executable files ...
N  checking for hidden files and directories
   Found the following hidden files and directories:
     joss-paper/.github
   These were most likely included in error. See section ‘Package
   structure’ in the ‘Writing R Extensions’ manual.
✔  checking for portable file names ...
✔  checking for sufficient/correct file permissions
✔  checking serialization versions
✔  checking whether package ‘konfound’ can be installed (3.9s)
✔  checking installed package size ...
✔  checking package directory
✔  checking for future file timestamps ...
✔  checking ‘build’ directory
N  checking top-level files
   Non-standard files/directories found at top level:
     ‘10.21105.joss.05779.pdf’ ‘joss-paper’
✔  checking for left-over files
✔  checking index information ...
✔  checking package subdirectories ...
✔  checking R files for non-ASCII characters ...
✔  checking R files for syntax errors ...
✔  checking whether the package can be loaded (480ms)
✔  checking whether the package can be loaded with stated dependencies (382ms)
✔  checking whether the package can be unloaded cleanly (382ms)
✔  checking whether the namespace can be loaded with stated dependencies (488ms)
✔  checking whether the namespace can be unloaded cleanly (576ms)
✔  checking loading without being on the library search path (568ms)
✔  checking startup messages can be suppressed (1s)
N  checking dependencies in R code (2.1s)
   Namespaces in Imports field not imported from:
     ‘mice’ ‘tibble’
     All declared Imports should be used.
✔  checking S3 generic/method consistency (555ms)
✔  checking replacement functions (462ms)
✔  checking foreign function calls (692ms)
─  checking R code for possible problems ... [21s/11s] NOTE (11.1s)
   test_cop: no visible binding for global variable ‘ModelLabel’
   test_cop: no visible binding for global variable ‘coef_X’
     ModelLabel coef_X

   Found if() conditions comparing class() to string:
   File ‘konfound/R/cop_pse_auxiliary.R’: if (class(flag_cov) == "lavaan") ...
   File ‘konfound/R/cop_pse_auxiliary.R’: if (class(flag_cor) == "lavaan") ...
   File ‘konfound/R/cop_pse_auxiliary.R’: if (class(flag_cor) == "lavaan" && class(flag_cov) == "lavaan") ...
   File ‘konfound/R/cop_pse_auxiliary.R’: if (class(flag_cov) == "lavaan") ...
   File ‘konfound/R/cop_pse_auxiliary.R’: if (class(flag_cov) == "lavaan") ...
   File ‘konfound/R/cop_pse_auxiliary.R’: if (class(flag_cov) == "lavaan") ...
   File ‘konfound/R/cop_pse_auxiliary.R’: if (class(flag_cov) == "lavaan") ...
   Use inherits() (or maybe is()) instead.
✔  checking Rd files ...
✔  checking Rd metadata ...
✔  checking Rd line widths ...
✔  checking Rd cross-references ...
✔  checking for missing documentation entries (495ms)
W  checking for code/documentation mismatches (748ms)
   Codoc mismatches from documentation object 'tkonfound_fig':
   tkonfound_fig
     Code: function(a, b, c, d, thr_p = 0.05, switch_trm = TRUE, test =
                    "fisher", replace = "control")
     Docs: function(a, b, c, d, thr_p = 0.05, switch_trm = T, test =
                    "fisher", replace = "control")
     Mismatches in argument default values:
       Name: 'switch_trm' Code: TRUE Docs: T

✔  checking Rd \usage sections (2s)
✔  checking Rd contents ...
✔  checking for unstated dependencies in examples ...
✔  checking contents of ‘data’ directory ...
✔  checking data for non-ASCII characters ...
✔  checking LazyData
✔  checking data for ASCII and uncompressed saves ...
✔  checking installed files from ‘inst/doc’ ...
✔  checking files in ‘vignettes’ ...
E  checking examples (1.5s)
   Running examples in ‘konfound-Ex.R’ failed
   The error most likely occurred in:

   > base::assign(".ptime", proc.time(), pos = "CheckExEnv")
   > ### Name: konfound
   > ### Title: Perform sensitivity analysis on fitted models
   > ### Aliases: konfound
   >
   > ### ** Examples
   >
   > # using lm() for linear models
   > m1 <- lm(mpg ~ wt + hp, data = mtcars)
   > konfound(m1, wt)
   Robustness of Inference to Replacement (RIR):
   To invalidate an inference,  66.521 % of the estimate would have to be due to bias.
   This is based on a threshold of -1.298 for statistical significance (alpha = 0.05).

   To invalidate an inference,  21  observations would have to be replaced with cases
   for which the effect is 0 (RIR = 21).

   See Frank et al. (2013) for a description of the method.

   Citation: Frank, K.A., Maroulis, S., Duong, M., and Kelcey, B. (2013).
   What would it take to change an inference?
   Using Rubin's causal model to interpret the robustness of causal inferences.
   Education, Evaluation and Policy Analysis, 35 437-460.
   For more detailed output, consider setting `to_return` to table
   To consider other predictors of interest, consider setting `test_all` to TRUE.
   > konfound(m1, wt, test_all = TRUE)
   Note that this output is calculated based on the correlation-based approach used in mkonfound()
   Warning in data.frame(t = est_eff/std_err, df = (n_obs - n_covariates -  :
     row names were found from a short variable and have been discarded
   # A tibble: 2 × 8
     var_name     t    df action       inference pct_bias_to_change_i…¹  itcv r_con
     <chr>    <dbl> <dbl> <chr>        <chr>                      <dbl> <dbl> <dbl>
   1 wt       -6.13    29 to_invalida… reject_n…                   52.7 0.614 0.784
   2 hp       -3.52    29 to_invalida… reject_n…                   35.1 0.298 0.546
   # ℹ abbreviated name: ¹​pct_bias_to_change_inference
   > konfound(m1, wt, to_return = "table")
   Dependent variable is mpg
   Warning: Unknown or uninitialised column: `itcv`.
   Error in abs(konfound(model_object, !!tested_variable, to_return = "raw_output")$itcv) :
     non-numeric argument to mathematical function
   Calls: konfound -> konfound_lm -> test_sensitivity -> output_table
   Execution halted
✔  checking for unstated dependencies in ‘tests’ ...
─  checking tests ...
─  Running ‘testthat.R’ (4.7s)
E  Some test files failed
   Last 13 lines of output:
     row names were found from a short variable and have been discarded
     Backtrace:
         ▆
      1. └─konfound::konfound(testmod1, texp, test_all = TRUE, to_return = "raw_output") at test-pkonfound.R:10:1
      2.   └─konfound:::konfound_lm(...)
      3.     └─base::data.frame(...)

     ── Failure ('test-konfound.R:30:5'): konfound works for lme4 model ─────────────
     output3$percent_bias_to_change_inference not equal to 84.826.
     target is NULL, current is numeric

     [ FAIL 1 | WARN 6 | SKIP 0 | PASS 16 ]
     Error: Test failures
     Execution halted
✔  checking for unstated dependencies in vignettes (346ms)
✔  checking package vignettes in ‘inst/doc’ ...
✔  checking re-building of vignette outputs (6.5s)
✔  checking for non-standard things in the check directory ...
✔  checking for detritus in the temp directory

   See
     ‘/home/fabians/reviews/JOSS/konfound.Rcheck/00check.log’
   for details.

── R CMD check results ──────────────────────────── konfound 0.4.0 ────
Duration: 44.4s

❯ checking examples ... ERROR
  Running examples in ‘konfound-Ex.R’ failed
  The error most likely occurred in:

  > base::assign(".ptime", proc.time(), pos = "CheckExEnv")
  > ### Name: konfound
  > ### Title: Perform sensitivity analysis on fitted models
  > ### Aliases: konfound
  >
  > ### ** Examples
  >
  > # using lm() for linear models
  > m1 <- lm(mpg ~ wt + hp, data = mtcars)
  > konfound(m1, wt)
  Robustness of Inference to Replacement (RIR):
  To invalidate an inference,  66.521 % of the estimate would have to be due to bias.
  This is based on a threshold of -1.298 for statistical significance (alpha = 0.05).

  To invalidate an inference,  21  observations would have to be replaced with cases
  for which the effect is 0 (RIR = 21).

  See Frank et al. (2013) for a description of the method.

  Citation: Frank, K.A., Maroulis, S., Duong, M., and Kelcey, B. (2013).
  What would it take to change an inference?
  Using Rubin's causal model to interpret the robustness of causal inferences.
  Education, Evaluation and Policy Analysis, 35 437-460.
  For more detailed output, consider setting `to_return` to table
  To consider other predictors of interest, consider setting `test_all` to TRUE.
  > konfound(m1, wt, test_all = TRUE)
  Note that this output is calculated based on the correlation-based approach used in mkonfound()
  Warning in data.frame(t = est_eff/std_err, df = (n_obs - n_covariates -  :
    row names were found from a short variable and have been discarded
  # A tibble: 2 × 8
    var_name     t    df action       inference pct_bias_to_change_i…¹  itcv r_con
    <chr>    <dbl> <dbl> <chr>        <chr>                      <dbl> <dbl> <dbl>
  1 wt       -6.13    29 to_invalida… reject_n…                   52.7 0.614 0.784
  2 hp       -3.52    29 to_invalida… reject_n…                   35.1 0.298 0.546
  # ℹ abbreviated name: ¹​pct_bias_to_change_inference
  > konfound(m1, wt, to_return = "table")
  Dependent variable is mpg
  Warning: Unknown or uninitialised column: `itcv`.
  Error in abs(konfound(model_object, !!tested_variable, to_return = "raw_output")$itcv) :
    non-numeric argument to mathematical function
  Calls: konfound -> konfound_lm -> test_sensitivity -> output_table
  Execution halted

❯ checking tests ...
  See below...

❯ checking for code/documentation mismatches ... WARNING
  Codoc mismatches from documentation object 'tkonfound_fig':
  tkonfound_fig
    Code: function(a, b, c, d, thr_p = 0.05, switch_trm = TRUE, test =
                   "fisher", replace = "control")
    Docs: function(a, b, c, d, thr_p = 0.05, switch_trm = T, test =
                   "fisher", replace = "control")
    Mismatches in argument default values:
      Name: 'switch_trm' Code: TRUE Docs: T

❯ checking for hidden files and directories ... NOTE
  Found the following hidden files and directories:
    joss-paper/.github
  These were most likely included in error. See section ‘Package
  structure’ in the ‘Writing R Extensions’ manual.

❯ checking top-level files ... NOTE
  Non-standard files/directories found at top level:
    ‘10.21105.joss.05779.pdf’ ‘joss-paper’

❯ checking dependencies in R code ... NOTE
  Namespaces in Imports field not imported from:
    ‘mice’ ‘tibble’
    All declared Imports should be used.

❯ checking R code for possible problems ... [21s/11s] NOTE
  test_cop: no visible binding for global variable ‘ModelLabel’
  test_cop: no visible binding for global variable ‘coef_X’
    ModelLabel coef_X

  Found if() conditions comparing class() to string:
  File ‘konfound/R/cop_pse_auxiliary.R’: if (class(flag_cov) == "lavaan") ...
  File ‘konfound/R/cop_pse_auxiliary.R’: if (class(flag_cor) == "lavaan") ...
  File ‘konfound/R/cop_pse_auxiliary.R’: if (class(flag_cor) == "lavaan" && class(flag_cov) == "lavaan") ...
  File ‘konfound/R/cop_pse_auxiliary.R’: if (class(flag_cov) == "lavaan") ...
  File ‘konfound/R/cop_pse_auxiliary.R’: if (class(flag_cov) == "lavaan") ...
  File ‘konfound/R/cop_pse_auxiliary.R’: if (class(flag_cov) == "lavaan") ...
  File ‘konfound/R/cop_pse_auxiliary.R’: if (class(flag_cov) == "lavaan") ...
  Use inherits() (or maybe is()) instead.

── Test failures ──────────────────────────────────────── testthat ────

> library(testthat)
> library(konfound)
Sensitivity analysis as described in Frank, Maroulis, Duong, and Kelcey (2013) and in Frank (2000).
For more information visit http://konfound-it.com.
>
> test_check("konfound")
[ FAIL 1 | WARN 6 | SKIP 0 | PASS 16 ]

══ Warnings ════════════════════════════════════════════════════════════════════
── Warning ('test-konfound.R:30:5'): konfound works for lme4 model ─────────────
Unknown or uninitialised column: `percent_bias_to_change_inference`.
Backtrace:
    ▆
 1. ├─testthat::expect_equal(...) at test-konfound.R:30:5
 2. │ └─testthat::quasi_label(enquo(object), label, arg = "object")
 3. │   └─rlang::eval_bare(expr, quo_get_env(quo))
 4. ├─output3$percent_bias_to_change_inference
 5. └─tibble:::`$.tbl_df`(output3, percent_bias_to_change_inference)
── Warning ('test-mkonfound.r:11:1'): (code run outside of `test_that()`) ──────
row names were found from a short variable and have been discarded
Backtrace:
    ▆
 1. └─konfound::konfound(testmod1, texp, test_all = TRUE, to_return = "raw_output") at test-mkonfound.r:11:1
 2.   └─konfound:::konfound_lm(...)
 3.     └─base::data.frame(...)
── Warning ('test-mkonfound.r:11:1'): (code run outside of `test_that()`) ──────
Use of .data in tidyselect expressions was deprecated in tidyselect 1.2.0.
ℹ Please use `"t"` instead of `.data$t`
Backtrace:
     ▆
  1. └─konfound::konfound(testmod1, texp, test_all = TRUE, to_return = "raw_output") at test-mkonfound.r:11:1
  2.   └─konfound:::konfound_lm(...)
  3.     └─konfound::mkonfound(d, .data$t, .data$df)
  4.       ├─dplyr::pull(select(d, !!t_enquo))
  5.       ├─dplyr::select(d, !!t_enquo)
  6.       └─dplyr:::select.data.frame(d, !!t_enquo)
  7.         └─tidyselect::eval_select(expr(c(...)), data = .data, error_call = error_call)
  8.           └─tidyselect:::eval_select_impl(...)
  9.             ├─tidyselect:::with_subscript_errors(...)
 10.             │ └─rlang::try_fetch(...)
 11.             │   └─base::withCallingHandlers(...)
 12.             └─tidyselect:::vars_select_eval(...)
 13.               └─tidyselect:::walk_data_tree(expr, data_mask, context_mask)
 14.                 └─tidyselect:::eval_c(expr, data_mask, context_mask)
 15.                   └─tidyselect:::reduce_sels(node, data_mask, context_mask, init = init)
 16.                     └─tidyselect:::walk_data_tree(new, data_mask, context_mask)
 17.                       └─tidyselect:::expr_kind(expr, context_mask, error_call)
 18.                         └─tidyselect:::call_kind(expr, context_mask, error_call)
── Warning ('test-mkonfound.r:11:1'): (code run outside of `test_that()`) ──────
Use of .data in tidyselect expressions was deprecated in tidyselect 1.2.0.
ℹ Please use `"df"` instead of `.data$df`
Backtrace:
     ▆
  1. └─konfound::konfound(testmod1, texp, test_all = TRUE, to_return = "raw_output") at test-mkonfound.r:11:1
  2.   └─konfound:::konfound_lm(...)
  3.     └─konfound::mkonfound(d, .data$t, .data$df)
  4.       ├─dplyr::pull(select(d, !!df_enquo))
  5.       ├─dplyr::select(d, !!df_enquo)
  6.       └─dplyr:::select.data.frame(d, !!df_enquo)
  7.         └─tidyselect::eval_select(expr(c(...)), data = .data, error_call = error_call)
  8.           └─tidyselect:::eval_select_impl(...)
  9.             ├─tidyselect:::with_subscript_errors(...)
 10.             │ └─rlang::try_fetch(...)
 11.             │   └─base::withCallingHandlers(...)
 12.             └─tidyselect:::vars_select_eval(...)
 13.               └─tidyselect:::walk_data_tree(expr, data_mask, context_mask)
 14.                 └─tidyselect:::eval_c(expr, data_mask, context_mask)
 15.                   └─tidyselect:::reduce_sels(node, data_mask, context_mask, init = init)
 16.                     └─tidyselect:::walk_data_tree(new, data_mask, context_mask)
 17.                       └─tidyselect:::expr_kind(expr, context_mask, error_call)
 18.                         └─tidyselect:::call_kind(expr, context_mask, error_call)
── Warning ('test-mkonfound.r:11:1'): (code run outside of `test_that()`) ──────
Use of .data in tidyselect expressions was deprecated in tidyselect 1.2.0.
ℹ Please use `"term"` instead of `.data$term`
Backtrace:
     ▆
  1. └─konfound::konfound(testmod1, texp, test_all = TRUE, to_return = "raw_output") at test-mkonfound.r:11:1
  2.   └─konfound:::konfound_lm(...)
  3.     ├─dplyr::select(tidy_output, var_name = .data$term)
  4.     └─dplyr:::select.data.frame(tidy_output, var_name = .data$term)
  5.       └─tidyselect::eval_select(expr(c(...)), data = .data, error_call = error_call)
  6.         └─tidyselect:::eval_select_impl(...)
  7.           ├─tidyselect:::with_subscript_errors(...)
  8.           │ └─rlang::try_fetch(...)
  9.           │   └─base::withCallingHandlers(...)
 10.           └─tidyselect:::vars_select_eval(...)
 11.             └─tidyselect:::walk_data_tree(expr, data_mask, context_mask)
 12.               └─tidyselect:::eval_c(expr, data_mask, context_mask)
 13.                 └─tidyselect:::reduce_sels(node, data_mask, context_mask, init = init)
 14.                   └─tidyselect:::walk_data_tree(new, data_mask, context_mask)
 15.                     └─tidyselect:::expr_kind(expr, context_mask, error_call)
 16.                       └─tidyselect:::call_kind(expr, context_mask, error_call)
── Warning ('test-pkonfound.R:10:1'): (code run outside of `test_that()`) ──────
row names were found from a short variable and have been discarded
Backtrace:
    ▆
 1. └─konfound::konfound(testmod1, texp, test_all = TRUE, to_return = "raw_output") at test-pkonfound.R:10:1
 2.   └─konfound:::konfound_lm(...)
 3.     └─base::data.frame(...)

── Failure ('test-konfound.R:30:5'): konfound works for lme4 model ─────────────
output3$percent_bias_to_change_inference not equal to 84.826.
target is NULL, current is numeric

[ FAIL 1 | WARN 6 | SKIP 0 | PASS 16 ]
Error: Test failures
Execution halted

2 errors ✖ | 1 warning ✖ | 4 notes ✖
Error: R CMD check found ERRORs
Execution halted

Exited with status 1.
```
which is unacceptable for R packages looking to be published on JOSS.

Additionally, none of the issues found by goodpractice::gp and pointed out by me in August seem to have been adressed sucessfully:
```
── GP konfound ────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────

It is good practice to

  ✖ write short and simple functions. These functions have high cyclomatic complexity (>50): test_sensitivity_ln
    (217), getswitch_fisher (79), tkonfound_fig (74), getswitch_chisq (70), tkonfound (70). You can make them easier to
[…]
