### 2024-01-09 konfound-project/konfound#50 @fabian-s
EDIT: maybe you did not push those latest changes to the repo --- the commit I'm on is:

commit 78ec6f95156855bf38d10ebf82379cc6c6b841df (HEAD -> master, origin/master, origin/HEAD)
Author: Qinyun Lin <qinyun.lin@gu.se>
Date:   Tue Dec 5 20:33:26 2023 +0100
