### 2024-01-09 konfound-project/konfound#50 @wwang93
@fabian-s Yes, the team hasn't merged some of the changes on the default master branch yet, we'll agree on that as soon as possible.
