### 2024-01-09 konfound-project/konfound#50 @fabian-s
another edit:

![image](https://github.com/konfound-project/konfound/assets/998541/534836fe-3ed5-4e10-a521-3680c3ec5ee6)

this is extraordinarily bad SWE hygiene -- you should not push code to main that breaks your package, that's what development branches are for...
