### 2024-01-09 konfound-project/konfound#50 @fabian-s
> @fabian-s Yes, the team hasn't merged some of the changes on the default master branch yet, we'll agree on that as soon as possible.

ok please don't @ me again before that is done and you have an actually working and standards-compliant package, ty
