+++ R/zzz.R
@@ -1,8 +1,18 @@
+#' Package Initialization Functions and Utilities
+#'
+#' These functions are used for initializing the package environment and
+#' providing utility functions for the package.
+#'
+#' @name zzz
+#' @aliases zzz
+#' @rdname zzz
+#' @importFrom utils globalVariables browseURL
+#'
 ## quiets concerns (notes) of R CMD check re: the vars that are evaluated 
 ## using non-standard evaluation
 if (getRversion() >= "2.15.1") 
     utils::globalVariables(c("inference", "key", "replace_null_cases", 
-                             "percent_bias", "val"))
+                             "percent_bias", "val","ModelLabel", "coef_X"))
 
 .onAttach <- function(libname, pkgname) {
     packageStartupMessage("Sensitivity analysis as described in Frank, 
