+++ DESCRIPTION
@@ -30,8 +30,8 @@ Description: Statistical methods that quantify the conditions necessary to
     inferences for a user's own model, a single published study and
     multiple studies respectively.
 License: MIT + file LICENSE
-URL: https://github.com/jrosen48/konfound
-BugReports: https://github.com/jrosen48/konfound/issues
+URL: https://github.com/konfound-project/konfound, https://konfound-project.github.io/konfound/
+BugReports: https://github.com/konfound-project/konfound/issues
 Depends: 
     R (>= 2.10)
 Imports: 
