+++ R/test_sensitivity_ln.R
@@ -382,7 +382,8 @@ test_sensitivity_ln <- function(est_eff,
                   p_start = p_start, p_final = p_final,
                   p_chi_start = p_start_chi, p_chi_final = p_final_chi,
                   p_fisher_start = p_start_fisher, p_fisher_final = p_final_fisher,
-		  a_1 = a1, b_1 = b1, c_1 = c1, d_1 = d1, 
+		  a_1 = a1, b_1 = b1, c_1 = c1, d_1 = d1,
+		  a = a, b = b, c = c, d = d,
                   # RIRway_phrase = RIRway_phrase,
                   final = final,
                   RIR_primary = RIR,
