+++ R/test_sensitivity_ln.R
@@ -384,6 +384,7 @@ test_sensitivity_ln <- function(est_eff,
                   p_fisher_start = p_start_fisher, p_fisher_final = p_final_fisher,
 		  a_1 = a1, b_1 = b1, c_1 = c1, d_1 = d1,
 		  a = a, b = b, c = c, d = d,
+		  table_bstart1 = table_bstart1,
                   # RIRway_phrase = RIRway_phrase,
                   final = final,
                   RIR_primary = RIR,
