+++ R/test_sensitivity_ln.R
@@ -391,6 +391,7 @@ test_sensitivity_ln <- function(est_eff,
                   RIR_supplemental = RIR_extra, 
                   RIR_perc = RIR_pi,  # need to discuss the denominator
 		  prob_replace = prob_replace,
+		  est_eff_final = final_solution$est_eff_final, std_err_final = final_solution$std_err_final, 
                   fragility_primary = final,
                   fragility_supplemental = final_extra,        
                   starting_table = final_solution$table_start,
