### 2024-01-25 812d311
Update test_sensitivity.R
should be changed after comp with stata
