+++ R/test_sensitivity.R
@@ -46,13 +46,13 @@ test_sensitivity <- function(est_eff,
 if (signsuppression == 1) warning("signsuppression is defined by a threshold of opposite sign of the estimated effect.")
   
  # if (nu != 0) warning("You entered a non-zero null hypothesis about an effect. 
-                       ITCV is calculated assuming omitted variable is equally 
-                       correlated with predictor of interest and outcome. 
-                       This approach maximizes the impact in the correlation metric 
-                       and could be applied to standardized variables. 
-                       An alternative approach is to preserve the original metric 
-                       and choose the two correlations to preserve the standard error. 
-                       See index = PSE. ")
+                      # ITCV is calculated assuming omitted variable is equally 
+                      # correlated with predictor of interest and outcome. 
+                      # This approach maximizes the impact in the correlation metric 
+                      # and could be applied to standardized variables. 
+                      # An alternative approach is to preserve the original metric 
+                      # and choose the two correlations to preserve the standard error. 
+                      # See index = PSE. ")
 
  # if (!is.na(eff_thr)) warning("The threshold you specified will be used without\n",
  #                            "regard for the standard error and statistical significance\n",
