### 2024-01-27 5583191
clean up for n_covariates == 0
if # covariates = 0, then only report unconditional partial correlations
