+++ R/test_sensitivity.R
@@ -166,15 +166,19 @@ test_sensitivity <- function(est_eff,
   r_con <- round(sqrt(abs(itcv)), 3)
 
   ## calculate the unconditional ITCV if user inputs sdx, sdy and R2
-  if (!is.na(sdx) & !is.na(sdy) & !is.na(R2)) {
+  if (!is.na(sdx) & !is.na(sdy) & !is.na(R2) & (n_covariates > 0)) {
     ## pull in the auxiliary function for R2yz and R2xz 
     r2yz <- cal_ryz(obs_r, R2)^2
     uncond_rycv <- r_con * sqrt(1 - r2yz)
     r2xz <- cal_rxz(sdx^2, sdy^2, R2, n_obs-n_covariates-2, std_err)^2
     uncond_rxcv <- r_con * sqrt(1 - r2xz)
-  } else {
-    r2yz = uncond_rycv = r2xz = uncond_rxcv = NA
-  }
+  } else if (n_covariates == 0) {
+      r2yz = r2xz = NA 
+      uncond_rycv <- r_con
+      uncond_rxcv <- r_con
+      } else {
+          r2yz = uncond_rycv = r2xz = uncond_rxcv = NA
+          }
   
   uncond_rycv = uncond_rycv * signITCV
   rycvGz = r_con * signITCV
@@ -184,6 +188,16 @@ test_sensitivity <- function(est_eff,
   ## calculate act_r using one less df or maybe -1 instead
   act_r_forVF <- act_t / sqrt(act_t^2 + n_obs - n_covariates - 2)
   r_final = (act_r_forVF - r_con * rycvGz)/sqrt((1 - r_con^2) * (1 - rycvGz^2))
+  
+  # clean up for n_covariates == 0
+  if (n_covariates == 0) {
+      rxcvGz = rycvGz = itcvGz = NA
+  } else {
+      rxcvGz = r_con
+      rycvGz = rycvGz
+      itcvGz = itcv
+  }
+  
   
   # if (component_correlations == FALSE){
   #     rsq <- # has to come from some kind of model object
@@ -235,8 +249,8 @@ test_sensitivity <- function(est_eff,
                        critical_r, r_final = r_final,
                        # rxcv always be positive, rycv goes with itcv
                        rxcv = uncond_rxcv, rycv = uncond_rycv, 
-                       rxcvGz = r_con, rycvGz = rycvGz, 
-                       itcvGz = itcv, itcv = uncond_rxcv * uncond_rycv, 
+                       rxcvGz = rxcvGz, rycvGz = rycvGz, 
+                       itcvGz = itcvGz, itcv = uncond_rxcv * uncond_rycv, 
                        r2xz = r2xz, r2yz = r2yz, 
                        delta_star = NA, delta_star_restricted = NA, 
                        delta_exact = NA, delta_pctbias = NA, 
