### 2024-01-31 826b65e
update output for n_cov == 0
