### 2024-02-02 32dcb78
Update test_sensitivity.R
align with Qunyin's update
