+++ R/test_sensitivity.R
@@ -168,18 +168,24 @@ if (signsuppression == 1) warning("signsuppression is defined by a threshold of 
   r_con <- round(sqrt(abs(itcv)), 3)
 
   ## calculate the unconditional ITCV if user inputs sdx, sdy and R2
-  if (!is.na(sdx) & !is.na(sdy) & !is.na(R2)) {
+   if (!is.na(sdx) & !is.na(sdy) & !is.na(R2) & (n_covariates > 0)) {
     ## pull in the auxiliary function for R2yz and R2xz 
     r2yz <- cal_ryz(obs_r, R2)^2
     uncond_rycv <- r_con * sqrt(1 - r2yz)
     r2xz <- cal_rxz(sdx^2, sdy^2, R2, n_obs-n_covariates-2, std_err)^2
     uncond_rxcv <- r_con * sqrt(1 - r2xz)
-  } else {
-    r2yz = uncond_rycv = r2xz = uncond_rxcv = NA
-  }
+  } else if (n_covariates == 0) {
+      r2yz = r2xz = NA 
+      uncond_rycv <- r_con
+      uncond_rxcv <- r_con
+      } else {
+          r2yz = uncond_rycv = r2xz = uncond_rxcv = NA
+          }
   
   uncond_rycv = uncond_rycv * signITCV
   rycvGz = r_con * signITCV
+  rxcvGz = r_con
+  itcvGz = itcv # conditional ITCV
   
   # verify 
   # act_r <- act_t / sqrt(act_t^2 + n_obs - n_covariates - 2)
@@ -241,8 +247,8 @@ if (signsuppression == 1) warning("signsuppression is defined by a threshold of 
     return(list(obs_r, critical_r, 
                        # rxcv always be positive, rycv goes with itcv
                        rxcv = uncond_rxcv, rycv = uncond_rycv, 
-                       rxcvGz = r_con, rycvGz = rycvGz, 
-                       itcvGz = itcv, itcv = uncond_rxcv * uncond_rycv, 
+                       rxcvGz = rxcvGz, rycvGz = rycvGz, 
+                       itcvGz = itcvGz, itcv = uncond_rxcv * uncond_rycv,
                        r2xz = r2xz, r2yz = r2yz, 
                        delta_star = NA, delta_star_restricted = NA, 
                        delta_exact = NA, delta_pctbias = NA, 
