+++ R/helper_output_print.R
@@ -101,7 +101,7 @@ if (requireNamespace("htmltools", quietly = TRUE)) {
       cat("The minimum impact of an omitted variable to invalidate an inference for a null hypothesis of 0 effect is based on a correlation of ", r_con)
       cat("\n")
       cat(" with the outcome and at ", r_con,
-          " with the predictor of interest (conditioning on observed covariates) based on a threshold of ")
+          " with the predictor of interest (conditioning on all observed covariates in the model) based on a threshold of ")
       cat("\n")
       cat(round(critical_r, 3), " for statistical significance (alpha = ", alpha, ").\n",
           sep = "")
@@ -113,7 +113,7 @@ if (requireNamespace("htmltools", quietly = TRUE)) {
       cat("The minimum (in absolute value) impact of an omitted variable to invalidate an inference for a null hypothesis of 0 effect is based on a correlation of ", -r_con)
       cat("\n")
       cat(" with the outcome and at ", r_con,
-          " with the predictor of interest (conditioning on observed covariates; signs are interchangable) based on a threshold of ")
+          " with the predictor of interest (conditioning on all observed covariates in the model; signs are interchangable) based on a threshold of ")
       cat("\n")
       cat(round(critical_r, 3), " for statistical significance (alpha = ", alpha, ").\n",
           sep = "")
@@ -125,7 +125,7 @@ if (requireNamespace("htmltools", quietly = TRUE)) {
       cat("The maximum impact (in absolute value) of an omitted variable to sustain an inference for a null hypothesis of 0 effect is based on a correlation of ", -r_con)
       cat("\n")
       cat(" with the outcome and at ", r_con,
-          " with the predictor of interest (conditioning on observed covariates; signs are interchangable) based on a threshold of ", round(beta_threshhold, 3))
+          " with the predictor of interest (conditioning on all observed covariates in the model; signs are interchangable) based on a threshold of ", round(beta_threshhold, 3))
       cat("\n")
       cat(" for statistical significance (alpha = ", alpha, ").\n",
           sep = "")
@@ -137,7 +137,7 @@ if (requireNamespace("htmltools", quietly = TRUE)) {
       cat("The maximum impact of an omitted variable to sustain an inference for a null hypothesis of 0 effect is based on a correlation of ", r_con)
       cat("\n")
       cat(" with the outcome and at ", r_con,
-          " with the predictor of interest (conditioning on observed covariates) based on a threshold of ", round(beta_threshhold, 3))
+          " with the predictor of interest (conditioning on all observed covariates in the model) based on a threshold of ", round(beta_threshhold, 3))
       cat("\n")
       cat(" for statistical significance (alpha = ", alpha, ").\n",
           sep = "")
