### 2024-02-02 b094004
Update helper_output_print.R
