### 2024-02-02 e72bc4e
Update helper_output_print.R
