+++ R/helper_output_print.R
@@ -135,12 +135,13 @@ if (requireNamespace("htmltools", quietly = TRUE)) {
       cat("\n")
       cat(-r_con, " X ", r_con, " = ", -round(r_con^2, 3), " to sustain an inference for a null hypothesis of 0 effect.\n", sep = "")
     } else if (abs(obs_r) < abs(critical_r) & obs_r < 0) {
-      cat("The maximum impact of an omitted variable to sustain an inference for a null hypothesis of 0 effect is based on a correlation of ", r_con)
+      cat("The maximum impact of an omitted variable to sustain an inference
+      cat(paste0("for a null hypothesis of 0 effect is based on a correlation of ", r_con))
       cat("\n")
-      cat(" with the outcome and at ", r_con,
-          " with the predictor of interest (conditioning on all observed covariates in the model) based on a threshold of ", round(beta_threshhold, 3))
+      cat(paste0("with the outcome and at ", r_con, " with the predictor of interest"))
+      cat(paste0("conditioning on all observed covariates in the model) based on a threshold of", round(critical_r, 3)))
       cat("\n")
-      cat(" for statistical significance (alpha = ", alpha, ").\n",
+      cat("for statistical significance (alpha = ", alpha, ").\n",
           sep = "")
       cat("\n")
       cat("Correspondingly the impact of an omitted variable (as defined in Frank 2000) must be ")
