### 2024-02-02 3be4bb6
Update helper_output_print.R
