### 2024-02-02 0a59141
Update helper_output_print.R
