### 2024-02-02 4cdacf4
Update helper_output_print.R
