+++ R/helper_output_print.R
@@ -102,9 +102,9 @@ if (requireNamespace("htmltools", quietly = TRUE)) {
       cat("\n")
       cat(paste0("for a null hypothesis of 0 effect is based on a correlation of ", r_con))
       cat("\n")
-      cat(paste0("with the outcome and at ", r_con, " with the predictor of interest"))
+      cat(paste0("with the outcome and at ", r_con, " with the predictor of interest (conditioning"))
       cat("\n")
-      cat(paste0("conditioning on all observed covariates in the model) based on a threshold of", round(critical_r, 3)))
+      cat(paste0("on all observed covariates in the model) based on a threshold of", round(critical_r, 3)))
       cat("\n")
       cat("for statistical significance (alpha = ", alpha, ").\n",
           sep = "")
@@ -117,13 +117,13 @@ if (requireNamespace("htmltools", quietly = TRUE)) {
       cat("\n")
       cat(paste0("invalidate an inference for a null hypothesis of 0 effect is based"))
       cat("\n")
-      cat(paste0("on a correlation of ", -r_con, " with the outcome and at ", r_con))
+      cat(paste0("on a correlation of ", -r_con, " with the outcome and at ", r_con, " with the predictor"))
       cat("\n")
-      cat("with the predictor of interest (conditioning on all observed covariates")
+      cat("of interest (conditioning on all observed covariates in the model; signs are interchangeable)")
       cat("\n")
-      cat(paste0("in the model; signs are interchangable) based on a threshold of ", round(critical_r, 3)))
+      cat(paste0("based on a threshold of ", round(beta_threshhold, 3), "for statistical significance))
       cat("\n")
-      cat("for statistical significance (alpha = ", alpha, ").\n", sep = "")
+      cat("(alpha = ", alpha, ").\n", sep = "")
       cat("\n")
       cat("Correspondingly the impact of an omitted variable (as defined in Frank 2000) must be ") 
       cat("\n")
@@ -133,13 +133,13 @@ if (requireNamespace("htmltools", quietly = TRUE)) {
       cat("\n")
       cat("sustain an inference for a null hypothesis of 0 effect is based on")
       cat("\n")
-      cat(paste0("a correlation of ", -r_con, "with the outcome an at ", r_con))
+      cat(paste0("a correlation of ", -r_con, " with the outcome an at ", r_con, " with the predictor"))
       cat("\n")
-      cat("with the predictor of interest (conditioning on all observed covariates")
+      cat("of interest (conditioning on all observed covariates in the model; signs are interchangeable)")
       cat("\n")
-      cat(paste0("in the model; signs are interchangable) based on a threshold of ", round(beta_threshhold, 3)))
+      cat(paste0("based on a threshold of ", round(beta_threshhold, 3), "for statistical significance))
       cat("\n")
-      cat("for statistical significance (alpha = ", alpha, ").\n", sep = "")
+      cat("(alpha = ", alpha, ").\n", sep = "")
       cat("\n")
       cat("Correspondingly the impact of an omitted variable (as defined in Frank 2000) must be ")
       cat("\n")
@@ -149,9 +149,9 @@ if (requireNamespace("htmltools", quietly = TRUE)) {
       cat("\n")
       cat(paste0("for a null hypothesis of 0 effect is based on a correlation of ", r_con))
       cat("\n")
-      cat(paste0("with the outcome and at ", r_con, " with the predictor of interest"))
+      cat(paste0("with the outcome and at ", r_con, " with the predictor of interest (conditioning"))
       cat("\n")
-      cat(paste0("conditioning on all observed covariates in the model) based on a threshold of", round(critical_r, 3)))
[…]
