### 2024-02-02 809ee8e
Update helper_output_print.R
