### 2024-02-02 02e256f
Update helper_output_print.R
