### 2024-02-02 737537f
Update helper_output_print.R
