### 2024-02-02 43cee87
Update helper_output_print.R
