+++ R/helper_output_print.R
@@ -178,10 +178,11 @@ if (requireNamespace("htmltools", quietly = TRUE)) {
     
     link_html <- '<html><body><a href="https://journals.sagepub.com/doi/10.1177/0049124100029002001">Click here for the article</a></body></html>'
     if (requireNamespace("htmltools", quietly = TRUE)) {
-      htmltools::html_print(htmltools::HTML(link_html))
+      invisible(htmltools::html_print(htmltools::HTML(link_html)))
     } else {
       message("htmltools package is required")
     }
+
 
   }
     
