### 2024-02-02 c6cb238
Update helper_output_print.R
