+++ R/helper_output_print.R
@@ -176,12 +176,12 @@ if (requireNamespace("htmltools", quietly = TRUE)) {
     cat("\n")
     cat("Accuracy of results increases with the number of decimals reported")
     
-    link_html <- '<html><body><a href="https://journals.sagepub.com/doi/10.1177/0049124100029002001">Click here for the article</a></body></html>'
-    if (requireNamespace("htmltools", quietly = TRUE)) {
-      capture.output(htmltools::html_print(htmltools::HTML(link_html)), file = NULL)
-    } else {
-      message("htmltools package is required")
-    }
+link_html <- '<html><body><a href="https://journals.sagepub.com/doi/10.1177/0049124100029002001">Click here for the article</a></body></html>'
+if (requireNamespace("htmltools", quietly = TRUE)) {
+  suppressed_output <- capture.output(htmltools::html_print(htmltools::HTML(link_html)), file = NULL)
+} else {
+  message("htmltools package is required")
+}
 
 
   }
