### 2024-02-02 2f9903c
Update helper_output_print.R
