### 2024-02-02 8678fd0
Update helper_output_print.R
