### 2024-02-02 23b9507
Update helper_output_print.R
