### 2024-02-02 343faf2
Update helper_output_print.R
