+++ R/helper_output_print.R
@@ -109,7 +109,7 @@ if (requireNamespace("htmltools", quietly = TRUE)) {
       cat("\n")
       cat(paste0("with the outcome and at ", r_con, " with the predictor of interest (conditioning"))
       cat("\n")
-      cat(paste0("on all observed covariates in the model) based on a threshold of", round(critical_r, 3)))
+      cat(paste0("on all observed covariates in the model) based on a threshold of ", round(critical_r, 3)))
       cat("\n")
       cat("for statistical significance (alpha = ", alpha, ").\n",
           sep = "")
@@ -126,9 +126,9 @@ if (requireNamespace("htmltools", quietly = TRUE)) {
       cat("\n")
       cat("of interest (conditioning on all observed covariates in the model;")
       cat("\n")
-      cat(paste0("signs are interchangeable) based on a threshold of ", round(beta_threshhold, 3)))
+      cat(paste0("signs are interchangeable) based on a threshold of ", round(beta_threshhold, 3), "for"))
       cat("\n")
-      cat("for statistical significance (alpha = ", alpha, ").\n", sep = "")
+      cat("statistical significance (alpha = ", alpha, ").\n", sep = "")
       cat("\n")
       cat("Correspondingly the impact of an omitted variable (as defined in Frank 2000) must be ") 
       cat("\n")
@@ -142,9 +142,9 @@ if (requireNamespace("htmltools", quietly = TRUE)) {
       cat("\n")
       cat("of interest (conditioning on all observed covariates in the model;")
       cat("\n")
-      cat(paste0("signs are interchangeable) based on a threshold of ", round(beta_threshhold, 3)))
+      cat(paste0("signs are interchangeable) based on a threshold of ", round(beta_threshhold, 3), "for"))
       cat("\n")
-      cat("for statistical significance (alpha = ", alpha, ").\n", sep = "")
+      cat("statistical significance (alpha = ", alpha, ").\n", sep = "")
       cat("\n")
       cat("Correspondingly the impact of an omitted variable (as defined in Frank 2000) must be ")
       cat("\n")
@@ -171,9 +171,9 @@ if (requireNamespace("htmltools", quietly = TRUE)) {
     cat("\n")
     cat(crayon::underline("Citation:"))
     cat("\n")
-    cat("Frank, K. (2000). Impact of a confounding variable on the")
+    cat("Frank, K. (2000). Impact of a confounding variable on the inference of a regression coefficient.")
     cat("\n")
-    cat("inference of a regression coefficient.", crayon::italic("Sociological Methods and Research, 29"), "(2), 147-194")
+    cat(crayon::italic("Sociological Methods and Research, 29"), "(2), 147-194")
     cat("\n")
     cat("\n")
     cat("Accuracy of results increases with the number of decimals reported.")
