### 2024-02-02 de82072
Update helper_output_print.R
