### 2024-02-02 9ed2a4b
Update helper_output_print.R
