+++ R/helper_output_print.R
@@ -126,7 +126,7 @@ if (requireNamespace("htmltools", quietly = TRUE)) {
       cat("\n")
       cat("of interest (conditioning on all observed covariates in the model;")
       cat("\n")
-      cat(paste0("signs are interchangeable) based on a threshold of ", round(beta_threshhold, 3), "for"))
+      cat(paste0("signs are interchangeable) based on a threshold of ", round(beta_threshhold, 3), " for"))
       cat("\n")
       cat("statistical significance (alpha = ", alpha, ").\n", sep = "")
       cat("\n")
@@ -142,7 +142,7 @@ if (requireNamespace("htmltools", quietly = TRUE)) {
       cat("\n")
       cat("of interest (conditioning on all observed covariates in the model;")
       cat("\n")
-      cat(paste0("signs are interchangeable) based on a threshold of ", round(beta_threshhold, 3), "for"))
+      cat(paste0("signs are interchangeable) based on a threshold of ", round(beta_threshhold, 3), " for"))
       cat("\n")
       cat("statistical significance (alpha = ", alpha, ").\n", sep = "")
       cat("\n")
@@ -171,9 +171,9 @@ if (requireNamespace("htmltools", quietly = TRUE)) {
     cat("\n")
     cat(crayon::underline("Citation:"))
     cat("\n")
-    cat("Frank, K. (2000). Impact of a confounding variable on the inference of a regression coefficient.")
+    cat("Frank, K. (2000). Impact of a confounding variable on the inference of")
     cat("\n")
-    cat(crayon::italic("Sociological Methods and Research, 29"), "(2), 147-194")
+    cat("a regression coefficient.", crayon::italic("Sociological Methods and Research, 29"), "(2), 147-194")
     cat("\n")
     cat("\n")
     cat("Accuracy of results increases with the number of decimals reported.")
