### 2024-02-02 84bfcf9
Update helper_output_print.R
