+++ R/helper_output_print.R
@@ -171,9 +171,9 @@ if (requireNamespace("htmltools", quietly = TRUE)) {
     cat("\n")
     cat(crayon::underline("Citation:"))
     cat("\n")
-    cat("Frank, K. (2000). Impact of a confounding variable on the inference of")
+    cat("Frank, K. (2000). Impact of a confounding variable on the inference of a")
     cat("\n")
-    cat("a regression coefficient.", crayon::italic("Sociological Methods and Research, 29"), "(2), 147-194")
+    cat("regression coefficient.", crayon::italic("Sociological Methods and Research, 29"), "(2), 147-194")
     cat("\n")
     cat("\n")
     cat("Accuracy of results increases with the number of decimals reported.")
