+++ R/test_sensitivity_ln.R
@@ -366,7 +366,7 @@ test_sensitivity_ln <- function(est_eff,
   
   # output dispatch
   if (to_return == "raw_output") {
-      return(output_list(obs_r = NA, act_r = NA, 
+      return(list(obs_r = NA, act_r = NA, 
                   critical_r = NA, r_final = NA,
                   rxcv = NA, rycv = NA, 
                   rxcvGz = NA, rycvGz = NA, 
@@ -383,25 +383,27 @@ test_sensitivity_ln <- function(est_eff,
                   fragility_primary = final,
                   fragility_supplemental = final_extra,
                   ## to see intermediate outputs
-                  # thr_t = thr_t, t_ob = t_ob,
-                  # t_start = final_solution$t_start, t_final = final_solution$t_final,   
-                  # p_start = p_start, p_final = p_final,
-                  # p_chi_start = p_start_chi, p_chi_final = p_final_chi,
-                  # p_fisher_start = p_start_fisher, p_fisher_final = p_final_fisher,
-		              # a_1 = a1, b_1 = b1, c_1 = c1, d_1 = d1,
-		              # a = a, b = b, c = c, d = d,
-		              # table_bstart1 = table_bstart1,
+                  thr_t = thr_t, t_ob = t_ob,
+                  t_start = final_solution$t_start, t_final = final_solution$t_final,   
+                  p_start = p_start, p_final = p_final,
+                  p_chi_start = p_start_chi, p_chi_final = p_final_chi,
+                  p_fisher_start = p_start_fisher, p_fisher_final = p_final_fisher,
+		  a_1 = a1, b_1 = b1, c_1 = c1, d_1 = d1,
+		  a_2 = a2, b_2 = b2, c_2 = c2, d_2 = d2,
+		  check1 = check1, check2 = check2,
+		  final_switch1 = final_switch1, final_switch2 = final_switch2,
+		  final = final, dcroddsratio_ob = dcroddsratio_ob, final_extra = final_extra,
+		  a = a, b = b, c = c, d = d,
+		  # table_bstart1 = table_bstart1,
                   # RIRway_phrase = RIRway_phrase,
-                  # final = final, dcroddsratio_ob = dcroddsratio_ob, final_extra = final_extra,
-                  # RIR_primary = RIR,
-                  # RIR_supplemental = RIR_extra, 
-                  # RIR_perc = RIR_pi,  # need to discuss the denominator
-		              # prob_replace = prob_replace, needtworows = final_solution$needtworows,
-		              # est_eff_start = final_solution$est_eff_start, std_err_start = final_solution$std_err_start, 
-		              # est_eff_final = final_solution$est_eff_final, std_err_final = final_solution$std_err_final, 
-                  # fragility_primary = final,
-                  # fragility_supplemental = final_extra,        
-
+                  RIR_primary = RIR,
+                  RIR_supplemental = RIR_extra, 
+                  RIR_perc = RIR_pi,  # need to discuss the denominator
+		  prob_replace = prob_replace, needtworows = final_solution$needtworows,
+		  est_eff_start = final_solution$est_eff_start, std_err_start = final_solution$std_err_start, 
+		  est_eff_final = final_solution$est_eff_final, std_err_final = final_solution$std_err_final, 
+                  fragility_primary = final,
+                  fragility_supplemental = final_extra,        
                   starting_table = final_solution$table_start,
                   final_table = final_solution$table_final,
                   user_SE = user_std_err,
