### 2024-02-06 8f4e67c
Update nonlinear_auxiliary.R
