### 2024-02-06 f369ae5
Update test_sensitivity_ln.R
temporary change of %d to %f for more precision
