+++ R/test_sensitivity_ln.R
@@ -495,7 +495,7 @@ test_sensitivity_ln <- function(est_eff,
           
           #conclusion1b <- 
           cat(paste0(
-            sprintf(" This is equivalent to transferring \n%d", final_solution$final_switch), 
+            sprintf(" This is equivalent to transferring \n.1%f", final_solution$final_switch), 
             c(" cases from "), transferway, 
             sprintf(" (Fragility = %d).", total_switch),
             c("\n\nNote that RIR = Fragility/[1-P("), RIRway_phrase, c(")]"))
@@ -529,7 +529,7 @@ test_sensitivity_ln <- function(est_eff,
           
           #conclusion1b <- 
           cat(paste0(
-            sprintf(" This is equivalent to transferring \n%d", final_solution$final_switch), 
+            sprintf(" This is equivalent to transferring \n.1%f", final_solution$final_switch), 
             c(" cases from "), transferway, 
             sprintf(" (Fragility = %d).", total_switch),
             c("\n\nNote that RIR = Fragility/[1-P("), RIRway_phrase, c(")]"))
@@ -587,7 +587,7 @@ test_sensitivity_ln <- function(est_eff,
           
           #conclusion1b <- 
           cat(paste0(
-            sprintf(" This is equivalent \nto transferring %d", final_solution$final_switch), 
+            sprintf(" This is equivalent \nto transferring .1%f", final_solution$final_switch), 
             c(" cases from "), transferway, 
             sprintf("\n(Fragility = %d).", total_switch),
             c("\n\nNote that RIR = Fragility/[1-P("), RIRway_phrase, c(")]"))
@@ -620,7 +620,7 @@ test_sensitivity_ln <- function(est_eff,
           
           #conclusion1b <- 
           cat(paste0(
-            sprintf(" This is equivalent to transferring \n%d", final_solution$final_switch), 
+            sprintf(" This is equivalent to transferring \n.1%f", final_solution$final_switch), 
             c(" cases from "), transferway, 
             sprintf(" (Fragility = %d).", total_switch),
             c("\n\nNote that RIR = Fragility/[1-P("), RIRway_phrase, c(")]"))
@@ -698,7 +698,7 @@ test_sensitivity_ln <- function(est_eff,
           
           #conclusion1b <- 
           cat(paste0(
-            sprintf(" This is equivalent to transferring \n%d", final_solution$final_switch), 
+            sprintf(" This is equivalent to transferring \n.1%f", final_solution$final_switch), 
             c(" cases from "), transferway, 
             sprintf(" (Fragility = %d).", total_switch),
             c("\n\nNote that RIR = Fragility/[1-P("), RIRway_phrase, c(")]"))
@@ -730,7 +730,7 @@ test_sensitivity_ln <- function(est_eff,
           
           #conclusion1b <- 
           cat(paste0(
-            sprintf(" This is equivalent to transferring \n%d", final_solution$final_switch), 
+            sprintf(" This is equivalent to transferring \n.1%f", final_solution$final_switch), 
             c(" cases from "), transferway, 
             sprintf(" (Fragility = %d).", total_switch),
             c("\n\nNote that RIR = Fragility/[1-P("), RIRway_phrase, c(")]"))
@@ -788,7 +788,7 @@ test_sensitivity_ln <- function(est_eff,
           
           #conclusion1b <- 
           cat(paste0(
-            sprintf(" This is equivalent \nto transferring %d", final_solution$final_switch), 
+            sprintf(" This is equivalent \nto transferring .1%f", final_solution$final_switch), 
             c(" cases from "), transferway, 
             sprintf("\n(Fragility = %d).", total_switch),
[…]
