### 2024-02-06 318954b
Update nonlinear_auxiliary.R
