### 2024-02-07 52ff5a7
Update nonlinear_auxiliary.R
