### 2024-02-07 4fbe9fa
Update nonlinear_auxiliary.R
