+++ tests/testthat/test-konfound.R
@@ -28,10 +28,26 @@ output4 <- konfound(m4, condition, two_by_two = TRUE, n_treat = 55, to_return = 
 
 ## these both need updating
 
-# test_that("konfound works for lme4 model", {
-#     expect_equal(output3$percent_bias_to_change_inference, 84.826, tolerance = .001)
-# })
+test_that("konfound works for lme4 model", {
+    expect_equal(output3$percent_bias_to_change_inference, 84.826, tolerance = .001)
+})
 
-# test_that("konfound works for glm, 2x2 model", {
-#     expect_equal(output4$RIR, 15)
-# })
+test_that("konfound works for glm, 2x2 model", {
+    expect_equal(output4$RIR, 15)
+})
+
+m5 <- lm(mpg ~ wt + hp, data = mtcars)
+output5 <- konfound(m1, wt, to_return = "table")
+
+test_that("konfound returns a tibble", {
+    expect_s3_class(output5, "tbl_df")
+})
+
+gss_cat$married <- ifelse(gss_cat$marital == "Married", 1, 0)
+
+m6 <- glm(married ~ age, data = gss_cat, family = binomial(link = "logit"))
+m6_output <- konfound(m6, age, to_return = "raw_output")
+
+test_that("konfound glm works", {
+    expect_equal(as.vector(m6_output$percent_bias_to_change_inference), 35.357)
+})
