+++ RIR, unconITCV and ITCV for linear-compare R with STATA.R
@@ -1,155 +1,0 @@
-library(evaluate)
-library(tidyverse)
-library(githubinstall)
-library(gridExtra)
-
-################################################################################
-########################## RIR and ITCV for linear ########################
-
-# loop to create testdata based on the different scenarios
-### estimated effect from -100, -2.5, -0.5, 0, 0.5, 2.5, 100
-### standard error is always 1
-### n_obs 20 and 10000
-### n_covariates 0 10 20 
-### tails 1 and 2
-### nu -1, 0, and 1
-### Later: alternative threshold estimated effect +0.25, -0.25, +0, *(-1) - this one is not ready
-
-# Step 1: first is to define the testdata
-##library(githubinstall)
-##gh_install_packages("konfound", ref = "newitcv")
-columns = c("est_eff", "std_err", "n_obs", "n_coviariates",
-            "tail", "nu", "sdx","sdy","rs","ITCV", "unconITCV","RIR", "error_message")
-
-df = data.frame(matrix(nrow = 0, ncol = length(columns))) 
-colnames(df) = columns
-
-est_eff_list = c(-100, -2.5, -0.5, 0, 0.5, 2.5, 100)
-std_err = 1
-n_obs_list = c(20, 10000)
-n_covariates_list = c(0, 10, 20)
-tail_list = c(1, 2)
-nu_list = c(-1, 0, 1)
-sdx_list = c(0.5, 0.75, 1, 1.25)
-sdy_list = c(0.5, 0.75, 1, 1.25)
-rsquare_list = c(0.1, 0.2, 0.5)
-nrow = 0
-
-for (i in est_eff_list) {
-  for (j in std_err) {
-    for (k in n_obs_list) {
-      for (l in n_covariates_list) {
-        for (m in tail_list) {
-          for (n in nu_list) {
-            for(o in sdx_list){
-              for(p in sdy_list){
-                for(q in rsquare_list){
-                  nrow = nrow + 1
-                  df[nrow,] = c(i, j, k, l, m, n, o, p, q, NA, NA, NA,NA)
-                }
-              }
-            }
-          }
-        }
-      }
-    }
-  }
-}
-
-df
-testdata_R <- df
-testdata_Stata <- df
-
-## Step 2: use the pkonfond method(new_itcv) in R to test
-### Note that if the function generates any error message and stop running then save the error message
-#### output$replace_null_cases >> RIR
-#### output$itcv >> itcvGz
-
-library(konfound)
-
-for(i in 1:nrow(testdata_R)){
-  e <- suppressWarnings(
-    try_capture_stack(
-      pkonfound(est_eff = testdata_R$est_eff[i], 
-                std_err = testdata_R$std_err[i], 
-                n_obs = testdata_R$n_obs[i], 
-                n_covariates = testdata_R$n_coviariates[i],
-                tails = testdata_R$tail[i],
-                nu = testdata_R$nu[i],
-                sdx = testdata_R$sdx[i],
-                sdy = testdata_R$sdy[i],
-                R2 = testdata_R$rs[i],
-                to_return = "raw_output"), 
-      env = parent.frame()))
-  if(is.error(e)) {
-    testdata_R$RIR[i] <- testdata_R$ITCV[i] <- NA
-    testdata_R$error_message[i] <- e$message
-  } else {
-    output <- pkonfound(est_eff = testdata_R$est_eff[i], 
-                        std_err = testdata_R$std_err[i], 
-                        n_obs = testdata_R$n_obs[i], 
[…]
