+++ NAMESPACE
@@ -7,7 +7,10 @@ export(mkonfound)
 export(pkonfound)
 export(tkonfound_fig)
 import(dplyr)
+import(ggplot2)
+import(lavaan)
 import(rlang)
+importFrom(dplyr,select)
 importFrom(rlang,.data)
 importFrom(stats,chisq.test)
 importFrom(stats,fisher.test)
