+++ NAMESPACE
@@ -1,10 +1,10 @@
 # Generated by roxygen2: do not edit by hand
 
-S3method(summary,konfound)
 export(konfound)
 export(launch_shiny)
 export(mkonfound)
 export(pkonfound)
+export(summary.konfound)
 export(tkonfound_fig)
 import(dplyr)
 import(ggplot2)
