### 2024-02-09 bd83c30
remove lme4 code from vignette
