### 2024-02-09 d57fce3
remove lme4 example from konfound
