### 2024-02-09 2810fd0
undo lme4 edits
trying another approach
