### 2024-02-09 56b2c28
Revert "undo lme4 edits"
This reverts commit 2810fd0ef397c3098e03f2527d9e23a98c6d12bd.
