+++ tests/testthat/test-pkonfound.R
@@ -5,10 +5,10 @@ library(mice)
 
 popmis <- popmis[1:100, ]
 testmod1 <- lm(teachpop ~ texp + sex, data = popmis)
-# testmod2 <- lmer(teachpop ~ texp + sex + (1 | school), data = popmis)
+testmod2 <- lmer(teachpop ~ texp + sex + (1 | school), data = popmis)
 
 output1 <- konfound(testmod1, texp, test_all = TRUE, to_return = "raw_output")
-#output2 <- konfound(testmod2, texp, test_all = TRUE, to_return = "raw_output")
+output2 <- konfound(testmod2, texp, test_all = TRUE, to_return = "raw_output")
 
 output3 <- pkonfound(a = 35, b = 17, c = 17, d = 38, to_return = "raw_output")
 
