+++ DESCRIPTION
@@ -52,7 +52,7 @@ Suggests:
     forcats,
     ggrepel,
     knitr,
-    lme4,
+    lme4 (>=1.1-35.1),
     Matrix (>=1.6-2),
     margins,
     pbkrtest,
