### 2024-02-10 59ee2a1
Update nonlinear_auxiliary.R
