### 2024-02-12 1f59555
Update nonlinear_auxiliary.R
