### 2024-02-12 02cf243
Update nonlinear_auxiliary.R
