### 2024-02-12 #5779 @jrosen48
@fabian-s apologies for the delay. We have been working hard on addressing the issues (see https://github.com/konfound-project/konfound/commits/master/). I will plan to update on the specific, remaining issues this week, with the plan being to address any lingering issues well before that Feb 28 date.
