### 2024-02-12 #5779 @jrosen48
One process note I'll add is that we had the incorrect badge for the R CMD CHECK via GitHub Actions, and now have the correct badge added to the repository. In addition to some sloppiness on our part regarding using the `master` branch, we also had an issue with a package dependency that was causing some builds to fail. Sorry again for the frustrating interactions - I take responsibility for them and we are moving forward productively with some lessons learned.
