+++ tests/testthat/test-pkonfound.R
@@ -12,44 +12,54 @@ output2 <- konfound(testmod2, texp, test_all = TRUE, to_return = "raw_output")
 
 output3 <- pkonfound(a = 35, b = 17, c = 17, d = 38, to_return = "raw_output")
 
+output3_b <- pkonfound(a = 5, b = 17, c = 8, d = 38, to_return = "raw_output")
+output3_c <- pkonfound(a = 5, b = 17, c = 8, d = 38, to_return = "raw_output", switch_trm = FALSE)
+output3_d <- pkonfound(a = 5, b = 17, c = 8, d = 38, to_return = "raw_output", 
+                       switch_trm = FALSE,
+                       test = "chisq")
+
+output3_e <- pkonfound(a = 5, b = 17, c = 8, d = 38, to_return = "raw_output", 
+                       switch_trm = FALSE,
+                       test = "fisher")
+
 test_that("pkonfound test positive, significant coefficient works", {
-  expect_equal(dplyr::pull(pkonfound(2, .4, 100, 3, to_return = "raw_output")[1, 3]), tolerance = .001, .603 * 100) # pct bias
-  expect_equal(dplyr::pull(pkonfound(2, .4, 100, 3, to_return = "raw_output")[1, 7]), tolerance = .001, 0.568) # r-corr
-  expect_equal(dplyr::pull(pkonfound(2, .4, 100, 3, to_return = "raw_output")[1, 8]), tolerance = .001, 0.322) # impact
+    expect_equal(dplyr::pull(pkonfound(2, .4, 100, 3, to_return = "raw_output")[1, 3]), tolerance = .001, .603 * 100) # pct bias
+    expect_equal(dplyr::pull(pkonfound(2, .4, 100, 3, to_return = "raw_output")[1, 7]), tolerance = .001, 0.568) # r-corr
+    expect_equal(dplyr::pull(pkonfound(2, .4, 100, 3, to_return = "raw_output")[1, 8]), tolerance = .001, 0.322) # impact
 })
 
 test_that("pkonfound test negative, significant coefficient works", {
-  expect_equal(dplyr::pull(pkonfound(-2, .4, 100, 3, to_return = "raw_output")[1, 3]), tolerance = .001, .603 * 100)
-  expect_equal(dplyr::pull(pkonfound(-2, .4, 100, 3, to_return = "raw_output")[1, 7]), tolerance = .001, 0.568)
-  expect_equal(dplyr::pull(pkonfound(-2, .4, 100, 3, to_return = "raw_output")[1, 8]), tolerance = .001, -0.322)
+    expect_equal(dplyr::pull(pkonfound(-2, .4, 100, 3, to_return = "raw_output")[1, 3]), tolerance = .001, .603 * 100)
+    expect_equal(dplyr::pull(pkonfound(-2, .4, 100, 3, to_return = "raw_output")[1, 7]), tolerance = .001, 0.568)
+    expect_equal(dplyr::pull(pkonfound(-2, .4, 100, 3, to_return = "raw_output")[1, 8]), tolerance = .001, -0.322)
 })
 
 test_that("pkonfound test positive, not significant coefficient works", {
-  # expect_equal(dplyr::pull(pkonfound(1, .4, 100, 3, to_return = "raw_output")[1, 3]), tolerance = .001, .206 * 100)
-  expect_equal(dplyr::pull(pkonfound(1, .4, 100, 3, to_return = "raw_output")[1, 7]), tolerance = .001, 0.248)
-  expect_equal(dplyr::pull(pkonfound(1, .4, 100, 3, to_return = "raw_output")[1, 8]), tolerance = .001, 0.061)
+    # expect_equal(dplyr::pull(pkonfound(1, .4, 100, 3, to_return = "raw_output")[1, 3]), tolerance = .001, .206 * 100)
[…]
