+++ R/pkonfound.R
@@ -67,6 +67,7 @@
 #' pkonfound(est_eff = .4, std_err = .1, n_obs = 290, sdx = 2, sdy = 6, R2 = .7, eff_thr = 0, FR2max = .8, index = "COP", to_return = "raw_output")
 #' # use pkonfound to calculate rxcv and rycv when preserving standard error
 #' pkonfound(est_eff = .5, std_err = .056, n_obs = 6174, eff_thr = .1, sdx = 0.22, sdy = 1, R2 = .3, index = "PSE", to_return = "raw_output")
+#' 
 #' @export
 
  pkonfound <- function(est_eff,
