+++ tests/testthat/test-pkonfound.R
@@ -1,26 +1,5 @@
 context("Checking pkonfound")
 
-library(lme4)
-library(mice)
-
-popmis <- popmis[1:100, ]
-testmod1 <- lm(teachpop ~ texp + sex, data = popmis)
-testmod2 <- lmer(teachpop ~ texp + sex + (1 | school), data = popmis)
-
-output1 <- konfound(testmod1, texp, test_all = TRUE, to_return = "raw_output")
-output2 <- konfound(testmod2, texp, test_all = TRUE, to_return = "raw_output")
-
-output3 <- pkonfound(a = 35, b = 17, c = 17, d = 38, to_return = "raw_output")
-
-output3_b <- pkonfound(a = 5, b = 17, c = 8, d = 38, to_return = "raw_output")
-output3_c <- pkonfound(a = 5, b = 17, c = 8, d = 38, to_return = "raw_output", switch_trm = FALSE)
-output3_d <- pkonfound(a = 5, b = 17, c = 8, d = 38, to_return = "raw_output", 
-                       switch_trm = FALSE,
-                       test = "chisq")
-
-output3_e <- pkonfound(a = 5, b = 17, c = 8, d = 38, to_return = "raw_output", 
-                       switch_trm = FALSE,
-                       test = "fisher")
 
 test_that("pkonfound test positive, significant coefficient works", {
     expect_equal(dplyr::pull(pkonfound(2, .4, 100, 3, to_return = "raw_output")[1, 3]), tolerance = .001, .603 * 100) # pct bias
@@ -46,31 +25,38 @@ test_that("pkonfound test negative, not significant coefficient works", {
     expect_equal(dplyr::pull(pkonfound(-1, .4, 100, 3, to_return = "raw_output")[1, 8]), tolerance = .001, -.061)
 })
 
-# test_that("pkonfound test positive, significant coefficient, small sample size works", {
-# expect_equal(dplyr::pull(pkonfound(2, .3, 20, 3, to_return = "raw_output")[1, 3]), tolerance = .001, .682 * 100)
-# expect_equal(dplyr::pull(pkonfound(2, .3, 20, 3, to_return = "raw_output")[1, 7]), tolerance = .001, 0.865)
-# expect_equal(dplyr::pull(pkonfound(2, .3, 20, 3, to_return = "raw_output")[1, 8]), tolerance = .001, 0.748)
-# })
-# above to be updated later 
-
 test_that("tkonfound for two_by_two works", {
+    output3 <- pkonfound(a = 35, b = 17, c = 17, d = 38, to_return = "raw_output")
+    
     expect_equal(output3$RIR, tolerance = .001, 14)
 })
 
 test_that("test_all works for lm and lmer outout", {
+    library(lme4)
+    library(mice)
+    
+    popmis <- popmis[1:100, ]
+    testmod1 <- lm(teachpop ~ texp + sex, data = popmis)
+    testmod2 <- lmer(teachpop ~ texp + sex + (1 | school), data = popmis)
+    
+    output1 <- konfound(testmod1, texp, test_all = TRUE, to_return = "raw_output")
+    output2 <- konfound(testmod2, texp, test_all = TRUE, to_return = "raw_output")
+    
     expect_is(output1, "data.frame")
     expect_is(output2, "data.frame")
 })
 
-thresh_plot <- pkonfound(2, .4, 100, 3, to_return = "thresh_plot")
 
 test_that("pkonfound creates the threshhold plot", {
+    thresh_plot <- pkonfound(2, .4, 100, 3, to_return = "thresh_plot")
+    
     expect_s3_class(thresh_plot, "ggplot")
 })
 
[…]
