+++ R/konfound-lm.R
@@ -1,6 +1,6 @@
 # konfound-lm
 
-konfound_lm <- function(model_object, tested_variable_string, test_all, alpha, tails, index, to_return) {
+konfound_lm <- function(model_object, tested_variable_string, alpha, tails, index, to_return) {
     tidy_output <- broom::tidy(model_object) # tidying output
     glance_output <- broom::glance(model_object)
     
