+++ tests/testthat/test-tkonfound.R
@@ -34,3 +34,9 @@ test_that("tkonfound creates the correlation plot when treatment is small", {
     
     expect_s3_class(tkonfound_fig5, "ggplot")
 })
+
+test_that("tkonfound for two_by_two works for raw output", {
+    output3 <- pkonfound(a = 35, b = 17, c = 17, d = 38, to_return = "raw_output")
+    
+    expect_equal(output3$RIR, tolerance = .001, 14)
+})
