### 2024-02-12 dc144dd
Update nonlinear_auxiliary.R
