### 2024-02-12 28e9a5a
Update nonlinear_auxiliary.R
