### 2024-02-12 62d8b99
Update nonlinear_auxiliary.R
