### 2024-02-13 34eb320
patch to konfound to address test_all = TRUE error
