+++ pkgdown.yml
@@ -1,0 +1,7 @@
+pandoc: 2.19.2
+pkgdown: 2.0.7
+pkgdown_sha: ~
+articles:
+  introduction-to-konfound: introduction-to-konfound.html
+last_built: 2024-02-13T01:48Z
+
