### 2024-02-13 164aa4e
address warnings in tests that were due to an unititialized column
