+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Type: Package
 Package: konfound
 Title: Quantify the Robustness of Causal Inferences
-Version: 0.4.0
+Version: 0.5.0
 Authors@R: c(
     person(c("Joshua", "M"), "Rosenberg", , "jmrosen48@gmail.com", role = c("aut", "cre")),
     person("Ran", "Xu", , "ranxu@msu.edu", role = "ctb"),
