+++ README.Rmd
@@ -15,6 +15,7 @@ knitr::opts_chunk$set(
 <!-- badges: start -->
 [![CRAN status](https://www.r-pkg.org/badges/version/konfound)](https://cran.r-project.org/package=konfound)
 [![R-CMD-check](https://github.com/konfound-project/konfound/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/konfound-project/konfound/actions/workflows/R-CMD-check.yaml)
+[![status](https://joss.theoj.org/papers/c53ffd4bff0f14c6533e45423d7deb82/status.svg)](https://joss.theoj.org/papers/c53ffd4bff0f14c6533e45423d7deb82)
 <!-- badges: end -->
 
 # konfound
