+++ NAMESPACE
@@ -1,7 +1,6 @@
 # Generated by roxygen2: do not edit by hand
 
 export(konfound)
-export(launch_shiny)
 export(mkonfound)
 export(pkonfound)
 export(summary.konfound)
