+++ tests/testthat/test-konfound.R
@@ -19,6 +19,10 @@ test_that("konfound works for glm, 2x2 model", {
     m4 <- glm(outcome ~ condition, data = binary_dummy_data, family = binomial(link = "logit"))
     output4 <- konfound(m4, condition, two_by_two = TRUE, n_treat = 55, to_return = "raw_output")
     expect_equal(output4$RIR, 15)
+    
+    m4 <- glm(outcome ~ condition, data = binary_dummy_data, family = binomial(link = "logit"))
+    output4_print <- capture.output(konfound(m4, condition, two_by_two = TRUE, n_treat = 55, to_return = "print"))
+    expect_true(length(output4_print) > 0)
 })
 
 
@@ -35,11 +39,11 @@ test_that("konfound returns a tibble", {
     expect_s3_class(output5b, "tbl_df")
 })
 
-gss_cat$married <- ifelse(gss_cat$marital == "Married", 1, 0)
-
-m6 <- glm(married ~ age, data = gss_cat, family = binomial(link = "logit"))
-m6_output <- konfound(m6, age, to_return = "raw_output")
-
 test_that("konfound glm works", {
+    gss_cat$married <- ifelse(gss_cat$marital == "Married", 1, 0)
+    
+    m6 <- glm(married ~ age, data = gss_cat, family = binomial(link = "logit"))
+    m6_output <- konfound(m6, age, to_return = "raw_output")
+    
     expect_equal(as.vector(m6_output$percent_bias_to_change_inference), 35.357)
 })
