+++ README.knit.md
@@ -1,9 +1,13 @@
+---
+output: github_document
+---
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
-<!-- badges: start -->
 
-[![CRAN
-status](https://www.r-pkg.org/badges/version/konfound)](https://cran.r-project.org/package=konfound)
+
+
+<!-- badges: start -->
+[![CRAN status](https://www.r-pkg.org/badges/version/konfound)](https://cran.r-project.org/package=konfound)
 [![R-CMD-check](https://github.com/konfound-project/konfound/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/konfound-project/konfound/actions/workflows/R-CMD-check.yaml)
 [![codecov](https://codecov.io/gh/konfound-project/konfound/graph/badge.svg?token=ARijYlxn7O)](https://codecov.io/gh/konfound-project/konfound)
 [![status](https://joss.theoj.org/papers/c53ffd4bff0f14c6533e45423d7deb82/status.svg)](https://joss.theoj.org/papers/c53ffd4bff0f14c6533e45423d7deb82)
@@ -11,27 +15,21 @@ status](https://www.r-pkg.org/badges/version/konfound)](https://cran.r-project.o
 
 # konfound
 
-In social science (and educational) research, we often wish to
-understand how robust inferences about effects are to unobserved (or
-controlled for) covariates, possible problems with measurement, and
-other sources of bias. The goal of `konfound` is to carry out
-sensitivity analysis to help analysts to *quantify how robust inferences
-are to potential sources of bias*. This R package provides tools to
-carry out sensitivity analysis as described in Frank, Maroulis, Duong,
-and Kelcey (2013) based on Rubin’s (1974) causal model as well as in
-Frank (2000) based on the impact threshold for a confounding variable.
+In social science (and educational) research, we often wish to understand how robust inferences about effects are to unobserved (or controlled for) covariates, possible problems with measurement, and other sources of bias.  The goal of `konfound` is to carry out sensitivity analysis to help analysts to *quantify how robust inferences are to potential sources of bias*. This R package provides tools to carry out sensitivity analysis as described in Frank, Maroulis, Duong, and Kelcey (2013) based on Rubin’s (1974) causal model as well as in Frank (2000) based on the impact threshold for a confounding variable.
 
 # Installation
 
 You can install the CRAN version of konfound with:
 
-``` r
+
+```r
 install.packages("konfound")
 ```
 
 You can install the development version from GitHub with:
 
-``` r
+
+```r
 install.packages("devtools")
 devtools::install_github("jrosen48/konfound")
 ```
@@ -40,18 +38,17 @@ devtools::install_github("jrosen48/konfound")
 
 ## pkonfound() for published studies
 
-`pkonfound()`, for published studies, calculates (1) how much bias there
-must be in an estimate to invalidate/sustain an inference; (2) the
-impact of an omitted variable necessary to invalidate/sustain an
-inference for a regression coefficient:
+`pkonfound()`, for published studies, calculates (1) how much bias there must be in an estimate to invalidate/sustain an inference; (2) the impact of an omitted variable necessary to invalidate/sustain an inference for a regression coefficient:
 
-``` r
+
+```r
 library(konfound)
 #> Sensitivity analysis as described in Frank, Maroulis, Duong, and Kelcey (2013) and in Frank (2000).
 #> For more information visit http://konfound-it.com.
 ```
 
-``` r
+
+```r
 pkonfound(est_eff = 2, 
           std_err = .4, 
           n_obs = 100, 
@@ -75,11 +72,10 @@ pkonfound(est_eff = 2,
 
 ## konfound() for models fit in R
 
-`konfound()` calculates the same for models fit in R. For example, here
[…]
