+++ README.knit.md
@@ -1,209 +1,0 @@
----
-output: github_document
----
-
-<!-- README.md is generated from README.Rmd. Please edit that file -->
-
-
-
-<!-- badges: start -->
-[![CRAN status](https://www.r-pkg.org/badges/version/konfound)](https://cran.r-project.org/package=konfound)
-[![R-CMD-check](https://github.com/konfound-project/konfound/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/konfound-project/konfound/actions/workflows/R-CMD-check.yaml)
-[![codecov](https://codecov.io/gh/konfound-project/konfound/graph/badge.svg?token=ARijYlxn7O)](https://codecov.io/gh/konfound-project/konfound)
-[![status](https://joss.theoj.org/papers/c53ffd4bff0f14c6533e45423d7deb82/status.svg)](https://joss.theoj.org/papers/c53ffd4bff0f14c6533e45423d7deb82)
-<!-- badges: end -->
-
-# konfound
-
-In social science (and educational) research, we often wish to understand how robust inferences about effects are to unobserved (or controlled for) covariates, possible problems with measurement, and other sources of bias.  The goal of `konfound` is to carry out sensitivity analysis to help analysts to *quantify how robust inferences are to potential sources of bias*. This R package provides tools to carry out sensitivity analysis as described in Frank, Maroulis, Duong, and Kelcey (2013) based on Rubin’s (1974) causal model as well as in Frank (2000) based on the impact threshold for a confounding variable.
-
-# Installation
-
-You can install the CRAN version of konfound with:
-
-
-```r
-install.packages("konfound")
-```
-
-You can install the development version from GitHub with:
-
-
-```r
-install.packages("devtools")
-devtools::install_github("jrosen48/konfound")
-```
-
-# Use of konfound
-
-## pkonfound() for published studies
-
-`pkonfound()`, for published studies, calculates (1) how much bias there must be in an estimate to invalidate/sustain an inference; (2) the impact of an omitted variable necessary to invalidate/sustain an inference for a regression coefficient:
-
-
-```r
-library(konfound)
-#> Sensitivity analysis as described in Frank, Maroulis, Duong, and Kelcey (2013) and in Frank (2000).
-#> For more information visit http://konfound-it.com.
-```
-
-
-```r
-pkonfound(est_eff = 2, 
-          std_err = .4, 
-          n_obs = 100, 
-          n_covariates = 3)
-#> Robustness of Inference to Replacement (RIR):
-#> To invalidate an inference,  60.29 % of the estimate would have to be due to bias. 
-#> This is based on a threshold of 0.794 for statistical significance (alpha = 0.05).
-#> 
-#> To invalidate an inference,  60  observations would have to be replaced with cases
-#> for which the effect is 0 (RIR = 60).
-#> 
-#> See Frank et al. (2013) for a description of the method.
-#> 
-#> Citation: Frank, K.A., Maroulis, S., Duong, M., and Kelcey, B. (2013).
-#> What would it take to change an inference?
-#> Using Rubin's causal model to interpret the robustness of causal inferences.
-#> Education, Evaluation and Policy Analysis, 35 437-460.
-#> For other forms of output, run ?pkonfound and inspect the to_return argument
-#> For models fit in R, consider use of konfound().
-```
-
-## konfound() for models fit in R
-
-`konfound()` calculates the same for models fit in R. For example, here are the coefficients for a linear model fit with `lm()` using the built-in dataset `mtcars`:
-
-
-```r
[…]
