### 2024-02-13 385496b
remove mention of functionality to call multiple forms of output at once
