+++ vignettes/introduction-to-konfound.Rmd
@@ -114,20 +114,6 @@ m1
 konfound(m1, hp)
 ```
 
-Like with `pkonfound()`, we can also output multiple forms of output at once with `konfound()`:
-
-```{r}
-konfound_output <- konfound(m1, hp, to_return = c("raw_output", "thresh_plot", "corr_plot"))
-summary(konfound_output)
-```
-
-Again, we can type each of those, i.e.:
-
-```{r}
-konfound_output$raw_output
-konfound_output$thresh_plot
-```
-
 Whereas this cannot be carried out with `pkonfound()`, with `konfound()` you can also return a table with some key output from the correlation-based approach.
 
 ```{r}
@@ -158,8 +144,6 @@ if (requireNamespace("forcats")) {
 m4 <- glm(outcome ~ condition, data = binary_dummy_data)
 konfound(m4, condition, two_by_two = TRUE)
 ```
-
-As with models fit with `lm()` (and use of `pkonfound()`), multiple forms of output can be specified with the `to_return` argument to `konfound()`, i.e. `konfound(m2, age, to_return = c("raw_output", "corr_plot", "thresh_plot"))`.
 
 **For mixed effects (or multi-level) models fit with the lmer() function from the lme4 package**
 
