+++ pkgdown.yml
@@ -3,5 +3,5 @@ pkgdown: 2.0.7
 pkgdown_sha: ~
 articles:
   introduction-to-konfound: introduction-to-konfound.html
-last_built: 2024-02-13T01:48Z
+last_built: 2024-02-13T02:59Z
 
