### 2024-02-13 konfound-project/konfound#53 @jrosen48
Hi both, as an update, we are now at 77%, and think we can get the coverage to above 90% with a little more digging.
