### 2024-02-13 konfound-project/konfound#53 @jrosen48
We've added the package to codecov, so the test coverage can now be examined here: https://app.codecov.io/gh/konfound-project/konfound/tree/master/R
