### 2024-02-13 458f912
remove calling of multiple forms of output
