### 2024-02-13 eb3f9e3
address NOTES in R CMD CHECK
