+++ NAMESPACE
@@ -3,7 +3,7 @@
 export(konfound)
 export(mkonfound)
 export(pkonfound)
-export(summary.konfound)
+export(summary)
 import(dplyr)
 import(ggplot2)
 import(lavaan)
