### 2024-02-13 84d3196
add a few more tests
