### 2024-02-13 edde33c
address note about class(flag_cov)
