### 2024-02-14 konfound-project/konfound#53 @jrosen48
As an update, here's where the test coverage stands - all functions > 70% coverage and 80% total coverage across the package:

```{r}
> covr::package_coverage()
konfound Coverage: 80.39%
R/nonlinear_auxiliary.R: 70.43%
R/test_sensitivity.R: 70.73%
R/test_sensitivity_ln.R: 73.81%
R/test_pse.R: 79.07%
R/tkonfound.R: 81.03%
R/pkonfound.R: 81.93%
R/tkonfound_fig.R: 83.02%
R/test_cop.R: 84.06%
R/core-sensitivity-mkonfound.R: 85.19%
R/helper_output_dataframe.R: 85.19%
R/cop_pse_auxiliary.R: 88.94%
R/konfound.R: 92.31%
R/helper_output_print.R: 95.60%
R/helper_plot_correlation.R: 95.83%
R/mkonfound.R: 95.83%
R/helper_plot_threshold.R: 97.73%
R/helper_output_table.R: 100.00%
R/konfound-glm-dichotomous.R: 100.00%
R/konfound-glm.R: 100.00%
R/konfound-lm.R: 100.00%
R/konfound-lmer.R: 100.00%
R/zzz.R: 100.00%
```

Noting that this is related to #50. We will use that issue primarily to discuss other code quality issues (cyclomatic complexity and non-compliant coding style).
