### 2024-02-14 7c6fe29
edits to JOSS paper in response to feedback from @fartist and @wjakethompson
