+++ R/test_sensitivity_ln.R
@@ -394,7 +394,8 @@ test_sensitivity_ln <- function(est_eff,
 		  final_switch1 = final_switch1, final_switch2 = final_switch2,
 		  final = final, dcroddsratio_ob = dcroddsratio_ob, final_extra = final_extra,
 		  a = a, b = b, c = c, d = d,
-		  # table_bstart1 = table_bstart1,
+		  table_bstart1 = table_bstart1,
+		  table_bstart2 = table_bstart2,
                   # RIRway_phrase = RIRway_phrase,
                   RIR_primary = RIR,
                   RIR_supplemental = RIR_extra, 
