### 2024-02-14 36300e9
simplify the function test_sensitivity_ln
now almost all calculations are moved to subfunctions in nonlinear_auxiliary
