### 2024-02-14 konfound-project/konfound#53 @jrosen48
We are at 81% and pausing due to asymptotic returns, in part due to the nature of some of the functions that calculate statistical indices and the error handling in their code. @wjakethompson and @fabian-s, can you offer any feedback on this? In your view, is this sufficient for ensuring the package functions as expected and that future bugs are likely to be detected? In our view, it is, as the coverage combined with the care we have taken to ensure that the statistical output is correct lend us confidence in the functioning of the package, though we recognize it could be possible to bolster further coverage (possibly to around 85-90% before severely diminishing returns set in!).

```{r}
> covr::package_coverage()
konfound Coverage: 80.49%
R/test_sensitivity.R: 70.73%
R/nonlinear_auxiliary.R: 70.84%
R/test_sensitivity_ln.R: 74.00%
R/test_pse.R: 79.07%
R/tkonfound.R: 81.03%
R/pkonfound.R: 81.93%
R/tkonfound_fig.R: 83.02%
R/test_cop.R: 84.06%
R/core-sensitivity-mkonfound.R: 85.19%
R/helper_output_dataframe.R: 85.19%
R/cop_pse_auxiliary.R: 88.94%
R/konfound.R: 92.31%
R/helper_output_print.R: 95.60%
R/helper_plot_correlation.R: 95.83%
R/mkonfound.R: 95.83%
R/helper_plot_threshold.R: 97.73%
R/helper_output_table.R: 100.00%
R/konfound-glm-dichotomous.R: 100.00%
R/konfound-glm.R: 100.00%
R/konfound-lm.R: 100.00%
R/konfound-lmer.R: 100.00%
R/zzz.R: 100.00%
```
