### 2024-02-14 92b1353
minor edits to DESCRIPTION, removing aes_string and data_frame to address warnings, and moving print output out of test_sensitivity_ln
