+++ R/test_sensitivity_ln.R
@@ -13,298 +13,323 @@ test_sensitivity_ln <- function(est_eff,
                                 to_return,
                                 model_object,
                                 tested_variable) {
-  
-  ## error message if input is inappropriate
-  if (!(std_err > 0)) {stop("Did not run! Standard error needs to be greater than zero.")}
-  if (!(n_obs > n_covariates + 3)) {stop("Did not run! There are too few observations relative to the number of observations and covariates. Please specify a less complex model to use KonFound-It.")}
-
-  thr_t <- cal_thr_t(est_eff, alpha, tails, n_obs, n_covariates)
-
-  # stop message
-  if (n_obs <= 0 || n_treat <= 0) {
-    stop("Please enter positive integers for sample size and number of treatment group cases.")
-  }
-  if (n_obs <= n_treat) {
-    stop("The total sample size should be larger than the number of treatment group cases.")
-  }
-  
-  odds_ratio <- exp(est_eff)
-  
-  # updated approach to deal with imaginary
-  minse <- cal_minse(n_obs, n_treat, odds_ratio)
-
-    # check if the implied table solution may contain imaginary numbers
-  changeSE <- FALSE
-  if (std_err < minse) {
-    haveimaginary <- TRUE
-    changeSE <- TRUE
-    user_std_err <- std_err
-    std_err <- minse
-  }
     
-  # n_treat is the number of observations in the treatment group (c+d)
-  # n_cnt is the number of observations in the control group (a+b)
-  n_cnt <- n_obs - n_treat
-  # t_ob is the t value calculated based on observed estimate and standard error
-  t_ob <- est_eff / std_err
-  # invalidate_ob is true - the observed result is significant - we are invalidating the observed result
-  invalidate_ob <- isinvalidate(thr_t, t_ob)
-  # dcroddsratio_ob is true - our goal is to decrease the odds ratio
-  dcroddsratio_ob <- isdcroddsratio(thr_t, t_ob)
-  
-  # previous approach to deal with imaginary 
-  # to record the original treatment cases in case we need to adjust it
-  # user_ntrm <- n_treat
-  # check if the implied table solution may contain imaginary numbers
-  # haveimaginary <- F
-  # changepi <- F
-  # set the default value for whether we need and can adjust pi (ratio of treatment cases)
-  # to remove the imaginary part
-  # keyimagin <- (4 + 4 * odds_ratio^2 + odds_ratio *
-  #                 (-8 + 4 * n_obs * std_err^2 - n_obs * n_treat * std_err^4 + n_treat^2 * std_err^4))
-  # minimgain <- 4 + 4 * odds_ratio^2 + odds_ratio * (-8 + n_obs * std_err^2 * (4 - 0.25 * n_obs * std_err^2))
-  # keyx1 <- 4 + 4 * odds_ratio^2 + odds_ratio * (-8 + 4 * n_obs * std_err^2)
-  # if (keyimagin > 0) {
+    ## error message if input is inappropriate
+    if (!(std_err > 0)) {stop("Did not run! Standard error needs to be greater than zero.")}
+    if (!(n_obs > n_covariates + 3)) {stop("Did not run! There are too few observations relative to the number of observations and covariates. Please specify a less complex model to use KonFound-It.")}
+    
+    thr_t <- cal_thr_t(est_eff, alpha, tails, n_obs, n_covariates)
+    
+    # stop message
+    if (n_obs <= 0 || n_treat <= 0) {
+        stop("Please enter positive integers for sample size and number of treatment group cases.")
+    }
+    if (n_obs <= n_treat) {
[…]
