+++ NAMESPACE
@@ -7,6 +7,7 @@ export(summary)
 export(tkonfound)
 import(ggplot2)
 import(lavaan)
+import(rlang)
 importFrom(broom,glance)
 importFrom(broom,tidy)
 importFrom(broom.mixed,tidy)
@@ -46,8 +47,10 @@ importFrom(ggplot2,theme_void)
 importFrom(ggplot2,xlab)
 importFrom(ggplot2,ylab)
 importFrom(ggrepel,geom_label_repel)
+importFrom(glue,glue)
 importFrom(lavaan,parameterEstimates)
 importFrom(lavaan,sem)
+importFrom(lifecycle,deprecated)
 importFrom(lme4,fixef)
 importFrom(lme4,lmer)
 importFrom(margins,margins)
