### 2024-02-15 konfound-project/konfound#50 @fabian-s
What actually happened here is you somewhat successfully refactored test_sensitivity_ln (217 -> 59),  while cyclocomp values for *every* other function remain completely unchanged since I opened this issue (e.g. git blame for getswitch_fisher tells me there have not been any structural changes at all in the last 4 years).

Most of this seems to be covered by tests now, so at least you may notice once your convoluted stuff breaks.

Close if you want, but IMO this is still rather terrible,  and I also do not appreciate at all being gaslit and misled like this.
