### 2024-02-15 konfound-project/konfound#53 @fabian-s
Seems ok, close if you want
