### 2024-02-15 konfound-project/konfound#50 @jrosen48
Also, I want to register that I was not attempting to mislead you! I wrote: "We have made changes to the code to reduce the cyclometric complexity, ...". I did not (and did not mean to!) suggest that we addressed the other issues or that this broader issue was resolved -- I was commenting on the cyclometric complexity aspect of these issues related to the code quality.
