### 2024-02-17 2192fc1
Update cop_pse_auxiliary.R
