### 2024-02-17 a51a414
Update cop_pse_auxiliary.R
