### 2024-02-17 ea2c3c5
Update cop_pse_auxiliary.R
