### 2024-02-17 5475749
Update helper_plot_correlation.R
