### 2024-02-18 a29c19e
Update helper_plot_threshold.R
