### 2024-02-18 ded8119
Update konfound-glm.R
