+++ R/konfound-lmer.R
@@ -31,7 +31,13 @@ get_kr_df <- function(model_object) {
 #' @return The results of the konfound analysis for the specified fixed effect(s).
 #' @importFrom broom.mixed tidy
 #' @importFrom dplyr filter bind_cols
-konfound_lmer <- function(model_object, tested_variable_string, test_all, alpha, tails, index, to_return) {
+konfound_lmer <- function(model_object, 
+                          tested_variable_string, 
+                          test_all, 
+                          alpha, 
+                          tails, 
+                          index, 
+                          to_return) {
     tidy_output <- broom.mixed::tidy(model_object) # tidying output
     
     coef_df <- tidy_output[tidy_output$term == tested_variable_string, ]
