### 2024-02-18 022adc7
Update konfound.R
