+++ R/test_cop.R
@@ -38,15 +38,25 @@ test_cop <- function(est_eff, # unstandardized
   var_z <- sdz <- 1
   
   ## error message if input is inappropriate
-  if (!(std_err > 0)) {stop("Did not run! Standard error needs to be greater than zero.")}
-  if (!(sdx > 0)) {stop("Did not run! Standard deviation of x needs to be greater than zero.")}
-  if (!(sdy > 0)) {stop("Did not run! Standard deviation of y needs to be greater than zero.")}
-  if (!(n_obs > n_covariates + 3)) {stop("Did not run! There are too few observations relative to the number of observations and covariates. Please specify a less complex model to use KonFound-It.")}
+  if (!(std_err > 0)) {stop("Did not run! Standard error needs to be 
+                            greater than zero.")}
+  if (!(sdx > 0)) {stop("Did not run! Standard deviation of x needs to be 
+                        greater than zero.")}
+  if (!(sdy > 0)) {stop("Did not run! Standard deviation of y needs to be 
+                        greater than zero.")}
+  if (!(n_obs > n_covariates + 3)) {
+    stop("Did not run! There are too few observations relative to the 
+         number of observations and covariates. Please specify a less 
+         complex model to use KonFound-It.")}
   if (!(R2 < FR2max)) {stop("Did not run! R2 Max needs to be greater than R2.")}
   if (!(FR2max < 1)) {stop("Did not run! R2 Max needs to be less than 1.")}
-  if (!(1-((sdy^2/sdx^2)*(1-R2)/((df+1) * std_err^2))>0)) {stop("Did not run! Entered values produced Rxz^2 <=0, consider adding more significant digits to your entered values.")}
+  if (!(1-((sdy^2/sdx^2)*(1-R2)/((df+1) * std_err^2))>0)) {
+    stop("Did not run! Entered values produced Rxz^2 <=0, 
+         consider adding more significant digits to your entered values.")}
 
-  negest <- 0 # an indicator of whether the use specified est_eff is negative, 1 is yes negative
+  # an indicator of whether the use specified est_eff is negative,
+  # 1 is yes negative
+  negest <- 0 
   if (est_eff < 0) {
     est_eff <- abs(est_eff)
     negest <- 1
@@ -73,7 +83,8 @@ test_cop <- function(est_eff, # unstandardized
   ## df + 1 because omitted variable is NOT included in M2 
   #### we change the n_obs to df to recover the rxz as in the particular sample
 
-  ## note that in the updated approach rxy is not necessary to calculate rxcv_exact, ryxcv_exact and delta_exact
+  ## note that in the updated approach rxy is not necessary to calculate 
+  ## rxcv_exact, ryxcv_exact and delta_exact
   rxy <- ryx <- cal_rxy(ryxGz, rxz, ryz)
   rxy_M2 <- cal_rxy(ryxGz_M2, rxz, ryz) 
   ## rxy_M2 is only for simulation to recover the exact number
@@ -86,8 +97,9 @@ test_cop <- function(est_eff, # unstandardized
   R2_uncond <- rxy^2
   
   ## calculate delta_star
-  delta_star <- cal_delta_star(FR2max, R2, R2_uncond, est_eff, eff_thr, var_x, var_y, eff_uncond, rxz, n_obs)
-  
+  delta_star <- cal_delta_star(FR2max, R2, R2_uncond, est_eff, 
+                               eff_thr, var_x, var_y, eff_uncond, 
+                               rxz, n_obs)
   ## now introduce cv
   sdcv <- var_cv <- 1
   rcvz <- rzcv <- 0
