### 2024-02-18 8ae3fef
Update test_cop.R
