+++ R/test_cop.R
@@ -120,9 +120,11 @@ test_cop <- function(est_eff, # unstandardized
   
   # prepare some other values in the final Table (long output)
   R2_M3_oster <- as.numeric(verify_oster[1])
-  eff_x_M3_oster <- as.numeric(verify_oster[2]) # should be equivalent or very close to eff_thr
+  eff_x_M3_oster <- as.numeric(verify_oster[2]) 
+  # should be equivalent or very close to eff_thr
   se_x_M3_oster <- as.numeric(verify_oster[3])
-  beta_x_M3_oster <- as.numeric(verify_oster[9]) # should be equivalent or very close to beta_thr
+  beta_x_M3_oster <- as.numeric(verify_oster[9]) 
+  # should be equivalent or very close to beta_thr
   t_x_M3_oster <- eff_x_M3_oster / se_x_M3_oster 
   eff_z_M3_oster <- as.numeric(verify_oster[4])
   se_z_M3_oster <- as.numeric(verify_oster[5])
@@ -131,7 +133,8 @@ test_cop <- function(est_eff, # unstandardized
   cov_oster <- verify_oster[[11]]
   cor_oster <- verify_oster[[12]]
   
-  ## calculate the exact/true rcvx, rcvy, delta (updated version that does not need rxy)
+  ## calculate the exact/true rcvx, rcvy, delta 
+  ## (updated version that does not need rxy)
   ### the idea is to calculate everything conditional on z
   sdxGz <- sdx * sqrt(1 - rxz^2)
   sdyGz <- sdy * sqrt(1 - ryz^2)
@@ -168,9 +171,11 @@ test_cop <- function(est_eff, # unstandardized
   
   # prepare some other values in the final Table (long output)
   R2_M3 <- as.numeric(verify_exact[1])
-  eff_x_M3 <- as.numeric(verify_exact[2]) # should be equivalent or very close to eff_thr
+  eff_x_M3 <- as.numeric(verify_exact[2]) 
+  # should be equivalent or very close to eff_thr
   se_x_M3 <- as.numeric(verify_exact[3])
-  beta_x_M3 <- as.numeric(verify_exact[9]) # should be equivalent or very close to beta_thr
+  beta_x_M3 <- as.numeric(verify_exact[9]) 
+  # should be equivalent or very close to beta_thr
   t_x_M3 <- eff_x_M3 / se_x_M3 
   eff_z_M3 <- as.numeric(verify_exact[4])
   se_z_M3 <- as.numeric(verify_exact[5])
@@ -181,8 +186,10 @@ test_cop <- function(est_eff, # unstandardized
   
   verify_pse_reg_M2 <- verify_reg_Gz(n_obs, sdx, sdy, sdz, rxy_M2, rxz, rzy)
   R2_M2 <- as.numeric(verify_pse_reg_M2[1])
-  eff_x_M2 <- as.numeric(verify_pse_reg_M2[2]) # should be equivalent or very close to est_eff
-  se_x_M2 <- as.numeric(verify_pse_reg_M2[3]) # should be equivalent or very close to std_err
+  eff_x_M2 <- as.numeric(verify_pse_reg_M2[2]) 
+  # should be equivalent or very close to est_eff
+  se_x_M2 <- as.numeric(verify_pse_reg_M2[3]) 
+  # should be equivalent or very close to std_err
   eff_z_M2 <- as.numeric(verify_pse_reg_M2[4]) 
   se_z_M2 <- as.numeric(verify_pse_reg_M2[5]) 
   t_x_M2 <- eff_x_M2 / se_x_M2 
@@ -222,7 +229,8 @@ test_cop <- function(est_eff, # unstandardized
                              "Intermediate(M2)", eff_x_M2, R2, "exact", 
[…]
