+++ R/test_sensitivity_ln.R
@@ -1,4 +1,5 @@
-# Main function to test sensitivity for non-linear models to be wrapped with pkonfound(), konfound(), and mkonfound()
+# Main function to test sensitivity for non-linear models to be wrapped 
+# with pkonfound(), konfound(), and mkonfound()
 
 test_sensitivity_ln <- function(est_eff,
                                 std_err,
@@ -15,17 +16,23 @@ test_sensitivity_ln <- function(est_eff,
                                 tested_variable) {
     
     ## error message if input is inappropriate
-    if (!(std_err > 0)) {stop("Did not run! Standard error needs to be greater than zero.")}
-    if (!(n_obs > n_covariates + 3)) {stop("Did not run! There are too few observations relative to the number of observations and covariates. Please specify a less complex model to use KonFound-It.")}
+    if (!(std_err > 0)) {
+      stop("Did not run! Standard error needs to be greater than zero.")}
+    if (!(n_obs > n_covariates + 3)) {
+      stop("Did not run! There are too few observations relative to 
+           the number of observations and covariates. Please specify a 
+           less complex model to use KonFound-It.")}
     
     thr_t <- cal_thr_t(est_eff, alpha, tails, n_obs, n_covariates)
     
     # stop message
     if (n_obs <= 0 || n_treat <= 0) {
-        stop("Please enter positive integers for sample size and number of treatment group cases.")
+        stop("Please enter positive integers for sample size 
+             and number of treatment group cases.")
     }
     if (n_obs <= n_treat) {
-        stop("The total sample size should be larger than the number of treatment group cases.")
+        stop("The total sample size should be larger than 
+             the number of treatment group cases.")
     }
     
     odds_ratio <- exp(est_eff)
@@ -197,9 +204,11 @@ test_sensitivity_ln <- function(est_eff,
             change, sprintf("you would need to replace %d", RIR), RIRway, "cases")
         
         if (replace == "control") {
-            conclusion1a <- sprintf("with cases for which the probability of failure in the control group applies (RIR = %d).", RIR)
+            conclusion1a <- sprintf("with cases for which the probability of 
+                                    failure in the control group applies (RIR = %d).", RIR)
         } else {
-            conclusion1a <- sprintf("with cases for which the probability of failure in the entire sample applies (RIR = %d).", RIR)
+            conclusion1a <- sprintf("with cases for which the probability of failure 
+                                    in the entire sample applies (RIR = %d).", RIR)
         }
         
         conclusion1b <- paste(
@@ -213,9 +222,11 @@ test_sensitivity_ln <- function(est_eff,
             change, sprintf("you would need to replace %d", RIR), RIRway, "cases")
         
         if (replace == "control") {
-            conclusion1a <- sprintf("with cases for which the probability of failure in the control group applies (RIR = %d).", RIR)
+            conclusion1a <- sprintf("with cases for which the probability of 
+                                    failure in the control group applies (RIR = %d).", RIR)
         } else {
-            conclusion1a <- sprintf("with cases for which the probability of failure in the entire sample applies (RIR = %d).", RIR)
+            conclusion1a <- sprintf("with cases for which the probability of failure 
+                                    in the entire sample applies (RIR = %d).", RIR)
         }
         
         conclusion1b <- paste(
@@ -228,22 +239,27 @@ test_sensitivity_ln <- function(est_eff,
         conclusion1 <- paste(
             change, c("only transferring cases from"), transferway, "is not enough.")
         
-        conclusion1b <- paste(sprintf("We also need to transfer %d cases from", final_solution$final_extra),
-                              transferway_extra, c("as shown, from the User-entered Table to the Transfer Table. "))
[…]
