### 2024-02-18 853ae6e
Update helper_output_table.R
