### 2024-02-18 5064575
Update nonlinear_auxiliary.R
