### 2024-02-18 2af153f
Update test_cop.R
