### 2024-02-19 6152278
Update helper_output_print.R
