+++ R/helper_output_print.R
@@ -6,21 +6,15 @@ output_print <- function(est_eff, beta_threshhold, bias = NULL, sustain = NULL, 
     if ((abs(est_eff) > abs(beta_threshhold)) & is.na(eff_thr) == TRUE) {
       cat("TO INVALIDATE:\n")
       cat("\n")
-      cat(paste0("RIR = ", round(recase, 3), "\n"))
+      cat(paste0("RIR = ", recase, "\n"))
       cat("\n")
-      cat(paste0("You entered an estimated effect of ", round(est_eff, 3), ". To invalidate"))
-      cat("\n")
-      cat(paste0("the inference of an effect using the threshold of ", round(beta_threshhold, 3), " for"))
-      cat("\n")
-      cat(paste0("statistical significance with alpha = ", alpha, ", ", round(bias, 3), "% of the"))
-      cat("\n")
-      cat(paste0("(", round(est_eff, 3), ") estimate would have to be due to bias. This implies"))
-      cat("\n")
+      cat(sprintf("You entered an estimated effect of %.3f. To invalidate\n", est_eff))
+      cat(sprintf("the inference of an effect using the threshold of %.3f for\n", beta_threshhold))
+      cat(sprintf("statistical significance with alpha = %s, %.3f%% of the\n", alpha, bias))
+      cat(sprintf("(%.3f) estimate would have to be due to bias. This implies\n", est_eff))
       cat("that to invalidate the inference one would expect to have to\n")
-      cat(paste0("replace ", round(recase, 3), " (", round(bias, 3), "%) observations with cases for which the"))
-      cat("\n")    
-      cat(paste0("treatment effect is ", nu, " (RIR = ", round(recase, 3), ").\n"))
-      cat("\n")    
+      cat(sprintf("replace %.3f (%.3f%%) observations with cases for which the\n", recase, bias))
+      cat(sprintf("treatment effect is %s (RIR = %.3f).\n\n", nu, recase))
     } else if ((abs(est_eff) > abs(beta_threshhold)) & is.na(eff_thr) == FALSE) {
       cat("TO INVALIDATE:\n")
       cat("\n")
