### 2024-02-19 4af81ea
Update helper_output_print.R
