### 2024-02-19 d423aac
Update helper_output_print.R
