+++ R/helper_output_print.R
@@ -18,53 +18,34 @@ output_print <- function(est_eff, beta_threshhold, bias = NULL, sustain = NULL, 
     } else if ((abs(est_eff) > abs(beta_threshhold)) & is.na(eff_thr) == FALSE) {
       cat("TO INVALIDATE:\n")
       cat("\n")
-      cat(paste0("RIR = ", round(recase, 3), "\n"))
-      cat("\n")
-      cat(paste0("You entered an effect of ", round(est_eff, 3), ", and specified a threshold"))
-      cat("\n")
-      cat(paste0("for inference of ", round(eff_thr, 3), ". To invalidate the inference based on your"))
-      cat("\n")
-      cat(paste0("estimate, ", round(bias, 3), "% of the (", round(est_eff, 3), ") estimate would have to be due to"))
-      cat("\n")
-      cat(paste0("bias. This implies that to invalidate the inference one would"))
-      cat("\n")
-      cat(paste0("expect to have to replace ", round(recase, 3), " (", round(bias, 3), "%) observations with"))
-      cat("\n")
-      cat(paste0("cases for which the treatment effect is ", nu, " (RIR = ", round(recase, 3), ").\n"))
-      cat("\n") 
+      cat(paste0("RIR = ", recase, "\n"))
+      cat(sprintf("You entered an effect of %.3f, and specified a threshold\n", est_eff))
+      cat(sprintf("for inference of %.3f. To invalidate the inference based on your\n", eff_thr))
+      cat(sprintf("estimate, %.3f%% of the (%.3f) estimate would have to be due to\n", bias, est_eff))
+      cat("bias. This implies that to invalidate the inference one would\n")
+      cat(sprintf("expect to have to replace %.0f (%.3f%%) observations with\n", recase, bias))
+      cat(sprintf("cases for which the treatment effect is %s (RIR = %.0f).\n\n", nu, recase))
     } else if ((abs(est_eff) < abs(beta_threshhold)) & is.na(eff_thr) == TRUE) {
       cat("TO SUSTAIN:\n", sep = "")
       cat("\n")
-      cat(paste0("RIR = ", round(recase, 3), "\n"))
-      cat("\n")
-      cat(paste0("You entered an estimated effect of ", round(est_eff, 3), ". The threshold value for"))
-      cat("\n")
-      cat(paste0("statistical significance is ", round(beta_threshhold, 3), " (alpha = ", alpha, "). To reach that"))
-      cat("\n")
-      cat(paste0("threshold, ", round(sustain, 3), "% of the (", round(est_eff, 3), ") estimate would have to be due to"))
-      cat("\n")
-      cat("bias. This implies that to sustain an inference one would expect to")
-      cat("\n")
-      cat(paste0("have to replace ", round(recase, 3), " (", round(sustain, 3), "%) observations with effect of ", nu, " with\n"))
-      cat(paste0("cases with effect of ", round(beta_threshhold, 3), " (RIR = ", round(recase, 3), ").\n"))
-      cat("\n")
+      cat(paste0("RIR = ", recase, "\n"))
+      cat(sprintf("You entered an estimated effect of %.3f. The threshold value for\n", est_eff))
+      cat(sprintf("statistical significance is %.3f (alpha = %s). To reach that\n", beta_threshhold, alpha))
+      cat(sprintf("threshold, %.3f%% of the (%.3f) estimate would have to be due to\n", sustain, est_eff))
+      cat("bias. This implies that to sustain an inference one would expect to\n")
+      cat(sprintf("have to replace %.0f (%.3f%%) observations with effect of %s with\n", recase, sustain, nu))
+      cat(sprintf("cases with effect of %.3f (RIR = %.0f).\n\n", beta_threshhold, recase))
[…]
