### 2024-02-19 c399019
Update helper_output_print.R
