### 2024-02-19 a54dc65
Update helper_output_print.R
