+++ R/helper_output_print.R
@@ -78,39 +78,39 @@ if (requireNamespace("htmltools", quietly = TRUE)) {
     cat(crayon::bold("Impact Threshold for a Confounding Variable:\n"))
     cat("\n")
     if (abs(obs_r) > abs(critical_r) & obs_r > 0) {
-      cat(paste0("The minimum impact of an omitted variable to invalidate an inference\n")
-      cat(paste0("for a null hypothesis of 0 effect is based on a correlation of ", sprintf("%.3f", r_con), "\n"))
-      cat(paste0("with the outcome and at ", sprintf("%.3f", r_con), " with the predictor of interest (conditioning\n"))
-      cat(paste0("on all observed covariates in the model) based on a threshold of ", sprintf("%.3f", critical_r), "\n"))
-      cat(paste0("for statistical significance (alpha = ", alpha, ").\n\n"))
-      cat(paste0("Correspondingly the impact of an omitted variable (as defined in Frank 2000) must be\n"))
-      cat(paste0(sprintf("%.3f X %.3f = %.3f", r_con, r_con, r_con^2), " to invalidate an inference for a null hypothesis of 0 effect.\n\n"))
-    } else if (abs(obs_r) > abs(critical_r) & obs_r < 0) {
-      cat(paste0("The minimum (in absolute value) impact of an omitted variable to\n"))
-      cat(paste0("invalidate an inference for a null hypothesis of 0 effect is based on\n"))
-      cat(paste0("a correlation of ", sprintf("%.3f", -r_con), " with the outcome and at ", sprintf("%.3f", r_con), " with the predictor\n"))
-      cat(paste0("of interest (conditioning on all observed covariates in the model;\n"))
-      cat(paste0("signs are interchangeable) based on a threshold of ", sprintf("%.3f", beta_threshhold), " for\n"))
-      cat(paste0("statistical significance (alpha = ", alpha, ").\n\n"))
-      cat(paste0("Correspondingly the impact of an omitted variable (as defined in Frank 2000) must be\n"))
-      cat(paste0(sprintf("%.3f X %.3f = %.3f", -r_con, r_con, -r_con^2), " to invalidate an inference for a null hypothesis of 0 effect.\n\n"))
-    } else if (abs(obs_r) < abs(critical_r) & obs_r > 0) {
-      cat(paste0("The maximum impact (in absolute value) of an omitted variable to\n"))
-      cat(paste0("sustain an inference for a null hypothesis of 0 effect is based on\n"))
-      cat(paste0("a correlation of ", sprintf("%.3f", -r_con), " with the outcome an at ", sprintf("%.3f", r_con), " with the predictor\n"))
-      cat(paste0("of interest (conditioning on all observed covariates in the model;\n"))
-      cat(paste0("signs are interchangeable) based on a threshold of ", sprintf("%.3f", beta_threshhold), " for\n"))
-      cat(paste0("statistical significance (alpha = ", alpha, ").\n\n"))
-      cat(paste0("Correspondingly the impact of an omitted variable (as defined in Frank 2000) must be\n"))
-      cat(paste0(sprintf("%.3f X %.3f = %.3f", -r_con, r_con, -r_con^2), " to sustain an inference for a null hypothesis of 0 effect.\n\n"))
-    } else if (abs(obs_r) < abs(critical_r) & obs_r < 0) {
-      cat(paste0("The maximum impact of an omitted variable to sustain an inference\n"))
-      cat(paste0("for a null hypothesis of 0 effect is based on a correlation of ", sprintf("%.3f", r_con), "\n"))
-      cat(paste0("with the outcome and at ", sprintf("%.3f", r_con), " with the predictor of interest (conditioning\n"))
[…]
