+++ R/helper_output_print.R
@@ -78,39 +78,66 @@ if (requireNamespace("htmltools", quietly = TRUE)) {
     cat(crayon::bold("Impact Threshold for a Confounding Variable:\n"))
     cat("\n")
     if (abs(obs_r) > abs(critical_r) & obs_r > 0) {
-      cat("The minimum impact of an omitted variable to invalidate an inference\n")
-      cat("for a null hypothesis of 0 effect is based on a correlation of ", sprintf("%.3f", r_con), "\n")
-      cat("with the outcome and at ", sprintf("%.3f", r_con), " with the predictor of interest (conditioning\n")
-      cat("on all observed covariates in the model) based on a threshold of ", sprintf("%.3f", critical_r), "\n")
-      cat("for statistical significance (alpha = ", alpha, ").\n\n")
-      cat("Correspondingly the impact of an omitted variable (as defined in Frank 2000) must be\n")
-      cat(sprintf("%.3f X %.3f = %.3f", r_con, r_con, r_con^2), " to invalidate an inference for a null hypothesis of 0 effect.\n\n")
-  } else if (abs(obs_r) > abs(critical_r) & obs_r < 0) {
-      cat("The minimum (in absolute value) impact of an omitted variable to\n")
-      cat("invalidate an inference for a null hypothesis of 0 effect is based on\n")
-      cat("a correlation of ", sprintf("%.3f", -r_con), " with the outcome and at ", sprintf("%.3f", r_con), " with the predictor\n")
-      cat("of interest (conditioning on all observed covariates in the model;\n")
-      cat("signs are interchangeable) based on a threshold of ", sprintf("%.3f", beta_threshhold), " for\n")
-      cat("statistical significance (alpha = ", alpha, ").\n\n")
-      cat("Correspondingly the impact of an omitted variable (as defined in Frank 2000) must be\n")
-      cat(sprintf("%.3f X %.3f = %.3f", -r_con, r_con, -r_con^2), " to invalidate an inference for a null hypothesis of 0 effect.\n\n")
-  } else if (abs(obs_r) < abs(critical_r) & obs_r > 0) {
-      cat("The maximum impact (in absolute value) of an omitted variable to\n")
-      cat("sustain an inference for a null hypothesis of 0 effect is based on\n")
-      cat("a correlation of ", sprintf("%.3f", -r_con), " with the outcome an at ", sprintf("%.3f", r_con), " with the predictor\n")
-      cat("of interest (conditioning on all observed covariates in the model;\n")
-      cat("signs are interchangeable) based on a threshold of ", sprintf("%.3f", beta_threshhold), " for\n")
-      cat("statistical significance (alpha = ", alpha, ").\n\n")
-      cat("Correspondingly the impact of an omitted variable (as defined in Frank 2000) must be\n")
-      cat(sprintf("%.3f X %.3f = %.3f", -r_con, r_con, -r_con^2), " to sustain an inference for a null hypothesis of 0 effect.\n\n")
-  } else if (abs(obs_r) < abs(critical_r) & obs_r < 0) {
-      cat("The maximum impact of an omitted variable to sustain an inference\n")
-      cat("for a null hypothesis of 0 effect is based on a correlation of ", sprintf("%.3f", r_con), "\n")
-      cat("with the outcome and at ", sprintf("%.3f", r_con), " with the predictor of interest (conditioning\n")
-      cat("on all observed covariates in the model) based on a threshold of ", sprintf("%.3f", beta_threshhold), "\n")
-      cat("for statistical significance (alpha = ", alpha, ").\n\n")
-      cat("Correspondingly the impact of an omitted variable (as defined in Frank 2000) must be\n")
[…]
