### 2024-02-19 e761827
Update test_cop.R
