### 2024-02-19 594332f
Update helper_output_print.R
round up 'r_con' for ITCV
