+++ R/helper_output_print.R
@@ -105,9 +105,9 @@ if (requireNamespace("htmltools", quietly = TRUE)) {
     if (abs(obs_r) > abs(critical_r) & obs_r > 0) {
       cat("The minimum impact of an omitted variable to invalidate an inference")
       cat("\n")
-      cat(paste0("for a null hypothesis of 0 effect is based on a correlation of ", r_con))
+      cat(paste0("for a null hypothesis of 0 effect is based on a correlation of ", round(r_con, 3)))
       cat("\n")
-      cat(paste0("with the outcome and at ", r_con, " with the predictor of interest (conditioning"))
+      cat(paste0("with the outcome and at ", round(r_con, 3), " with the predictor of interest (conditioning"))
       cat("\n")
       cat(paste0("on all observed covariates in the model) based on a threshold of ", round(critical_r, 3)))
       cat("\n")
@@ -116,13 +116,13 @@ if (requireNamespace("htmltools", quietly = TRUE)) {
       cat("\n")
       cat("Correspondingly the impact of an omitted variable (as defined in Frank 2000) must be ") 
       cat("\n")
-      cat(r_con, " X ", r_con, " = ", round(r_con^2, 3), " to invalidate an inference for a null hypothesis of 0 effect.\n", sep = "")
+      cat(round(r_con, 3), " X ", round(r_con, 3), " = ", round(r_con^2, 3), " to invalidate an inference for a null hypothesis of 0 effect.\n", sep = "")
     } else if (abs(obs_r) > abs(critical_r) & obs_r < 0) {
       cat("The minimum (in absolute value) impact of an omitted variable to")
       cat("\n")
       cat(paste0("invalidate an inference for a null hypothesis of 0 effect is based on"))
       cat("\n")
-      cat(paste0("a correlation of ", -r_con, " with the outcome and at ", r_con, " with the predictor"))
+      cat(paste0("a correlation of ", -round(r_con, 3), " with the outcome and at ", round(r_con, 3), " with the predictor"))
       cat("\n")
       cat("of interest (conditioning on all observed covariates in the model;")
       cat("\n")
@@ -132,13 +132,13 @@ if (requireNamespace("htmltools", quietly = TRUE)) {
       cat("\n")
       cat("Correspondingly the impact of an omitted variable (as defined in Frank 2000) must be ") 
       cat("\n")
-      cat(-r_con, " X ", r_con, " = ", -round(r_con^2, 3), " to invalidate an inference for a null hypothesis of 0 effect.\n", sep = "")
+      cat(-round(r_con, 3), " X ", round(r_con, 3), " = ", -round(r_con^2, 3), " to invalidate an inference for a null hypothesis of 0 effect.\n", sep = "")
     } else if (abs(obs_r) < abs(critical_r) & obs_r > 0) {
       cat("The maximum impact (in absolute value) of an omitted variable to") 
       cat("\n")
       cat("sustain an inference for a null hypothesis of 0 effect is based on")
       cat("\n")
-      cat(paste0("a correlation of ", -r_con, " with the outcome an at ", r_con, " with the predictor"))
+      cat(paste0("a correlation of ", -round(r_con, 3), " with the outcome an at ", round(r_con, 3), " with the predictor"))
       cat("\n")
       cat("of interest (conditioning on all observed covariates in the model;")
       cat("\n")
@@ -148,13 +148,13 @@ if (requireNamespace("htmltools", quietly = TRUE)) {
       cat("\n")
       cat("Correspondingly the impact of an omitted variable (as defined in Frank 2000) must be ")
       cat("\n")
[…]
