### 2024-02-19 fda8af9
Update helper_output_print.R
