### 2024-02-19 b6d5d13
Update cop_pse_auxiliary.R
