+++ R/cop_pse_auxiliary.R
@@ -5,9 +5,9 @@
 #' @return R2yz value
 #' @importFrom lavaan parameterEstimates
 cal_ryz <- function(ryxGz, R2){
-    R2yz = (ryxGz^2 - R2)/(ryxGz^2 - 1)
+    R2yz <- (ryxGz^2 - R2)/(ryxGz^2 - 1)
     if (R2yz >= 0) {
-        ryz = sqrt(R2yz)
+        ryz <- sqrt(R2yz)
     } else {
         stop("Error! R2yz < 0!")
     }
@@ -24,10 +24,10 @@ cal_ryz <- function(ryxGz, R2){
 #' @return R2xz value
 #' @importFrom lavaan parameterEstimates
 cal_rxz <- function(var_x, var_y, R2, df, std_err){
-    R2xz = 1 - ((var_y * (1 - R2))/(var_x * df * std_err^2))
+    R2xz <- 1 - ((var_y * (1 - R2))/(var_x * df * std_err^2))
     if (R2xz <= 0) {stop("Error! R2xz < 0!")} 
     ## Note output the number of R2xz
-    rxz = sqrt(R2xz)
+    rxz <- sqrt(R2xz)
     return(rxz)
 }
 
@@ -39,7 +39,7 @@ cal_rxz <- function(var_x, var_y, R2, df, std_err){
 #' @return rxy value
 #' @importFrom lavaan parameterEstimates
 cal_rxy <- function(ryxGz, rxz, ryz){
-    rxy = ryxGz * sqrt((1 - rxz^2)*(1 - ryz^2)) + rxz * ryz
+    rxy <- ryxGz * sqrt((1 - rxz^2)*(1 - ryz^2)) + rxz * ryz
     return(rxy)
 }
 
@@ -66,33 +66,33 @@ cal_delta_star <- function(FR2max,
                            est_uncond, 
                            rxz, 
                            n_obs) {
-    if (FR2max > .99) {FR2max = .99}
+    if (FR2max > .99) {FR2max <- .99}
     # if (FR2max < R2 + inci) {FR2max = R2 + inci} check with Ken what this means
-    if (FR2max > R2) {D = sqrt(FR2max - R2)}
+    if (FR2max > R2) {D <- sqrt(FR2max - R2)}
     
     #elements for computing Oster's delta_star
-    bt_m_b = est_eff - eff_thr
-    rt_m_ro_t_syy = (R2 - R2_uncond) * var_y
-    b0_m_b1 = est_uncond - est_eff
-    rm_m_rt_t_syy = (FR2max - R2) * var_y
+    bt_m_b <- est_eff - eff_thr
+    rt_m_ro_t_syy <- (R2 - R2_uncond) * var_y
+    b0_m_b1 <- est_uncond - est_eff
+    rm_m_rt_t_syy <- (FR2max - R2) * var_y
     
-    t_x = var_x * (n_obs / (n_obs - 1)) * (1 - rxz^2)
+    t_x <- var_x * (n_obs / (n_obs - 1)) * (1 - rxz^2)
     ## adjust df for var_x 
     ## var_x is population variance, need sample variance from x
     ## this adjustment is to get closer to what robomit generates as they 
     ## run regression using the sample data 
-    num1 = bt_m_b * rt_m_ro_t_syy * t_x
-    num2 = bt_m_b * var_x * t_x * b0_m_b1^2
-    num3 = 2 * bt_m_b^2 * (t_x * b0_m_b1 * var_x)
-    num4 = bt_m_b^3 * (t_x * var_x - t_x^2)
-    num = num1 + num2 + num3 + num4
[…]
