+++ R/concord1.R
@@ -5,6 +5,7 @@
 #' @docType data
 #' @name concord1
 #' @references Hamilton, Lawrence C. 1983. Saving water: 
-#' A causal model of household conservation. Sociological Perspectives 26(4):355-374.
+#' A causal model of household conservation. 
+#' Sociological Perspectives 26(4):355-374.
 #' @format A data.frame with 496 rows and 10 variables.
 NULL
