### 2024-02-19 5e4a9d7
Update cop_pse_auxiliary.R
