### 2024-02-19 fc0c914
Update konfound-glm.R
