+++ R/konfound-lmer.R
@@ -10,7 +10,8 @@
 get_kr_df <- function(model_object) {
     L <- diag(rep(1, length(lme4::fixef(model_object))))
     L <- as.data.frame(L)
-    out <- suppressWarnings(purrr::map_dbl(L, pbkrtest::get_Lb_ddf, object = model_object))
+    out <- suppressWarnings(purrr::map_dbl(L, pbkrtest::get_Lb_ddf, 
+                                           object = model_object))
     names(out) <- names(lme4::fixef(model_object))
     out
 }
