### 2024-02-19 519609c
Update mkonfound.R
