+++ R/mkonfound.R
@@ -51,7 +51,8 @@ mkonfound <- function(d, t, df, alpha = .05, tails = 2, return_plot = FALSE) {
     )
 
     p <- ggplot2::ggplot(results_df, 
-                         ggplot2::aes(x = results_df$pct_bias_to_change_inference, 
+                         ggplot2::aes(
+                           x = results_df$pct_bias_to_change_inference, 
                                       fill = results_df$action)) +
       ggplot2::geom_histogram() +
       ggplot2::scale_fill_manual("", values = c("#1F78B4", "#A6CEE3")) +
