### 2024-02-19 bcc60d2
Update nonlinear_auxiliary.R
