+++ R/zzz.R
@@ -12,5 +12,7 @@ if (getRversion() >= "2.15.1") utils::globalVariables(c("inference", "key",
                                                         "percent_bias", "val"))
 
 .onAttach <- function(libname, pkgname) {
-  packageStartupMessage("Sensitivity analysis as described in Frank, Maroulis, Duong, and Kelcey (2013) and in Frank (2000).\nFor more information visit http://konfound-it.com.")
+  packageStartupMessage("Sensitivity analysis as described in Frank, 
+Maroulis, Duong, and Kelcey (2013) and in 
+Frank (2000).\nFor more information visit http://konfound-it.com.")
 }
