### 2024-02-19 0b39585
Update cop_pse_auxiliary.R（F - FALSE） (#78)
* Update cop_pse_auxiliary.R

* Update cop_pse_auxiliary.R

* Update concord1.R

* Update cop_pse_auxiliary.R

* Update helper_output_print.R

* Update helper_output_dataframe.R

* Update helper_plot_correlation.R

* Update helper_plot_threshold.R

* Update konfound-glm-dichotomous.R

* Update konfound-glm.R

* Update konfound-lm.R

* Update konfound-lmer.R

* Update konfound.R

* Update mkonfound-data.R

* Update mkonfound.R

* Update core-sensitivity-mkonfound.R

* Update pkonfound.R

* Update test_cop.R

* Update test_cop.R

* Update test_pse.R

*
[…]
