+++ R/tkonfound_fig.R
@@ -24,12 +24,14 @@
 #'
 #' @return Returns two plots showing the effect of hypothetical case switches 
 #'         on the effect size in a 2x2 table.
-tkonfound_fig <- function(a, b, c, d, thr_p = 0.05, switch_trm = TRUE, test = "fisher", replace = "control"){
+tkonfound_fig <- function(a, b, c, d, thr_p = 0.05, switch_trm = TRUE, 
+                          test = "fisher", replace = "control"){
 
 n_obs <- a + b + c + d
 ###***generate the log odds for each step of switch
 meta <- data.frame(matrix(ncol = 10, nrow = n_obs-3))
-colnames(meta) <- c("a", "b", "c", "d", "nobs", "switch", "logodds","cntrl_p","tr_p","pdif")
+colnames(meta) <- c("a", "b", "c", "d", "nobs", "switch", 
+                    "logodds","cntrl_p","tr_p","pdif")
 if (switch_trm == TRUE) {
   for (i in 1:(n_obs-3)){ 
     if (i <= a){
@@ -123,7 +125,8 @@ if (test == "chisq"){
 if (test == "fisher"){
   solution <- getswitch_fisher(a, b, c, d, thr_p, switch_trm)
   for (i in 1:(n_obs-3)){
-  meta$logodds[i] <- log(fisher_oddsratio(meta$a[i], meta$b[i], meta$c[i], meta$d[i]))
+  meta$logodds[i] <- log(fisher_oddsratio(meta$a[i], meta$b[i], 
+                                          meta$c[i], meta$d[i]))
   }
 }
 
@@ -272,24 +275,32 @@ if (!switch_trm && !dcroddsratio_ob) {
 }
 
 zoom$label <- ifelse(zoom$sigpoint=="positive", 
-                     paste("sig pos:RIR=", zoom[zoom$sigpoint=="positive",]$RIR),NA)
+                     paste("sig pos:RIR=", 
+                           zoom[zoom$sigpoint=="positive",]$RIR),
+                     NA)
 zoom$label <- ifelse(zoom$sigpoint=="negative", 
-                     paste("sig neg:RIR=", zoom[zoom$sigpoint=="negative",]$RIR),zoom$label)
+                     paste("sig neg:RIR=", 
+                           zoom[zoom$sigpoint=="negative",]$RIR),
+                     zoom$label)
 zoom$label <- ifelse(zoom$sigpoint=="current", 
                      paste("current"),zoom$label)
 
 fig2 <- ggplot2::ggplot(zoom, ggplot2::aes_string(x="RIR",y="pdif"))+
   ggplot2::geom_line(ggplot2::aes_string(y="pdif"), size = 1) +
-  ggplot2::geom_point(ggplot2::aes_string(y="pdif", shape = "current",fill = "sigpoint"), 
+  ggplot2::geom_point(ggplot2::aes_string(y="pdif", 
+                                          shape = "current",
+                                          fill = "sigpoint"), 
                       size = 1)+
   ggrepel::geom_label_repel(ggplot2::aes_string(label="label"))+
   ggplot2::scale_fill_manual(values=fillcol)+
   ggplot2::scale_shape_manual(values=pointshape)+
   ggplot2::scale_y_continuous(name="Difference in probability of successful outcome (treatment - control)")+
   ggplot2::scale_x_continuous(name="RIR (Fragility)", 
-                              breaks= c(zoom$RIR[1], zoom$RIR[as.integer(length(zoom$RIR)/2)], 
+                              breaks= c(zoom$RIR[1], 
+                                        zoom$RIR[as.integer(length(zoom$RIR)/2)], 
                                         zoom$RIR[as.integer(length(zoom$RIR))]),
[…]
