+++ R/mkonfound.R
@@ -37,10 +37,12 @@ mkonfound <- function(d, t, df, alpha = .05, tails = 2, return_plot = FALSE) {
   df_vec <- pull(select(d, !!df_enquo))
 
   if (length(t_vec) <= 1) {
-    stop("To carry out sensitivity analysis for a single study, use pkonfound()")
+    stop("To carry out sensitivity analysis 
+         for a single study, use pkonfound()")
   }
   
-  results_df <- purrr::map2_dfr(.x = t_vec, .y = df_vec, .f = core_sensitivity_mkonfound)
+  results_df <- purrr::map2_dfr(.x = t_vec, .y = df_vec, 
+                                .f = core_sensitivity_mkonfound)
 
   if (return_plot == TRUE) {
     results_df$action <- dplyr::case_when(
@@ -48,7 +50,10 @@ mkonfound <- function(d, t, df, alpha = .05, tails = 2, return_plot = FALSE) {
       results_df$action == "to_sustain" ~ "To Sustain"
     )
 
-    p <- ggplot2::ggplot(results_df, ggplot2::aes(x = results_df$pct_bias_to_change_inference, fill = results_df$action)) +
+    p <- ggplot2::ggplot(results_df, 
+                         ggplot2::aes(
+                           x = results_df$pct_bias_to_change_inference, 
+                                      fill = results_df$action)) +
       ggplot2::geom_histogram() +
       ggplot2::scale_fill_manual("", values = c("#1F78B4", "#A6CEE3")) +
       ggplot2::theme_bw() +

[…]
