### 2024-02-19 478e031
Revert "Update cop_pse_auxiliary.R（F - FALSE） (#78)"
This reverts commit 0b395859b37fadff6e758a8bda6b18114319ccac.
