+++ R/konfound-glm.R
@@ -18,19 +18,13 @@
 #' @importFrom dplyr select filter bind_cols
 #' @importFrom stats glm
 #' @importFrom margins margins
-konfound_glm <- function(model_object, 
-                         tested_variable_string, 
-                         alpha, tails, 
-                         index = "RIR", 
-                         to_return) {
+konfound_glm <- function(model_object, tested_variable_string, alpha, tails, index = "RIR", to_return) {
     tidy_output <- broom::tidy(model_object) # tidying output
     glance_output <- broom::glance(model_object)
     
     coef_df <- tidy_output[tidy_output$term == tested_variable_string, ]
     est_eff <- coef_df$estimate
-    est_eff <- suppressWarnings(
-      summary(margins::margins(model_object))$AME[names(summary(
-        margins::margins(model_object))$AME) == tested_variable_string])
+    est_eff <- suppressWarnings(summary(margins::margins(model_object))$AME[names(summary(margins::margins(model_object))$AME) == tested_variable_string])
     std_err <- coef_df$std.error
     n_obs <- glance_output$nobs
     n_covariates <- glance_output$df.null - glance_output$df.residual
