+++ R/konfound-lmer.R
@@ -10,8 +10,7 @@
 get_kr_df <- function(model_object) {
     L <- diag(rep(1, length(lme4::fixef(model_object))))
     L <- as.data.frame(L)
-    out <- suppressWarnings(purrr::map_dbl(L, pbkrtest::get_Lb_ddf, 
-                                           object = model_object))
+    out <- suppressWarnings(purrr::map_dbl(L, pbkrtest::get_Lb_ddf, object = model_object))
     names(out) <- names(lme4::fixef(model_object))
     out
 }
@@ -32,13 +31,7 @@ get_kr_df <- function(model_object) {
 #' @return The results of the konfound analysis for the specified fixed effect(s).
 #' @importFrom broom.mixed tidy
 #' @importFrom dplyr filter bind_cols
-konfound_lmer <- function(model_object, 
-                          tested_variable_string, 
-                          test_all, 
-                          alpha, 
-                          tails, 
-                          index, 
-                          to_return) {
+konfound_lmer <- function(model_object, tested_variable_string, test_all, alpha, tails, index, to_return) {
     tidy_output <- broom.mixed::tidy(model_object) # tidying output
     
     coef_df <- tidy_output[tidy_output$term == tested_variable_string, ]
