### 2024-02-19 konfound-project/konfound#50 @jrosen48
#78 addresses:

- [ ] lots of T and F instead of TRUE/FALSE
- [ ] Namespaces in Imports field not imported from: ‘cli’ ‘withr’
- [ ] It also addresses many of the long code lines issue, but not all, with most of the remaining being lengthy questions or comments

We have NOT yet addressed this:
- [ ] All declared Imports should be used.

It is not immediately clear to us what is causing that not yet addressed issue, as we have used `@importFrom` roxygen2 tags to specify which specific functions we import in our NAMESPACE file. We are trying to track down what is causing this, but the `goodpractice::gp()` output doesn't provide a clear pointer in terms of a marker or its summary output.
