### 2024-02-20 konfound-project/konfound#50 @jrosen48
The `NAMESPACE` file suggest it is:

```
import(ggplot2)
import(lavaan)
import(rlang)
```
