### 2024-02-20 08632ee
addressing namespace issue - trying to rewrite the NAMESPACE and docs
