+++ man/mkonfound_ex.Rd
@@ -20,6 +20,7 @@ mkonfound_ex
 }
 \description{
 A dataset containing t and df values from example studies from Educational Evaluation
-and Policy Analysis (as detailed in Frank et al., 2013): https://drive.google.com/file/d/1aGhxGjvMvEPVAgOA8rrxvA97uUO5TTMe/view
+and Policy Analysis (as detailed in Frank et al., 2013): 
+https://drive.google.com/file/d/1aGhxGjvMvEPVAgOA8rrxvA97uUO5TTMe/view
 }
 \keyword{datasets}
