+++ man/konfound-package.Rd
@@ -1,33 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/konfound-package.R
-\docType{package}
-\name{konfound-package}
-\alias{konfound}
-\alias{konfound-package}
-\title{konfound: Quantify the Robustness of Causal Inferences}
-\description{
-Statistical methods that quantify the conditions necessary to alter inferences, also known as sensitivity analysis, are becoming increasingly important to a variety of quantitative sciences. A series of recent works, including Frank (2000) \doi{10.1177/0049124100029002001} and Frank et al. (2013) \doi{10.3102/0162373713493129} extend previous sensitivity analyses by considering the characteristics of omitted variables or unobserved cases that would change an inference if such variables or cases were observed. These analyses generate statements such as "an omitted variable would have to be correlated at xx with the predictor of interest (e.g., treatment) and outcome to invalidate an inference of a treatment effect". Or "one would have to replace pp percent of the observed data with null hypothesis cases to invalidate the inference". We implement these recent developments of sensitivity analysis and provide modules to calculate these two robustness indices and generate such statements in R. In particular, the functions konfound(), pkonfound() and mkonfound() allow users to calculate the robustness of inferences for a user's own model, a single published study and multiple studies respectively.
-}
-\seealso{
-Useful links:
-\itemize{
-  \item \url{https://github.com/konfound-project/konfound}
-  \item \url{https://konfound-project.github.io/konfound/}
-  \item Report bugs at \url{https://github.com/konfound-project/konfound/issues}
-}
-
-}
-\author{
-\strong{Maintainer}: Joshua M Rosenberg \email{jmrosen48@gmail.com}
-
-Other contributors:
-\itemize{
-  \item Ran Xu \email{ranxu@msu.edu} [contributor]
-  \item Qinyun Lin \email{linqinyun1001@gmail.com} [contributor]
-  \item Spiro Maroulis \email{Spiro.Maroulis@asu.edu} [contributor]
-  \item Sarah Narvaiz \email{snarvaiz@alum.utk.edu} [contributor]
-  \item Kenneth A Frank \email{kenfrank@msu.edu} [contributor]
-}
-
-}
-\keyword{internal}
