### 2024-02-20 7951370
remove glue and lifecycle
