+++ DESCRIPTION
@@ -48,9 +48,7 @@ Imports:
     margins,
     tibble,
     ggrepel,
-    pbkrtest,
-    lifecycle,
-    glue
+    pbkrtest
 Suggests: 
     covr,
     devtools,
