### 2024-02-20 95df1c7
make edits to remove import lavaan and import ggplot2
