+++ R/pkonfound.R
@@ -27,8 +27,6 @@
 #' @param to_return whether to return a data.frame (by specifying this argument to equal "raw_output" for use in other analyses) or a plot ("plot"); default is to print ("print") the output to the console; can specify a vector of output to return
 #' @importFrom stats fisher.test
 #' @importFrom dplyr select
-#' @import lavaan
-#' @import ggplot2
 #' @return prints the bias and the number of cases that would have to be replaced with cases for which there is no effect to invalidate the inference
 #' @examples
 #' # using pkonfound for linear models
