+++ NAMESPACE
@@ -5,8 +5,6 @@ export(mkonfound)
 export(pkonfound)
 export(summary)
 export(tkonfound)
-import(ggplot2)
-import(lavaan)
 importFrom(broom,glance)
 importFrom(broom,tidy)
 importFrom(broom.mixed,tidy)
