### 2024-02-20 d87f0d5
remove summary function to simplify codebase and address remaining NAMESPACE issue
