### 2024-02-20 konfound-project/konfound#50 @jrosen48
We have addressed the remaining goodpractice issue; as an update:

- [ x ] lots of T and F instead of TRUE/FALSE
- [ x ] Namespaces in Imports field not imported from: ‘cli’ ‘withr’
- [ x ] remaining assignment operator issues (for assignment, changing `=` to `<-`)
- [ x ] All declared Imports should be used.
- [  x ] also many of the long code lines issue, but not all, with most of the remaining being lengthy questions or comments

From `goodpractice::gp()`, we are seeing:

```{r}
It is good practice to

  ✖ write unit tests for all functions, and all
    package code in general. 80% of code lines are covered
    by test cases.

    R/cop_pse_auxiliary.R:12:NA
    R/cop_pse_auxiliary.R:28:NA
    R/cop_pse_auxiliary.R:69:NA
    R/cop_pse_auxiliary.R:141:NA
    R/cop_pse_auxiliary.R:142:NA
    ... and 495 more lines

  ✖ write short and simple functions. These
    functions have high cyclomatic complexity (>50):
    check_starting_table (79), getswitch_fisher (79),
    tkonfound_fig (74), getswitch_chisq (70), tkonfound
    (70), test_sensitivity_ln (59). You can make them
    easier to reason about by encapsulating distinct steps
    of your function into subfunctions.
  ✖ avoid long code lines, it is bad for
    readability. Also, many people prefer editor windows
    that are about 80 characters wide. Try make your lines
    shorter than 80 characters

    R/cop_pse_auxiliary.R:233:81
    R/cop_pse_auxiliary.R:234:81
    R/cop_pse_auxiliary.R:235:81
    R/cop_pse_auxiliary.R:236:81
    R/cop_pse_auxiliary.R:237:81
    ... and 285 more lines
```
