### 2024-02-20 #5779 @jrosen48
As a status update, we have a comment pending review from @fabian-s in https://github.com/konfound-project/konfound/issues/50. We are also almost complete with the paper edits and aim to have an update before the end of the day tomorrow.
