### 2024-02-20 konfound-project/konfound#81 @jrosen48
JOSS paper edits
- [ x ] 1. Line 149, SE should have a hat and standard error should be standard error estimate. *NOTE: `SE` was a function argument that we have since renamed `std_err. We have changed standard error to estimated standard error.*
- [ x ] 2. Number of digits should be consistent throughout the text: e.g., .568 in line 173 vs. .52 in line 109. *NOTE: This is due to rounding in the underlying code (.52 is equivalent to .520). @fartist would it be preferable to ensure there are a fixed number of digits, i.e. the output would present .520?*
- [ x ] 3. APA style were not followed at multiple places e.g., line 108, line 134. *NOTE: This is addressed apart from K.A. appearing before Frank's last name - we will address this as we make final edits to the paper*.
- [ x ] 4. The statement of need does not address a specific target audience. To me the implied audience is applied researchers, but that's not explicitly stated anywhere.
- [ x ] 5. The paper does not currently describe how this software compares to existing packages for sensitivity analyses, or mention what other packages are available. *Note: we have notes on this to integrate into the text:

Vanderweel’s E-value package deals with sensitivity for omitted variables for dichotomous outcomes.  It is an improvement on Rosenbaum. We address dichotomous outcomes through case replacement and fragility.
Cinelli’s Robustness index deals with sensitivity of an estimate for continuous outcomes in terms of omitted variables.  ITCV accounts for simultaneous changes in estimates and standard errors.  RIR is an entirely different approach
Many others used simulation based approaches (e.g., Carnegie et al).  E.g. Franks et al.  Baer et al.  Our approach uses closed form expressions to generate a single term representing sensitivity.
Oster’s coefficient of proportionality focuses only on selection into the treatment based on unobservables versus observables necessary to nullify an estimate.  ITCV focuses on relationship of the unobservable to the predictor of interest and  to the outcome. Sensitivity analyses broadly reviewed here.
Frank, K.A., Lin, Q., Xu, R., Maroulis, S.J., Mueller, A. (on-line first). [Quantifying the Robustness of Causal Inferences: Sensitivity Analysis for Pragmatic Social Science](https://www.dropbox.com/s/4vvpvgqhpe2h1rv/Quantifying%20the%20robustness%20of%20causal%20inferences%20as%20published%20sensitivity%20analysis%20for%20pragmatic%20social%20science.pdf?dl=0).  Social Science Research. 110, 102815. ERIC:ED628601

- [ x ] 6. Some of the output in the example uses runs off the edge of the page (this may be a typesetting issue?) *Note: We have mostly addressed this, but there are still a few cases where the output approaches running off the page.*
