### 2024-02-20 konfound-project/konfound#81 @jrosen48
another issue that would be helpful to look at - issue 3 above. For some reason, Ken's first initials are appearing in the in-text citations, so they read as e.g. K.A. Frank et al. (2022) instead of Frank et al. (2022). Could you look into how the references are entered in the `.bib` file for the paper?
