### 2024-02-20 konfound-project/konfound#81 @jrosen48
@qinyun-lin if you have time, would you be able to look at issue 5? The text above is from Ken, so it just needs to be integrated into the first few paragraphs of the paper (into the introduction of the paper). Aiming to do this by tomorrow.
