### 2024-02-20 konfound-project/konfound#81 @wwang93
> @wwang93 could you look at issue 6 above? You can see the latest version of the paper here: https://github.com/openjournals/joss-papers/blob/joss.05779/joss.05779/10.21105.joss.05779.pdf

I could
