### 2024-02-20 konfound-project/konfound#81 @jrosen48
Yep https://github.com/konfound-project/konfound/commit/a2ee5b9068e721e1cf5f442d508be00615918be8 from you addresses issue 6, editing above
