### 2024-02-20 konfound-project/konfound#81 @fartist
I am fine with the current version.

  *   [ ? ] 2. Number of digits should be consistent throughout the text: e.g., .568 in line 173 vs. .52 in line 109. NOTE: This is due to rounding in the underlying code (.52 is equivalent to .520). @fartist<https://github.com/fartist> would it be preferable to ensure there are a fixed number of digits, i.e. the output would present .520?

From: Joshua Rosenberg ***@***.***>
Date: Tuesday, February 20, 2024 at 4:46 PM
To: konfound-project/konfound ***@***.***>
Cc: Feeeeeeeeeeng ***@***.***>, Mention ***@***.***>
Subject: Re: [konfound-project/konfound] JOSS paper edits (Issue #81)

Yep a2ee5b9<https://github.com/konfound-project/konfound/commit/a2ee5b9068e721e1cf5f442d508be00615918be8> from you addresses issue 6, editing above
