+++ joss-paper/paper.bib
@@ -44,14 +44,14 @@
 	doi = {10.1177/0049124100029002001},
 	number = {2},
 	journal = {Sociological Methods \& Research},
-	author = {{Frank}, K.~A.},
+	author = {Frank, Kenneth A},
 	month = nov,
 	year = 2000,
 	pages = {147-194}
 }
 
 @article{frank2007,
-author = {{Frank}, K.~A. and {Min}, K.},
+author = {Frank, Kenneth A. and {Min}, K.},
 title ={10. Indices of Robustness for Sample Representation},
 journal = {Sociological Methodology},
 volume = {37},
@@ -65,7 +65,7 @@ eprint = {https://doi.org/10.1111/j.1467-9531.2007.00186.x}
 
 @article{frank2008,
  URL = {http://www.jstor.org/stable/30128050},
- author = {{Frank}, K.~A. and {Sykes}, G. and {Anagnostopoulos}, D. and {Cannata}, M. and {Chard}, L. and {Krause}, A. and {McCrory}, R.},
+ author = {Frank, Kenneth A. and {Sykes}, G. and {Anagnostopoulos}, D. and {Cannata}, M. and {Chard}, L. and {Krause}, A. and {McCrory}, R.},
  journal = {Educational Evaluation and Policy Analysis},
  number = {1},
  pages = {3-30},
@@ -83,7 +83,7 @@ eprint = {https://doi.org/10.1111/j.1467-9531.2007.00186.x}
 	doi = {10.3102/0162373713493129},
 	number = {4},
 	journal = {Educational Evaluation and Policy Analysis},
-	author = {{Frank}, K.~A. and {Maroulis}, S.~J. and {Duong}, M.~Q. and {Kelcey}, B.~M.},
+	author = {Frank, Kenneth A. and {Maroulis}, S.~J. and {Duong}, M.~Q. and {Kelcey}, B.~M.},
 	month = dec,
 	year = 2013,
 	note = {Publisher: American Educational Research Association},
@@ -99,14 +99,14 @@ year = 2021,
 issn = {0895-4356},
 doi = {10.1016/j.jclinepi.2021.01.025},
 url = {https://www.sciencedirect.com/science/article/pii/S0895435621000366},
-author = {{Frank}, K.~A. and {Lin}, Q. and {Maroulis}, S. and {Mueller}, A.~S. and {Xu}, R. and {Rosenberg}, J.~M. and {Hayter}, C.~S. and {Mahmoud}, R.~A. and {Kolak}, M. and {Dietz}, T. and {Zhang}, L.},
+author = {Frank, Kenneth A. and {Lin}, Q. and {Maroulis}, S. and {Mueller}, A.~S. and {Xu}, R. and {Rosenberg}, J.~M. and {Hayter}, C.~S. and {Mahmoud}, R.~A. and {Kolak}, M. and {Dietz}, T. and {Zhang}, L.},
 keywords = {Robustness of findings, Randomized controlled trials, Fragility, Case replacement, Statistical significance, Clinical importance}
 }
 
 @article{frank2023,
 title = {Quantifying the robustness of causal inferences: Sensitivity analysis for pragmatic social science},
 keywords = "Causal inference, Pragmatic sociology, Sensitivity analysis",
-author = {{Frank}, K.~A. and {Lin}, Q. and {Xu}, R.  and {Maroulis}, S.  and {Mueller}, A.},
+author = {Frank, Kenneth A. and {Lin}, Q. and {Xu}, R.  and {Maroulis}, S.  and {Mueller}, A.},
 year = 2023,
 month = feb,
 doi = {10.1016/j.ssresearch.2022.102815},
[…]
