+++ joss-paper/paper.bib
@@ -14,7 +14,7 @@
 
 @article{frank2023,
   title={Quantifying the robustness of causal inferences: Sensitivity analysis for pragmatic social science},
-  author={Frank, Kenneth A and Lin, Qinyun and Xu, Ran and Maroulis, Spiro and Mueller, Anna},
+  author={Frank and Lin, Qinyun and Xu, Ran and Maroulis, Spiro and Mueller, Anna},
   journal={Social Science Research},
   volume={110},
   pages={102815},
@@ -44,7 +44,7 @@
 	doi = {10.1177/0049124100029002001},
 	number = {2},
 	journal = {Sociological Methods \& Research},
-	author = {Frank, Kenneth A},
+	author = {Frank},
 	month = nov,
 	year = 2000,
 	pages = {147-194}
