+++ joss-paper/paper.bib
@@ -51,7 +51,7 @@
 }
 
 @article{frank2007,
-author = {{Frank} and {Min}, K.},
+author = {{Frank} and {Min}},
 title ={10. Indices of Robustness for Sample Representation},
 journal = {Sociological Methodology},
 volume = {37},
