### 2024-02-21 5454ce3
edits to paper.bib to address APA issues
