+++ joss-paper/paper.bib
@@ -1,6 +1,6 @@
 
 @article{razavi2021,
-	title = {The Future of Sensitivity Analysis: An essential discipline for systems modeling and policy support},
+	title = {The future of sensitivity analysis: An essential discipline for systems modeling and policy support},
 	volume = {137},
 	issn = {1364-8152},
 	url = {https://www.sciencedirect.com/science/article/pii/S1364815220310112},
@@ -14,7 +14,7 @@
 
 @article{frank2023,
   title={Quantifying the robustness of causal inferences: Sensitivity analysis for pragmatic social science},
-  author={Frank and Lin, Qinyun and Xu, Ran and Maroulis, Spiro and Mueller, Anna},
+  author={Frank, K.~A. and Lin, Qinyun and Xu, Ran and Maroulis, Spiro and Mueller, Anna},
   journal={Social Science Research},
   volume={110},
   pages={102815},
@@ -23,7 +23,7 @@
 }
 
 @article{cinelli2019,
-    author = {{Cinelli}, C. and {Hazlett}, C.},
+    author = {Cinelli, C. and Hazlett, C.},
     title = "{Making Sense of Sensitivity: Extending Omitted Variable Bias}",
     journal = {Journal of the Royal Statistical Society Series B: Statistical Methodology},
     volume = {82},
@@ -44,15 +44,15 @@
 	doi = {10.1177/0049124100029002001},
 	number = {2},
 	journal = {Sociological Methods \& Research},
-	author = {Frank},
+	author = {Frank, K.~A.},
 	month = nov,
 	year = 2000,
 	pages = {147-194}
 }
 
 @article{frank2007,
-author = {{Frank} and {Min}},
-title ={10. Indices of Robustness for Sample Representation},
+author = {Frank, K.~A. and Min, K.-S.},
+title ={Indices of Robustness for Sample Representation},
 journal = {Sociological Methodology},
 volume = {37},
 number = {1},
@@ -65,7 +65,7 @@ eprint = {https://doi.org/10.1111/j.1467-9531.2007.00186.x}
 
 @article{frank2008,
  URL = {http://www.jstor.org/stable/30128050},
- author = {{Frank} and {Sykes}, G. and {Anagnostopoulos}, D. and {Cannata}, M. and {Chard}, L. and {Krause}, A. and {McCrory}, R.},
+ author = {Frank, K. A. and Sykes, G. and Anagnostopoulos, D. and Cannata, M. and Chard, L. and Krause, A. and McCrory, R.},
  journal = {Educational Evaluation and Policy Analysis},
  number = {1},
  pages = {3-30},
@@ -83,7 +83,7 @@ eprint = {https://doi.org/10.1111/j.1467-9531.2007.00186.x}
 	doi = {10.3102/0162373713493129},
 	number = {4},
 	journal = {Educational Evaluation and Policy Analysis},
-	author = {{Frank} and {Maroulis}, S.~J. and {Duong}, M.~Q. and {Kelcey}, B.~M.},
+	author = {Frank, K.~A. and Maroulis, S.~J. and Duong, M.~Q. and Kelcey, B.~M.},
 	month = dec,
 	year = 2013,
 	note = {Publisher: American Educational Research Association},
@@ -99,14 +99,14 @@ year = 2021,
 issn = {0895-4356},
 doi = {10.1016/j.jclinepi.2021.01.025},
 url = {https://www.sciencedirect.com/science/article/pii/S0895435621000366},
[…]
