### 2024-02-21 09c675f
edits to the paper and .bib to related to issues 3 and 5 in #80
