+++ joss-paper/paper.bib
@@ -12,16 +12,6 @@
 	keywords = {Decision making, Machine learning, Mathematical modeling, Model robustness, Model validation and verification, Policy support, Sensitivity analysis, Uncertainty quantification}
 }
 
-@article{frank2023,
-  title={Quantifying the robustness of causal inferences: Sensitivity analysis for pragmatic social science},
-  author={Frank, K.~A. and Lin, Qinyun and Xu, Ran and Maroulis, Spiro and Mueller, Anna},
-  journal={Social Science Research},
-  volume={110},
-  pages={102815},
-  year={2023},
-  publisher={Elsevier}
-}
-
 @article{cinelli2019,
     author = {Cinelli, C. and Hazlett, C.},
     title = "{Making Sense of Sensitivity: Extending Omitted Variable Bias}",
