+++ joss-paper/paper.md
@@ -45,27 +45,35 @@ affiliations:
   index: 5
 ---
 
-# Quantifying the Robustness of Inferences
+## Quantifying the Robustness of Inferences
 
-Statistical methods which quantify the conditions necessary to alter inferences are important to a variety of disciplines [@razavi2021]. One line of work is rooted in linear models and foregrounds the sensitivity of inferences to the strength of omitted variables [@frank2000; @cinelli2019]. A more recent approach is rooted in the potential outcomes framework for causal inference and foregrounds how hypothetical changes in a sample would alter an inference if such variables or cases were otherwise observed [@frank2007; @frank2008; @frank2013; @xu2019). It differs from other sensitivity analysis approaches and packages (e.g., in the variety of dependent variable types it accomodates, its internal logic, and its analytic (rather than simulation-based) approach [@frank2023].
+Sensitivity analysis, a statistical method crucial for validating inferences across disciplines, quantifies the conditions that could alter conclusions [@razavi2021]. One line of work is rooted in linear models and foregrounds the sensitivity of inferences to the strength of omitted variables [@frank2000; @cinelli2019]. A more recent approach is rooted in the potential outcomes framework for causal inference and foregrounds how hypothetical changes in a sample would alter an inference if such variables or cases were otherwise observed [@frank2007; @frank2008; @frank2013; @xu2019].
 
-We have implemented two measures from these lines of work within R via the `konfound` R package. The audience for this R package is primarily social scientists, as well as interested individuals in other fields. One measure is the Impact Threshold of a Confounding Variable (ITCV), which generates statements such as "to invalidate an inference of an effect, an omitted variable would have to be correlated at \_\_ with the predictor of interest and with the outcome" [@frank2000]. This sensitivity analysis can be calculated for any linear model. The second measure is the Robustness of an Inference to Replacement (RIR) which generates statements such as "to invalidate the inference, \_\_ % of the cases would have to be replaced with counterfactual cases with zero effect of the treatment" [@frank2013]. The RIR represents a more generally applicable approach not limited to linear models, and is recommended for all cases that use binary outcomes [@frank2021].
+One measure is the Impact Threshold of a Confounding Variable (ITCV), which generates statements such as "to invalidate an inference of an effect, an omitted variable would have to be correlated at \_\_ with the predictor of interest and with the outcome" [@frank2000]. This sensitivity analysis can be calculated for any linear model. The Robustness of an Inference to Replacement (RIR) assesses how replacing a certain percentage of cases with counterfactuals of zero treatment effect could undermine an inference.  The Robustness of an Inference to Replacement (RIR) is more general than the ITCV index. It assesses how replacing a certain percentage of cases with counterfactuals of zero treatment effect could undermine an inference [@frank2013]; it is particularly appropriate for models with binary outcomes [@frank2021].
 
-# Statement of Need: The Need for an R Package
+The sensitivity analysis technique we describe in this paper and implement in an R package differ from others in several ways. Unlike @linden2020conducting, which focuses on dichotomous outcomes and omitted variable sensitivity, our approach extends to continuous outcomes and evaluates both changes in estimates and standard errors. @oster2019unobservable focuses only on selection into the treatment based on unobservable variables versus observable variables necessary to nullify an estimate. The ITCV index focuses on relationship of the unobservable to the predictor of interest and to the outcome. More generally, many others used simulation based approaches, while our approach uses closed form expressions to generate a single term representing sensitivity. These techniques along with others are reviewed and discussed along with the ITCV and RIR approaches that are implemented in the konfound package in @frank2023. 
 
[…]
