### 2024-02-21 konfound-project/konfound#81 @jrosen48
issues 3 and 5 should now be addressed in https://github.com/konfound-project/konfound/commit/09c675f6d079ca77f067df77902078ff5033867c; waiting for the checks to pass before merging and re-rendering the PDF
