### 2024-02-21 80283da
minor edits to .bib
