+++ joss-paper/paper.bib
@@ -12,6 +12,29 @@
 	keywords = {Decision making, Machine learning, Mathematical modeling, Model robustness, Model validation and verification, Policy support, Sensitivity analysis, Uncertainty quantification}
 }
 
+
+@article{linden2020conducting,
+  title={Conducting sensitivity analysis for unmeasured confounding in observational studies using E-values: the evalue package},
+  author={Linden, Ariel and Mathur, Maya B and VanderWeele, Tyler J},
+  journal={The Stata Journal},
+  volume={20},
+  number={1},
+  pages={162--175},
+  year={2020},
+  publisher={SAGE Publications Sage CA: Los Angeles, CA}
+}
+
+@article{oster2019unobservable,
+  title={Unobservable selection and coefficient stability: Theory and evidence},
+  author={Oster, Emily},
+  journal={Journal of Business \& Economic Statistics},
+  volume={37},
+  number={2},
+  pages={187--204},
+  year={2019},
+  publisher={Taylor \& Francis}
+}
+
 @article{cinelli2019,
     author = {Cinelli, C. and Hazlett, C.},
     title = "{Making Sense of Sensitivity: Extending Omitted Variable Bias}",
