+++ .github/CONTRIBUTING.md
@@ -1,7 +1,6 @@
 # Contributing to konfound
 
 This outlines how to propose a change to konfound.
-For a detailed discussion on contributing to this and other tidyverse packages, please see the [development contributing guide](https://rstd.io/tidy-contrib) and our [code review principles](https://code-review.tidyverse.org/).
 
 ## Fixing typos
 
@@ -11,7 +10,8 @@ You can find the `.R` file that generates the `.Rd` by reading the comment in th
 
 ## Bigger changes
 
-If you want to make a bigger change, it's a good idea to first file an issue and make sure someone from the team agrees that it’s needed. 
+If you want to make a bigger change, it's a good idea to first file an issue to notify the team.
+
 If you’ve found a bug, please file an issue that illustrates the bug with a minimal 
 [reprex](https://www.tidyverse.org/help/#reprex) (this will also help you write a unit test, if needed).
 See our guide on [how to create a great issue](https://code-review.tidyverse.org/issues/) for more advice.
@@ -23,12 +23,17 @@ You may also wish to contact the development team for bigger changes. Please see
 *   Fork the package and clone onto your computer. If you haven't done this before, we recommend using `usethis::create_from_github("konfound-project/konfound", fork = TRUE)`.
 
 *   Install all development dependencies with `devtools::install_dev_deps()`, and then make sure the package passes R CMD check by running `devtools::check()`. 
-    If R CMD check doesn't pass cleanly, it's a good idea to ask for help before continuing. 
-*   Create a Git branch for your pull request (PR). We recommend using `usethis::pr_init("brief-description-of-change")`.
+    If R CMD check doesn't pass cleanly, it's a good idea to ask for help before continuing.
+    
+*   Create a Git branch for your pull request (PR). We recommend using `usethis::pr_init("brief-description-of-change")`. This is important: even for team members, please make commits to branches, first.
+
+*   Ensure that all checks are passing. You can see this information within your PR. If a check is not passing, correct the issue or contact a package maintainer for help.
 
 *   Make your changes, commit to git, and then create a PR by running `usethis::pr_push()`, and following the prompts in your browser.
     The title of your PR should briefly describe the change.
     The body of your PR should contain `Fixes #issue-number`.
+
+*   A member of the team will then review your PR.
 
 *  For user-facing changes, add a bullet to the top of `NEWS.md` (i.e. just below the first header). Follow the style described in <https://style.tidyverse.org/news.html>.
 
