### 2024-02-21 #5779 @jrosen48
@wjakethompson we have added a `CONTRIBUTING.md` file: https://github.com/konfound-project/konfound/blob/master/.github/CONTRIBUTING.md

This includes notes on using branches for any non-typo edits and the importance of ensuring that all checks successfully pass.

We think this will help us as a team to ensure good coding practice (and communicate in writing to open-source contributors about our development process).
