### 2024-02-21 8f4897d
a few more edits to address formatting issues with the paper
