### 2024-02-21 bb53752
a few more minor edits to the paper after viewing the rendered PDF
