### 2024-02-21 #5779 @jrosen48
We have made all of the requested edits and have addressed the outstanding issues. See https://github.com/konfound-project/konfound/issues/50 and https://github.com/konfound-project/konfound/issues/81 for the two issues that need approval to be closed.
