### 2024-02-21 konfound-project/konfound#81 @jrosen48
These issues are now all addressed. The paper can be previewed here: https://github.com/openjournals/joss-papers/blob/joss.05779/joss.05779/10.21105.joss.05779.pdf
