### 2024-02-21 konfound-project/konfound#54 @fabian-s
there is no CONTRIBUTING guide, just a code of conduct. this seems to  have been fixed only partially
