### 2024-02-21 #5779 @fabian-s
now that software quality issues seem to have been sufficiently improved,  please adress the remaining open points raised by @wjakethompson, see https://github.com/openjournals/joss-reviews/issues/5779#issuecomment-1726064309 (and https://github.com/konfound-project/konfound/issues/54)
