### 2024-02-21 837ec57
update README to link to CONTRIBUTING.md and minor edits to paper.md based on Ken's feedback
