+++ joss-paper/paper.md
@@ -5,7 +5,7 @@ tags:
 - Sensitivity analysis
 - Causal inference
 - R
-date: "2024-02-13"
+date: "2024-02-21"
 output:
   pdf_document: defaulted and 
 authors:
@@ -49,9 +49,9 @@ affiliations:
 
 Sensitivity analysis, a statistical method crucial for validating inferences across disciplines, quantifies the conditions that could alter conclusions [@razavi2021]. One line of work is rooted in linear models and foregrounds the sensitivity of inferences to the strength of omitted variables [@frank2000; @cinelli2019]. A more recent approach is rooted in the potential outcomes framework for causal inference and foregrounds how hypothetical changes in a sample would alter an inference if such variables or cases were otherwise observed [@frank2007; @frank2008; @frank2013; @xu2019].
 
-One measure is the Impact Threshold of a Confounding Variable (ITCV), which generates statements about the correlation of an omitted, confounding variable with both a predictor of interest and with the outcome [@frank2000]. This sensitivity analysis can be calculated for any linear model. The Robustness of an Inference to Replacement (RIR) assesses how replacing a certain percentage of cases with counterfactuals of zero treatment effect could undermine an inference. The RIR index is more general than the ITCV index. It assesses how replacing a certain percentage of cases with counterfactuals of zero treatment effect could undermine an inference [@frank2013].
+One sensitivity measure is the *Impact Threshold of a Confounding Variable*, or ITCV, which generates statements about the correlation of an omitted, confounding variable with both a predictor of interest and with the outcome [@frank2000]. The ITCV index can be calculated for any linear model. The *Robustness of an Inference to Replacement*, RIR, assesses how replacing a certain percentage of cases with counterfactuals of zero treatment effect could undermine an inference [@frank2013]. The RIR index is more general than the ITCV index.
 
-The sensitivity analysis technique we describe in this paper and implement in an R package differ from others in several ways. Unlike @linden2020conducting, whose approach focuses on dichotomous outcomes and omitted variable sensitivity, our approach extends to continuous outcomes and evaluates both changes in estimates and standard errors. @oster2019unobservable focuses only on selection into the treatment based on unobservable variables versus observable variables necessary to nullify an estimate. The ITCV index focuses on relationship of the unobservable to the predictor of interest and to the outcome. More generally, many others used simulation based approaches, while our approach uses closed form expressions to generate a single term representing sensitivity. These techniques along with others are reviewed and discussed along with the ITCV and RIR approaches that are implemented in the konfound package in @frank2023. 
+The sensitivity analysis techniques we describe in this paper and implement in an R package differ from others in several ways. Unlike @linden2020conducting, whose approach focuses on dichotomous outcomes and omitted variable sensitivity, our approach extends to continuous outcomes and evaluates both changes in estimates and standard errors. @oster2019unobservable focuses only on selection into the treatment based on unobservable variables versus observable variables necessary to nullify an estimate. The ITCV index focuses on relationship of the unobservable to the predictor of interest and to the outcome. More generally, many others used simulation based approaches, while our approach uses closed form expressions to generate a single term representing sensitivity. These techniques along with others are reviewed and discussed along with the ITCV and RIR approaches that are implemented in the konfound package in @frank2023. 
 
[…]
