### 2024-02-21 konfound-project/konfound#54 @jrosen48
Oye, sorry, the CONTRIBUTING.md file exists here: https://github.com/konfound-project/konfound/blob/master/.github/CONTRIBUTING.md

The file we have added notes on using branches for any non-typo edits and the importance of ensuring that all checks successfully pass. We think this will help us as a team to ensure good coding practice (and communicate in writing to open-source contributors about our development process).

We have now mentioned and linked to this file in the README.md file; see here: https://github.com/konfound-project/konfound
