### 2024-02-21 konfound-project/konfound#54 @fabian-s
great, thx.  this is usually a top level file but linking in the README is fine as well.
