### 2024-02-21 77eec32
Update helper_output_print.R
