+++ R/helper_output_print.R
@@ -107,7 +107,7 @@ if (requireNamespace("htmltools", quietly = TRUE)) {
       cat("\n")
       cat(paste0("for a null hypothesis of 0 effect is based on a correlation of ", round(r_con, 3)))
       cat("\n")
-      cat(paste0("with the outcome and at ", round(r_con, 3), " with the predictor of interest (conditioning"))
+      cat(paste0("with the outcome and ", round(r_con, 3), " with the predictor of interest (conditioning"))
       cat("\n")
       cat(paste0("on all observed covariates in the model) based on a threshold of ", round(critical_r, 3)))
       cat("\n")
@@ -122,7 +122,7 @@ if (requireNamespace("htmltools", quietly = TRUE)) {
       cat("\n")
       cat(paste0("invalidate an inference for a null hypothesis of 0 effect is based on"))
       cat("\n")
-      cat(paste0("a correlation of ", -round(r_con, 3), " with the outcome and at ", round(r_con, 3), " with the predictor"))
+      cat(paste0("a correlation of ", -round(r_con, 3), " with the outcome and ", round(r_con, 3), " with the predictor"))
       cat("\n")
       cat("of interest (conditioning on all observed covariates in the model;")
       cat("\n")
@@ -138,7 +138,7 @@ if (requireNamespace("htmltools", quietly = TRUE)) {
       cat("\n")
       cat("sustain an inference for a null hypothesis of 0 effect is based on")
       cat("\n")
-      cat(paste0("a correlation of ", -round(r_con, 3), " with the outcome an at ", round(r_con, 3), " with the predictor"))
+      cat(paste0("a correlation of ", -round(r_con, 3), " with the outcome and ", round(r_con, 3), " with the predictor"))
       cat("\n")
       cat("of interest (conditioning on all observed covariates in the model;")
       cat("\n")
@@ -154,7 +154,7 @@ if (requireNamespace("htmltools", quietly = TRUE)) {
       cat("\n")
       cat(paste0("for a null hypothesis of 0 effect is based on a correlation of ", round(r_con, 3)))
       cat("\n")
-      cat(paste0("with the outcome and at ", round(r_con, 3), " with the predictor of interest (conditioning"))
+      cat(paste0("with the outcome and ", round(r_con, 3), " with the predictor of interest (conditioning"))
       cat("\n")
       cat(paste0("on all observed covariates in the model) based on a threshold of ", round(beta_threshhold, 3)))
       cat("\n")
