### 2024-02-22 #5779 @fabian-s
@jrosen48 please verify and add the missing DOIs above to your references
