### 2024-02-22 308fbc3
edits to paper.bib and paper.md in response to request from @fabian-s: https://github.com/openjournals/joss-reviews/issues/5779#issuecomment-1959077141
