### 2024-02-22 #5779 @jrosen48
I think I've addressed. I will regenerate and re-check DOIs to confirm that.
