+++ joss-paper/paper.bib
@@ -21,7 +21,7 @@
   number={1},
   pages={162--175},
   year={2020},
-  doi={https://doi.org/10.1177/1536867X20909696}
+  doi={https://doi.org/10.1177/1536867X20909696},
   publisher={SAGE Publications Sage CA: Los Angeles, CA}
 }
 
