### 2024-02-22 95cd69a
additional edits to .bib
