+++ joss-paper/paper.bib
@@ -21,7 +21,7 @@
   number={1},
   pages={162--175},
   year={2020},
-  doi={https://doi.org/10.1177/1536867X20909696},
+  doi={10.1177/1536867X20909696},
   publisher={SAGE Publications Sage CA: Los Angeles, CA}
 }
 
@@ -46,7 +46,7 @@
     year = 2019,
     month = dec,
     doi = {10.1111/rssb.12348},
-    url = {https://doi.org/10.1111/rssb.12348},
+    url = {10.1111/rssb.12348},
     eprint = {https://academic.oup.com/jrsssb/article-pdf/82/1/39/49320681/jrsssb\_82\_1\_39.pdf}
 }
 
@@ -134,7 +134,7 @@ publisher = {Academic Press Inc.}
   author={Frank, K.~A. and Lin, Qinyun and Maroulis, Spiro and Dai, Shimeng and Jess, Nicole and Lin, Hung-chang and Maestrales, Sarah and Searle, Ellen and Tait, Jordan and others},
   journal={Ellen and Tait, Jordan, Improving Oster’s $\delta$*: Exact Calculation for the Coefficient of Proportionality Without Subjective Specification of a Baseline Model (December 16, 2022)},
   year={2022},
-  doi ={http://dx.doi.org/10.2139/ssrn.4305243},
+  doi ={10.2139/ssrn.4305243},
   url = {https://papers.ssrn.com/sol3/papers.cfm?abstract_id=4305243}
 }
 
