### 2024-02-22 #5779 @jrosen48
> @jrosen48 please verify and add the missing DOIs above to your references

added an re-checked!
