### 2024-02-22 97d6e08
fix bug when single predictor and test_all == T
