### 2024-02-22 d0131aa
minor copy edits to paper
