### 2024-02-22 dbe97dd
Update helper_output_print.R
