+++ R/helper_output_print.R
@@ -8,7 +8,7 @@ output_print <- function(est_eff, beta_threshhold, bias = NULL, sustain = NULL, 
       cat("\n")
       cat(paste0("RIR = ", round(recase, 3), "\n"))
       cat("\n")
-      cat(paste0("You entered an estimated effect of ", round(est_eff, 3), ". To invalidate"))
+      cat(paste0("Given an estimated effect of ", round(est_eff, 3), ". To invalidate"))
       cat("\n")
       cat(paste0("the inference of an effect using the threshold of ", round(beta_threshhold, 3), " for"))
       cat("\n")
@@ -26,7 +26,7 @@ output_print <- function(est_eff, beta_threshhold, bias = NULL, sustain = NULL, 
       cat("\n")
       cat(paste0("RIR = ", round(recase, 3), "\n"))
       cat("\n")
-      cat(paste0("You entered an effect of ", round(est_eff, 3), ", and specified a threshold"))
+      cat(paste0("Given an effect of ", round(est_eff, 3), ", and specified a threshold"))
       cat("\n")
       cat(paste0("for inference of ", round(eff_thr, 3), ". To invalidate the inference based on your"))
       cat("\n")
@@ -43,7 +43,7 @@ output_print <- function(est_eff, beta_threshhold, bias = NULL, sustain = NULL, 
       cat("\n")
       cat(paste0("RIR = ", round(recase, 3), "\n"))
       cat("\n")
-      cat(paste0("You entered an estimated effect of ", round(est_eff, 3), ". The threshold value for"))
+      cat(paste0("Given an estimated effect of ", round(est_eff, 3), ". The threshold value for"))
       cat("\n")
       cat(paste0("statistical significance is ", round(beta_threshhold, 3), " (alpha = ", alpha, "). To reach that"))
       cat("\n")
@@ -59,7 +59,7 @@ output_print <- function(est_eff, beta_threshhold, bias = NULL, sustain = NULL, 
       cat("\n")
       cat(paste0("RIR = ", round(recase, 3), "\n"))
       cat("\n")
-      cat(paste0("You entered an effect size of ", round(est_eff, 3), ", and specified a threshold"))
+      cat(paste0("Given an effect size of ", round(est_eff, 3), ", and specified a threshold"))
       cat("\n")
       cat(paste0("for inference of ", round(eff_thr, 3), ". To reach that threshold, ", round(sustain, 3), "% of the" ))
       cat("\n")
