### 2024-02-22 7cdbcec
Update helper_output_print.R
