+++ R/helper_output_print.R
@@ -141,6 +141,180 @@ if (requireNamespace("htmltools", quietly = TRUE)) {
     } else if (obs_r == critical_r) {
       warning("The correlation is exactly equal to the threshold.\n")
     }
+
+    cat("For calculation of unconditional ITCV, include the R2, sdx and sdy as input and request raw output.")
+    cat("See Frank (2000) for a description of the method.")
+    cat("\n")
+    cat("\n")
+    cat(crayon::underline("Citation:"))
+    cat("\n")
+    cat("Frank, K. (2000). Impact of a confounding variable on the inference of a")
+    cat("\n")
+    cat("regression coefficient.", crayon::italic("Sociological Methods and Research, 29"), "(2), 147-194")
+    cat("\n")
+    cat("\n")
+    cat("Accuracy of results increases with the number of decimals reported.")
+    cat("\n")
+
+link_html <- '<html><body><a href="https://journals.sagepub.com/doi/10.1177/0049124100029002001">Click here for the article</a></body></html>'
+if (requireNamespace("htmltools", quietly = TRUE)) {
+  # Redirect output to null
+  sink("/dev/null")
+  invisible(htmltools::html_print(htmltools::HTML(link_html)))
+  sink()
+} else {
+  message("htmltools package is required")
+}
+
+
+  }
+    
+}
+
+
+output_print_knf <- function(est_eff, beta_threshhold, bias = NULL, sustain = NULL, nu, eff_thr, recase, obs_r, critical_r, r_con, itcv, alpha, index) {
+  if (index == "RIR"){ 
+    cat(crayon::bold("Robustness of Inference to Replacement (RIR):\n"))
+    if ((abs(est_eff) > abs(beta_threshhold)) & is.na(eff_thr) == TRUE) {
+      cat("TO INVALIDATE:\n")
+      cat("\n")
+      cat(paste0("RIR = ", recase, "\n"))
+      cat("\n")
+      cat(sprintf("Given an estimated effect of %.3f. To invalidate\n", est_eff))
+      cat(sprintf("the inference of an effect using the threshold of %.3f for\n", beta_threshhold))
+      cat(sprintf("statistical significance with alpha = %s, %.3f%% of the\n", alpha, bias))
+      cat(sprintf("(%.3f) estimate would have to be due to bias. This implies\n", est_eff))
+      cat("that to invalidate the inference one would expect to have to\n")
+      cat(sprintf("replace %.0f (%.3f%%) observations with cases for which the\n", recase, bias))
+      cat(sprintf("treatment effect is %s (RIR = %.0f).\n\n", nu, recase))
+    } else if ((abs(est_eff) > abs(beta_threshhold)) & is.na(eff_thr) == FALSE) {
+      cat("TO INVALIDATE:\n")
+      cat("\n")
+      cat(paste0("RIR = ", recase, "\n"))
+      cat(sprintf("Given an effect of %.3f, and specified a threshold\n", est_eff))
+      cat(sprintf("for inference of %.3f. To invalidate the inference based on your\n", eff_thr))
+      cat(sprintf("estimate, %.3f%% of the (%.3f) estimate would have to be due to\n", bias, est_eff))
+      cat("bias. This implies that to invalidate the inference one would\n")
+      cat(sprintf("expect to have to replace %.0f (%.3f%%) observations with\n", recase, bias))
+      cat(sprintf("cases for which the treatment effect is %s (RIR = %.0f).\n\n", nu, recase))
+    } else if ((abs(est_eff) < abs(beta_threshhold)) & is.na(eff_thr) == TRUE) {
+      cat("TO SUSTAIN:\n", sep = "")
+      cat("\n")
[…]
