### 2024-02-22 7cd46dc
a few more edits in response to Ken's feedback
