### 2024-02-26 #5779 @fabian-s
Wonderful --

Please fix/check the following typos/weird phrasings indicated by Grammarly:

- "We will model estimate the effect of the following variables on household water consumption in 1981:"  just "model" or "estimate" or "estimate the effect by modeling ..." ?
- "We have created guidelines on evaluating" --> standard usage is "for evaluating", no? (not sure about this TBH)
- "This function quantifies the sensitivity for analyses which have already" --> "that have already"
- "where impact is defined as the correlation between omitted variable and focal predictor x correlation between omitted variable and outcome. " --> "where impact is defined as the correlation between the omitted variable and the focal predictor multiplied with the correlation between the omitted variable and the outcome."
- " ## 2000) must be 0.568 X 0.568 = 0.323 to invalidate an inference for a null
    ## hypothesis of 0 effect.See Frank (2000) for a description of the method." --> add missing space  ".[ ]See Frank"

and commit these changes.

once that is done, please:

- Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title)
and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCIDs.
