### 2024-02-27 #5779 @jrosen48
Noticed a small typo I corrected - changed `&beta` (thought that would render to β) to β.
