### 2024-03-02 #5779 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
