### 2024-03-09 #5779 @oliviaguest
Huge thanks to the editor: @fabian-s and the reviewers: @fartist, @wjakethompson! ✨  JOSS appreciates your work and effort. ✨ Also, big congratulations to the @jrosen48! 🥳 🍾
